% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{annotate_src_text}
\alias{annotate_src_text}
\title{Creates srcML XML}
\usage{
annotate_src_text(srcml_path, src_folder, srcml_filepath)
}
\arguments{
\item{srcml_path}{The path to srcML binary}

\item{src_folder}{The path to the source code zip, folder or file of analysis}

\item{srcml_filepath}{The path, filename and extension (.xml) of the output XML file.}
}
\value{
The path where the output xml was saved (i.e. srcml_filepath)
}
\description{
Parses src code zip, folder or file and outputs the annotated
XML representation of the file.
}
\references{
For details, see \url{https://www.srcml.org/tutorials/creating-srcml.html}.
}
\seealso{
\code{\link{query_src_text}} to query the output file.
}
