% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{assign_exact_identity}
\alias{assign_exact_identity}
\title{Identify authors with different names and emails}
\usage{
assign_exact_identity(unique_name_email)
}
\arguments{
\item{unique_name_email}{A single string containing name and email as used in git and mailinglists.}
}
\description{
Identify authors with different names and emails
}
