% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{assign_network_identity}
\alias{assign_network_identity}
\title{Assign Common Id to Nodes and Edgelists to Git and Mbox}
\usage{
assign_network_identity(project_git, project_mbox, assign_identity_function)
}
\arguments{
\item{project_git}{the parsed mbox file from \code{\link{parse_gitlog}}}

\item{project_mbox}{the parsed mbox file from \code{\link{parse_mbox}}}

\item{assign_identity_function}{The heuristic function which decides common IDs
(currently only available: \code{\link{assign_network_identity}})}
}
\description{
Assign Common Id to Nodes and Edgelists to Git and Mbox
}
