% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_clean.R
\name{clean_jira_data}
\alias{clean_jira_data}
\title{Clean JIRA Data}
\usage{
clean_jira_data(df)
}
\arguments{
\item{df}{A data.table containing JIRA data. Must include a 'reply_from' column.}
}
\value{
A cleaned data.table with only relevant human-generated messages.
}
\description{
Removes bot-generated messages and 'Hudson' from the JIRA dataset.
}
