% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smells.R
\name{community_metric_mean_communicability}
\alias{community_metric_mean_communicability}
\title{Mean value of Decision Communicability.}
\usage{
community_metric_mean_communicability(mail.graph, code.graph)
}
\arguments{
\item{mail.graph}{A uni-modal network G = (V,E), where V are
developers (authors or contributors) and E is a collaboration between developers
in the mailing list}

\item{code.graph}{A uni-modal network G = (V,E), where V are
developers (authors or contributors) and E is a collaboration between developers
in the git log.}
}
\description{
Each collaboration between two developers (A and B) in the
software development network is considered as a possible source of
architectural and design decision, therefore a developer is considered
aware of a decision if he or she is strongly connected to at least one
of the two developers whom generated the decision. In-communicability
is related to every collaboration within the collaboration DSN and it
is based on Tamburri et al.’s formulation:

MAI = DEM - DAM

DEM = n_collaborators_of_two_developers/n_developers

DAM = n_collaborators_of_two_developers_who_communicate_with_them/n_developers

Therefore, global in-communicability can be defined as the mean MAI over
the entire collaboration network. Communicability is a global indicator
which consists n the mean of all local communicability measures, calculated
for every collaboration between two developers within the collaboration
Developer Social Network in the range in analysis.
Communicability was preferred to in-communicability in order to simplify
measurement comprehension, because in-communicability tend to be characterised
by measurements that tend to zero. Communicability is computed as:

(1 - mean(incommunicability))
}
\references{
D. A. Tamburri, F. Palomba and R. Kazman,
"Exploring Community Smells in Open-Source: An Automated Approach"
in IEEE Transactions on Software Engineering, vol. 47, no. 3,
pp. 630-652, 1 March 2021, doi: 10.1109/TSE.2019.2901490.

Simone Magnoni (2016). An approach to measure Community
Smells in software development communities. (Doctoral dissertation, Politecnico Milano).
}
