% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smells.R
\name{community_smell_primadonnas}
\alias{community_smell_primadonnas}
\title{Prima-donnas Social Smell}
\usage{
community_smell_primadonnas(
  mail.graph,
  clusters,
  code.graph,
  collaboration = 0.2,
  precomputed.black = NA
)
}
\arguments{
\item{mail.graph}{A uni-modal network G = (V,E), where V are
developers (authors or contributors) and E is a collaboration between developers
in the mailing list}

\item{clusters}{Clusters obtained via community detection of mail.graph network.}

\item{code.graph}{A uni-modal network G = (V,E), where V are
developers (authors or contributors) and E is a collaboration between developers
in the git log.}

\item{collaboration}{Total possible collaborations threshold}

\item{precomputed.black}{Optional, the return of \code{community_smell_potential_black_cloud}}
}
\description{
Given the communication graph, its sub-communities and the collaboration
graph, return the sub-community ids behaving as prima-donnas.
It is possible to specify the threshold to consider two sub-communities collaborating.
It is possible to re-use pre-computed potential black links
}
\references{
D. A. Tamburri, F. Palomba and R. Kazman,
"Exploring Community Smells in Open-Source: An Automated Approach"
in IEEE Transactions on Software Engineering, vol. 47, no. 3,
pp. 630-652, 1 March 2021, doi: 10.1109/TSE.2019.2901490.

Simone Magnoni (2016). An approach to measure Community
Smells in software development communities. (Doctoral dissertation, Politecnico Milano).
}
