% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\name{convert_pipermail_to_mbox}
\alias{convert_pipermail_to_mbox}
\title{Convert pipermail archive files (.txt and .txt.gz) into an mbox format for use with \code{\link{parse_mbox}}}
\usage{
convert_pipermail_to_mbox(filelist)
}
\arguments{
\item{filelist}{A vector of pipermail archive files from \code{\link{download_pipermail}}}
}
\value{
Returns `output`, the name of the resulting .mbox file in the current working directory
}
\description{
Convert pipermail archive files (.txt and .txt.gz) into an mbox format for use with \code{\link{parse_mbox}}
}
