% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{create_assignee}
\alias{create_assignee}
\title{Create Assignee}
\usage{
create_assignee(jira_domain_url, assignee_name)
}
\arguments{
\item{jira_domain_url}{URL of JIRA domain}

\item{assignee_name}{name of assignee}
}
\value{
A list named 'assignee' which contains the assignee's information
}
\description{
Creates a assignee cell for \code{\link{make_jira_issue}}.
}
\keyword{internal}
