% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{create_base_info}
\alias{create_base_info}
\title{Create base_info_cell}
\usage{
create_base_info(
  id = 11,
  self = "https://project.org/jira/rest/api/latest/issue/11",
  key = "PROJECT-68",
  JirAgileR_id = 1
)
}
\arguments{
\item{id}{ID of Jira issue}

\item{self}{URL of API endpoint}

\item{key}{key for Jira issue}

\item{JirAgileR_id}{JirAgileR id}
}
\value{
A list named 'base_info_cell' containing all the parameters
}
\description{
Creates and formats base_info_cell for fake Jira issue
}
