% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{create_comment}
\alias{create_comment}
\title{Create Comment}
\usage{
create_comment(comment_details)
}
\arguments{
\item{comment_details}{A list of comment_details, similar to component_details}
}
\value{
comments
}
\description{
Creates comment field with information
}
