% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{create_comments}
\alias{create_comments}
\title{Create Comment}
\usage{
create_comments(comments_body)
}
\arguments{
\item{comments_body}{A list of comment_details, similar to component_details}
}
\value{
comments_list
}
\description{
Creates comment field with information
}
