% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{create_components}
\alias{create_components}
\title{Create Components}
\usage{
create_components(components_list)
}
\arguments{
\item{components_list}{A list with the lists of components}
}
\value{
A list named 'components' which contains each component
}
\description{
Creates a list populated with components and their information from components_list
}
