% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{create_ext_info_issues}
\alias{create_ext_info_issues}
\title{Create ext_info_cell for issues (without comments)}
\usage{
create_ext_info_issues()
}
\value{
ext info cell
}
\description{
Creates and formats ext_info_cell for fake Jira issue
}
