% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{create_ext_info_issues_comments}
\alias{create_ext_info_issues_comments}
\title{Create ext_info_cell for issues (with comments)}
\usage{
create_ext_info_issues_comments()
}
\value{
ext_info_cell with comment fields
}
\description{
Creates the ext_info_cell with issues with comments
}
