% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\name{create_fake_mbox}
\alias{create_fake_mbox}
\title{Creates fake mbox data}
\usage{
create_fake_mbox(
  mlist,
  reply_from,
  reply_to,
  reply_cc,
  reply_datetime,
  timezone,
  reply_subject,
  reply_body
)
}
\arguments{
\item{mlist}{specific mailing list associated with this mbox data}

\item{reply_from}{email of sender}

\item{reply_to}{email of receiver}

\item{reply_cc}{carbon copy of email to additional recipients}

\item{reply_datetime}{Date, time in the following format: 2023-02-11T09:30:00}

\item{timezone}{Abbreviated timezone of the reply eg: HST or EST}

\item{reply_subject}{subject of the email}

\item{reply_body}{the body of the email}
}
\value{
the content of the fake mbox you created
}
\description{
Creates fake mbox data
}
