% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{create_file_directory}
\alias{create_file_directory}
\title{Creates file directories needed by the configuration file (.yml).}
\usage{
create_file_directory(conf, verbose = FALSE)
}
\arguments{
\item{conf}{The parsed config file the function is creating the directory for.}

\item{verbose}{A boolean variable that prints operational messages when set to TRUE.
Obtained from `parse_config` function.}
}
\description{
The input is the parsed config file that the user will be using.
It will take the parsed config to get the filepaths that need to be created to
use the config file. The function relies on the helper function `create_file_directory`
to see if any of the directories already exist, and create those that don't.
}
