% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{create_file_path}
\alias{create_file_path}
\title{A helper function for `create_file_directory`.}
\usage{
create_file_path(filepath, verbose = TRUE)
}
\arguments{
\item{filepath}{The filepath to create.}

\item{verbose}{A boolean variable that prints the operational messages when set to TRUE.}
}
\description{
The function checks if a filepath exists on the local device.
If the filepath does not exist, then the function creates it.
}
