% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{create_fix_versions}
\alias{create_fix_versions}
\title{Create Fix Version}
\usage{
create_fix_versions(jira_domain_url, version_names)
}
\arguments{
\item{jira_domain_url}{URL of JIRA domain}

\item{version_names}{list of version names for the issue}
}
\value{
A list named 'fixVersions' with a list of versions.
}
\description{
Create a fixVersions cell for \code{\link{make_jira_issue}}.
}
