% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_clean.R
\name{create_if_comment_column}
\alias{create_if_comment_column}
\title{Add Comment Classification and Uniquify Comment IDs (creates column)}
\usage{
create_if_comment_column(df)
}
\arguments{
\item{df}{A data frame containing columns 'reply_id' and 'reply_subject'.}
}
\value{
a data  frame with a new 'if_comment' column (TRUE if it is a comment, otherwise false).
possibly modified 'reply_id' values for rows where 'if_comment' is TRUE.
}
\description{
Adds a 'if_comment' column to the data frame to indicate whether a row is a reply to
a subject or a new subject. If it is a comment, the `reply_id` is modified to ensure uniqueness.
}
