% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{create_issue_comments}
\alias{create_issue_comments}
\title{Create Issue Comments}
\usage{
create_issue_comments(comments)
}
\arguments{
\item{comments}{A character vector containing the comment body.}
}
\description{
Create issue comments cell for \code{\link{make_jira_issue}}.
Other parameters associated to the comments, such as the author
and update author are currently hardcoded.
}
\keyword{internal}
