% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{create_issue_type}
\alias{create_issue_type}
\title{Create Issue Type}
\usage{
create_issue_type(jira_domain_url, issue_type_description)
}
\arguments{
\item{jira_domain_url}{URL of JIRA}

\item{issue_type_description}{description of the issue type}
}
\value{
A list named 'issue_type' that represents the issue type of the Jira issue
}
\description{
Create issue type cell for fake Jira issue
}
