% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{create_issuetype}
\alias{create_issuetype}
\title{Create Issue Type}
\usage{
create_issuetype(
  self_url = "https://project.org/jira/rest/api/2/issuetype/2",
  id = "2",
  description = "A new feature of the product, which has yet to be developed.",
  iconUrl =
    "https://domain.org/jira/secure/viewavatar?size=xsmall&avatarId=21141&avatarType=issuetype",
  name = "New Feature",
  subtask = FALSE,
  avatarId = 21141
)
}
\arguments{
\item{self_url}{URL of issue type's endpoint}

\item{id}{ID of Jira issue}

\item{description}{Description of issue type}

\item{iconUrl}{URL of icon representing issue type}

\item{name}{Name of issue type}

\item{subtask}{FALSE, determines if the issue is a subtask}

\item{avatarId}{ID for issue type's avatar image}
}
\value{
A list named 'issuetype' that represents the issue type of the Jira issue
}
\description{
Create issue type cell for fake Jira issue
}
