% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\name{create_mbox_from_replies}
\alias{create_mbox_from_replies}
\title{Takes in mbox replies and creates a .mbox file}
\usage{
create_mbox_from_replies(folder_path = "/tmp", folder_name, replies)
}
\arguments{
\item{folder_path}{Folder path for the .mbox file being created. Defaulted at /tmp}

\item{folder_name}{Name of the folder that will store the .mbox file}

\item{replies}{A list of replies that have been created with create_fake_mbox_replies() function that will be added to .mbox file}
}
\value{
the path of the .mbox file that was created, now that each reply has been added
}
\description{
Takes a list of mbox replies generated with create_fake_mbox_replies function and compiles them all into a single
fake .mbox file
}
