% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{create_parent}
\alias{create_parent}
\title{Create Parent}
\usage{
create_parent(jira_domain_url, parent_issue_key, status, priority, issue_type)
}
\arguments{
\item{jira_domain_url}{URL of JIRA domain}

\item{parent_issue_key}{issue key of the parent issue of the current JIRA issue}

\item{status}{status of issue for development}

\item{priority}{the name of the priority of the issue}

\item{issue_type}{type of JIRA issue}
}
\value{
A list named 'parent' that contains a parent issue and it's fields
}
\description{
Create a parent cell for \code{\link{make_jira_issue}}.
}
\keyword{internal}
