% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{create_priority}
\alias{create_priority}
\title{Create Priority}
\usage{
create_priority(jira_domain_url, priority)
}
\arguments{
\item{jira_domain_url}{URL of JIRA domain}

\item{priority}{the name of the priority of the issue (Major, Minor, Trivial)}
}
\value{
A list named 'priority' containing the priority of the issue
}
\description{
Create a priority cell for \code{\link{make_jira_issue}}.
}
\keyword{internal}
