% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{create_project}
\alias{create_project}
\title{Create Project}
\usage{
create_project(jira_domain_url, issue_key, project_type)
}
\arguments{
\item{jira_domain_url}{URL of JIRA domain}

\item{issue_key}{issue key of JIRA issue (e.g. "PROJECT-68" or "GERONIMO-6723)}

\item{project_type}{the type of the project}
}
\value{
A list named 'project' that contains project type and other project information
}
\description{
Create a project cell for \code{\link{make_jira_issue}}.
}
