% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{create_reporter}
\alias{create_reporter}
\title{Create Reporter}
\usage{
create_reporter(
  self_url = "https://domain.org/jira/rest/api/2/user?username=user1",
  name = "user1",
  key = "user1",
  displayName = "Fake User1",
  active = TRUE,
  timeZone = "Etc/UTC"
)
}
\arguments{
\item{self_url}{URL of API endpoint}

\item{name}{Name of reporter}

\item{key}{Key associated with reporter}

\item{displayName}{Display name of reporter}

\item{active}{TRUE, indicates if reporter's account is active or not}

\item{timeZone}{Timezone associated with reporter}
}
\value{
A list named 'reporter' which contains the reporter's information
}
\description{
Creates a reporter field with information
}
