% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{create_resolution}
\alias{create_resolution}
\title{Create Resolution}
\usage{
create_resolution(
  self_url = "https://domain.org/jira/rest/api/2/resolution/1",
  id = "1",
  description = "A fix for this issue is checked into the tree and tested.",
  name = "Fixed"
)
}
\arguments{
\item{self_url}{URL of API endpoint}

\item{id}{ID associated with resolution}

\item{description}{Description of resolution}

\item{name}{Name of Resolution}
}
\value{
A list named 'resolution' which contains the resolution's information
}
\description{
Creates a resolution cell for \code{\link{make_jira_issue}}.
}
\keyword{internal}
