% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{create_status}
\alias{create_status}
\title{Create Status}
\usage{
create_status(
  self_url = "https://domain.org/jira/rest/api/2/status/6",
  description =
    "The issue is considered finished, the resolution is correct. Issues which are not closed can be reopened.",
  iconUrl = "https://domain.org/jira/images/icons/statuses/closed.png",
  name = "Closed",
  id = "6",
  statusCategory_self = "https://domain.org/jira/rest/api/2/statuscategory/3",
  statusCategory_id = 3,
  statusCategory_key = "done",
  statusCategory_colorName = "green",
  statusCategory_name = "Done"
)
}
\arguments{
\item{self_url}{URL of API endpoint}

\item{description}{Description of status}

\item{iconUrl}{URL to icon associated with status}

\item{name}{Name of status}

\item{id}{ID associated with status}

\item{statusCategory_self}{URL to status category API endpoint}

\item{statusCategory_id}{ID associated with status category}

\item{statusCategory_key}{key associated with status category}

\item{statusCategory_colorName}{name of color associated with status category}

\item{statusCategory_name}{name of status category}
}
\value{
A list named 'status' containing status's information
}
\description{
Creates a status with information
}
