% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{create_versions}
\alias{create_versions}
\title{Create Versions}
\usage{
create_versions(jira_domain_url, affects_versions)
}
\arguments{
\item{jira_domain_url}{URL of JIRA domain}

\item{affects_versions}{list of version names for the issue}
}
\value{
A list named 'versions' with a list of versions
}
\description{
Create a versions cell for \code{\link{make_jira_issue}}. This cell represents
the 'Affects Version/s' label in JIRA
}
\keyword{internal}
