% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{create_votes}
\alias{create_votes}
\title{Create Votes}
\usage{
create_votes(jira_domain_url, issue_key)
}
\arguments{
\item{jira_domain_url}{URL of JIRA domain}

\item{issue_key}{issue key of JIRA issue}
}
\value{
A list named 'votes' that has the number of votes for the issue
}
\description{
Create a votes cell for \code{\link{make_jira_issue}}.
}
\keyword{internal}
