% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{create_watches}
\alias{create_watches}
\title{Create Watches}
\usage{
create_watches(jira_domain_url, issue_key)
}
\arguments{
\item{jira_domain_url}{URL of JIRA domain}

\item{issue_key}{issue key of JIRA issue}
}
\value{
A list named 'watches' that has the number of watchers for the issue
}
\description{
Create a watches cell for \code{\link{make_jira_issue}}.
}
\keyword{internal}
