% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{download_and_save_jira_issues}
\alias{download_and_save_jira_issues}
\title{Define function to fetch issues from JIRA REST API and save as JSON}
\usage{
download_and_save_jira_issues(
  domain,
  username = NULL,
  password = NULL,
  jql_query,
  fields,
  save_path_issue_tracker_issues,
  maxResults = 50,
  verbose = FALSE,
  maxDownloadsPerHour
)
}
\arguments{
\item{domain}{Custom JIRA domain URL set in config file}

\item{username}{Atlassian username as a string}

\item{password}{Atlassian token}

\item{jql_query}{Specific query string to specify criteria for fetching}

\item{fields}{List of fields that are downloaded in each issue}

\item{save_path_issue_tracker_issues}{Path that files will be save along}

\item{maxResults}{(optional) the maximum number of results to download per page.
Default is 50}

\item{verbose}{boolean flag to specify printing operational
messages or not}

\item{maxDownloadsPerHour}{Maximum downloads per hour so as to not get API blocked}
}
\description{
Download issue data from "rest/api/lastest/search" endpoint
}
