% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{download_and_save_jira_issues_by_issueKey}
\alias{download_and_save_jira_issues_by_issueKey}
\title{Define function to fetch issues from 'created' date ranges from JIRA REST API and save as JSON}
\usage{
download_and_save_jira_issues_by_issueKey(
  domain,
  credentials,
  jql_query,
  fields,
  save_path_issue_tracker_issues,
  maxResults,
  verbose,
  maxDownloads,
  issueKey_lower_bound = NULL,
  issueKey_upper_bound = NULL
)
}
\arguments{
\item{domain}{Custom JIRA domain URL set in config file}

\item{credentials}{a path to text file containing your username/api token}

\item{jql_query}{Specific query string to specify criteria for fetching}

\item{fields}{List of fields that are downloaded in each issue}

\item{save_path_issue_tracker_issues}{Path that files will be save along}

\item{maxResults}{(optional) the maximum number of results to download per page.
Default is 50}

\item{verbose}{boolean flag to specify printing operational
messages or not}

\item{maxDownloads}{Maximum downloads per function call}

\item{issueKey_lower_bound}{an optional API parameter that alters the GET request}

\item{issueKey_upper_bound}{an optional API parameter that alters the GET request}
}
\description{
Download issue data from "rest/api/lastest/search" endpoint
}
