% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{download_and_save_jira_issues_refresh}
\alias{download_and_save_jira_issues_refresh}
\title{Define function to extract greatest issueKey from a file and pass it to download_and_save_jira_issues_by_issueKey}
\usage{
download_and_save_jira_issues_refresh(
  domain,
  credentials,
  jql_query,
  fields,
  save_path_issue_tracker_issues,
  maxResults,
  verbose,
  maxDownloads,
  file_name,
  unaltered_file_path
)
}
\arguments{
\item{domain}{Custom JIRA domain URL set in config file}

\item{credentials}{a path to text file containing your username/api token}

\item{jql_query}{Specific query string to specify criteria for fetching}

\item{fields}{List of fields that are downloaded in each issue}

\item{save_path_issue_tracker_issues}{Path that files will be save along}

\item{maxResults}{(optional) the maximum number of results to download per page.
Default is 50}

\item{verbose}{boolean flag to specify printing operational
messages or not}

\item{maxDownloads}{Maximum downloads per function call}

\item{file_name}{The filename that contains the greatest issueKey. Returned by a parser}

\item{unaltered_file_path}{the unaltered filepath set in the config file}
}
\description{
Download issue data from "rest/api/2/search" endpoint
}
