% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_bugzilla}
\alias{download_bugzilla}
\title{Download Bugzilla issues and comments using Perceval.}
\usage{
download_bugzilla(
  perceval_path,
  bugzilla_site,
  datetime,
  backend = "bugzilla",
  max_bugs = 500
)
}
\arguments{
\item{perceval_path}{path to perceval binary}

\item{bugzilla_site}{link to specific bugzilla site}

\item{datetime}{fetch bugs updated since this date (in any ISO 8601 format, e.g., 'YYYY-MM-DD HH:mm:SS+|-HH:MM'))}

\item{backend}{either traditional "bugzilla" or "bugzillarest" for Bugzilla >= 5.0 servers}

\item{max_bugs}{the maximum number of bugs requested on the same query. This acts as the limit parameter
in the Bugzilla REST API. Bugzilla sites may have specific limits set, so make sure to change the max_bugs
parameter accordingly to correctly download the data when using the "bugzillarest" backend.}
}
\value{
json object
}
\description{
Note that for the Bugzilla REST API backend, Bugzilla sites may limit the number of bugs that can be retrieved at one time.
Thus, the max_bugs parameter needs to be set correctly to ensure all bugs are retrieved and that
the json data is not broken. If you get an error trying to parse the data downloaded with this
}
\seealso{
\code{\link{parse_bugzilla}} a parser function to parse the bugzilla data
}
