% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_bugzilla_comments_from_rest_api}
\alias{download_bugzilla_comments_from_rest_api}
\title{Download Bugzilla comments using the Bugzilla REST API}
\usage{
download_bugzilla_comments_from_rest_api(
  bugzilla_site,
  bug_ids,
  save_folder_path
)
}
\arguments{
\item{bugzilla_site}{URL to specific bugzilla site}

\item{bug_ids}{the ids of the bugs to extract comments for from \code{\link{download_bugzilla_issues_from_rest_api}}}

\item{save_folder_path}{the full *folder* path where the bugzilla comments will be stored}
}
\description{
Downloads comments associated with each bug id returned from \code{\link{download_bugzilla_issues_from_rest_api}}.
Each file saved contains a group of comments associated with a particular bug, where the filename is the corresponding bug id.
}
\seealso{
\code{\link{download_bugzilla_issues_from_rest_api}} a downloader function to download the bugzilla issue with GET request

\code{\link{parse_bugzilla_crawler_comments}} a parser funciton to parse bugzilla comments
}
