% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_bugzilla_from_rest_api}
\alias{download_bugzilla_from_rest_api}
\title{Download Bugzilla issues and comments using the Bugzilla REST API}
\usage{
download_bugzilla_from_rest_api(
  bugzilla_site,
  start_timestamp,
  end_timestamp,
  project_name,
  comments = FALSE
)
}
\arguments{
\item{bugzilla_site}{link to specific bugzilla site}

\item{start_timestamp}{when to start bug retrieval (ex. 2023-01-01T00:14:57Z)}

\item{end_timestamp}{when to end bug retrieval (ex. 2023-01-01T00:14:57Z)}

\item{project_name}{the name of the bugzilla project (ex. redhat for https://bugzilla.redhat.com")}

\item{comments}{whether to download comments}
}
\description{
Download Bugzilla issues and comments using the Bugzilla REST API
}
