% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugzilla.R
\name{download_bugzilla_rest_comments}
\alias{download_bugzilla_rest_comments}
\title{Download Bugzilla comments using the Bugzilla REST API}
\usage{
download_bugzilla_rest_comments(bugzilla_site, bug_ids, save_folder_path)
}
\arguments{
\item{bugzilla_site}{URL to specific bugzilla site}

\item{bug_ids}{the ids of the bugs to extract comments for from \code{\link{download_bugzilla_rest_issues}}}

\item{save_folder_path}{the full *folder* path where the bugzilla comments will be stored}
}
\description{
Downloads comments associated with each bug id returned from \code{\link{download_bugzilla_rest_issues}}.
Each file saved contains a group of comments associated with a particular bug, where the filename is the corresponding bug id.
}
\seealso{
\code{\link{download_bugzilla_rest_issues}} a downloader function to download the bugzilla issues data
}
