% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugzilla.R
\name{download_bugzilla_rest_issues_comments}
\alias{download_bugzilla_rest_issues_comments}
\title{Download project data (issues and comments) from bugzilla site
Note: The first comment in every issue is the issue description}
\usage{
download_bugzilla_rest_issues_comments(
  bugzilla_site,
  start_timestamp,
  save_folder_path,
  limit_upperbound = 500
)
}
\arguments{
\item{bugzilla_site}{URL to specific bugzilla site}

\item{start_timestamp}{when to start bug retrieval (ex. 2023-01-01T00:14:57Z)}

\item{save_folder_path}{the full *folder* path where the bugzilla issues will be stored}

\item{limit_upperbound}{the number of issues saved in each page file. Some bugzilla sites have limits set on how many bugs
can be retrieved in one GET request, in which case, the limit set by the bugzilla site will be used in place of
limit_upperbound to ensure full bug retrieval.}
}
\description{
Download project data (issues and comments) from bugzilla site
Note: The first comment in every issue is the issue description
}
\seealso{
\code{\link{parse_bugzilla_rest_issues_comments}} a parser function to parse Bugzilla issues and comments data
}
