% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugzilla.R
\name{download_bugzilla_rest_issues_comments_by_date}
\alias{download_bugzilla_rest_issues_comments_by_date}
\title{Download Bugzilla Issues and Issues Comments By Date Using REST API}
\usage{
download_bugzilla_rest_issues_comments_by_date(
  bugzilla_site,
  start_timestamp,
  save_folder_path,
  project_key,
  limit_upperbound = 20,
  comments = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{bugzilla_site}{URL to specific bugzilla site}

\item{start_timestamp}{when to start bug retrieval (ex. 2023-01-01T00:14:57Z)}

\item{save_folder_path}{the full folder path where the bugzilla issues will be stored}

\item{project_key}{the project key of the project which can be found in the respective config file}

\item{limit_upperbound}{the number of issues saved in each page file. Some bugzilla sites have limits set on how many bugs
can be retrieved in one GET request, in which case, the limit set by the bugzilla site will be used in place of
limit_upperbound to ensure full bug retrieval.}

\item{comments}{set true to download issues with comments, leave as false to download only issues (without comments)}

\item{verbose}{set true to print execution details}
}
\description{
Downloads Bugzilla issues or issues with comments into a folder, where each file is a
json containing a page of issues. The files downloaded start from the "start_timestamp" parameter
until the current time.
}
\seealso{
\code{\link{parse_bugzilla_rest_issues_comments}} a parser function to parse Bugzilla issues and comments data
}
