% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{download_jira_issues_comments_by_issuekey}
\alias{download_jira_issues_comments_by_issuekey}
\title{Define function to fetch issues from 'created' date ranges from JIRA REST API and save as JSON}
\usage{
download_jira_issues_comments_by_issuekey(
  domain,
  credentials,
  jql_query,
  fields,
  save_path_issue_tracker_issues,
  maxresults,
  verbose,
  maxdownloads,
  issuekey_lower_bound = NULL,
  issuekey_upper_bound = NULL
)
}
\arguments{
\item{domain}{Custom JIRA domain URL set in config file}

\item{credentials}{a path to text file containing your username/api token}

\item{jql_query}{Specific query string to specify criteria for fetching}

\item{fields}{List of fields that are downloaded in each issue}

\item{save_path_issue_tracker_issues}{Path that files will be save along}

\item{maxresults}{(optional) the maximum number of results to download per page.
Default is 50}

\item{verbose}{boolean flag to specify printing operational
messages or not}

\item{maxdownloads}{Maximum downloads per function call}

\item{issuekey_lower_bound}{an optional API parameter that alters the GET request}

\item{issuekey_upper_bound}{an optional API parameter that alters the GET request}
}
\description{
Download issue data from "rest/api/lastest/search" endpoint
}
\seealso{
\code{link{download_jira_issues_comments}} to download all JIRA issues data,

\code{link{download_jira_issues_comments_by_date}} to download JIRA data using created as a search criteria,

\code{link{refresh_jira_issues_comments_by_issuekey}} to download only JIRA data that has not already been downloaded,

\code{link{parse_jira}} to parse jira issues along a file-path

Other jira: 
\code{\link{download_jira_issues_comments_by_date}()},
\code{\link{download_jira_issues_comments}()},
\code{\link{refresh_jira_issues_comments_by_issuekey}()}

Other downloaders: 
\code{\link{download_jira_issues_comments_by_date}()},
\code{\link{download_jira_issues_comments}()},
\code{\link{refresh_jira_issues_comments_by_issuekey}()}
}
\concept{downloaders}
\concept{jira}
