% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_mod_mbox}
\alias{download_mod_mbox}
\title{Compose mod_mbox archives (.mbox) into a single mbox file for use with \code{\link{parse_mbox}}}
\usage{
download_mod_mbox(base_url, mailinglist, from, to)
}
\arguments{
\item{mailinglist}{Name of the project mailing list (e.g. apr-dev) in the mod_mbox directory}

\item{from}{First year in the range to be downloaded}

\item{to}{Last year in the range to be downloaded}

\item{url}{An url pointing to the mod_mbox directory (e.g. "http://mail-archives.apache.org/mod_mbox") without trailing slashes}
}
\value{
Returns `output`, the name of the resulting .mbox file in the current working directory
}
\description{
Compose mod_mbox archives (.mbox) into a single mbox file for use with \code{\link{parse_mbox}}
}
