% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\name{download_mod_mbox_per_month}
\alias{download_mod_mbox_per_month}
\title{Compose mod_mbox archives (.mbox) into a single mbox file for use with \code{\link{parse_mbox}}}
\usage{
download_mod_mbox_per_month(
  archive_url,
  mailing_list,
  archive_type,
  from_year,
  to_year,
  save_folder_path,
  verbose = FALSE
)
}
\arguments{
\item{archive_url}{A url pointing to the mod_mbox mailing list directory (e.g. "http://mail-archives.apache.org/mod_mbox/apr-dev") without trailing slashes}

\item{mailing_list}{Name of the project mailing list (e.g. apr-dev) in the mod_mbox directory}

\item{archive_type}{Name of the archive that the project mailing list is archived in (e.g. apache)}

\item{from_year}{First year in the range to be downloaded}

\item{to_year}{Last year in the range to be downloaded}

\item{save_folder_path}{the full *folder* path where the monthly downloaded mbox will be stored.}

\item{verbose}{Prints progress during execution}
}
\value{
Returns the path of the downloaded mbox file.
}
\description{
Compose mod_mbox archives (.mbox) into a single mbox file for use with \code{\link{parse_mbox}}
}
