% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\name{download_pipermail}
\alias{download_pipermail}
\title{Download all pipermail files in an archive as mbox files}
\usage{
download_pipermail(
  mailing_list,
  start_year_month,
  end_year_month,
  save_folder_path
)
}
\arguments{
\item{mailing_list}{The name of the mailing list being downloaded}

\item{start_year_month}{The year and month of the first file to be downloaded}

\item{end_year_month}{The year and month of the last file to be downloaded}

\item{save_folder_path}{The folder path in which all the downloaded pipermail files will be stored}
}
\value{
Returns `downloaded_files`, a vector of the downloaded files in the current working directory
}
\description{
Download all pipermail files in an archive as mbox files
}
