% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv8.R
\name{dv8_clsxj_to_clsxb}
\alias{dv8_clsxj_to_clsxb}
\title{Convert a json cluster file, clsxj, to a binary cluster file, clsxb.}
\usage{
dv8_clsxj_to_clsxb(dv8_path, clsxj_path, clsxb_path)
}
\arguments{
\item{dv8_path}{path to dv8 binary}

\item{clsxj_path}{path to JSON cluster created by \code{\link{dv8_mdsmb_to_hierclsxb}}}

\item{clsxb_path}{path to save the DV8 binary cluster}
}
\value{
the clsxb_path
}
\description{
An input file \*-hier.json is converted to \*-clsx.dv8-clsx.
}
\seealso{
Other dv8:
\code{\link[=dependencies_to_sdsmj]{dependencies_to_sdsmj()}},
\code{\link[=dv8_clsxb_to_clsxj]{dv8_clsxb_to_clsxj()}},
\code{\link[=dv8_depends_to_sdsmj]{dv8_depends_to_sdsmj()}},
\code{\link[=dv8_dsmj_to_dsmb]{dv8_dsmj_to_dsmb()}},
\code{\link[=dv8_hdsmb_sdsmb_to_mdsmb]{dv8_hdsmb_sdsmb_to_mdsmb()}},
\code{\link[=dv8_mdsmb_to_flaws]{dv8_mdsmb_to_flaws()}},
\code{\link[=gitlog_to_hdsmj]{gitlog_to_hdsmj()}},
\code{\link[=graph_to_dsmj]{graph_to_dsmj()}},
\code{\link[=parse_dv8_architectural_flaws]{parse_dv8_architectural_flaws()}},
\code{\link[=parse_dv8_clusters]{parse_dv8_clusters()}},
\code{\link[=transform_dependencies_to_sdsmj]{transform_dependencies_to_sdsmj()}},
\code{\link[=transform_gitlog_to_hdsmj]{transform_gitlog_to_hdsmj()}},
\code{\link[=transform_temporal_gitlog_to_adsmj]{transform_temporal_gitlog_to_adsmj()}}
}
\concept{dv8}
