% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv8.R
\name{dv8_depends_to_sdsmj}
\alias{dv8_depends_to_sdsmj}
\title{Creates dependencies using Depends}
\usage{
dv8_depends_to_sdsmj(depends_jar_path, git_repo_path, language, sdsmj_path)
}
\arguments{
\item{depends_jar_path}{path to depends jar}

\item{git_repo_path}{path to git repo (ends in .git)}

\item{language}{the language of the .git repo (accepts cpp, java, ruby, python, pom)}

\item{sdsmj_path}{the path to save the sdsm-dv8.dsm file.}
}
\value{
the sdsmj_path
}
\description{
Creates dependencies using Depends
}
\seealso{
\code{\link{dv8_dsmj_to_dsmb}} to convert to `*-sdsm.dv8-dsm` and
\code{\link{dv8_hdsmb_sdsmb_to_mdsmb}} to merge DSMs into `*-merge.dv8-dsm`.

Other dv8:
\code{\link[=dependencies_to_sdsmj]{dependencies_to_sdsmj()}},
\code{\link[=dv8_clsxb_to_clsxj]{dv8_clsxb_to_clsxj()}},
\code{\link[=dv8_clsxj_to_clsxb]{dv8_clsxj_to_clsxb()}},
\code{\link[=dv8_dsmj_to_dsmb]{dv8_dsmj_to_dsmb()}},
\code{\link[=dv8_hdsmb_sdsmb_to_mdsmb]{dv8_hdsmb_sdsmb_to_mdsmb()}},
\code{\link[=dv8_mdsmb_to_flaws]{dv8_mdsmb_to_flaws()}},
\code{\link[=gitlog_to_hdsmj]{gitlog_to_hdsmj()}},
\code{\link[=graph_to_dsmj]{graph_to_dsmj()}},
\code{\link[=parse_dv8_architectural_flaws]{parse_dv8_architectural_flaws()}},
\code{\link[=parse_dv8_clusters]{parse_dv8_clusters()}},
\code{\link[=transform_dependencies_to_sdsmj]{transform_dependencies_to_sdsmj()}},
\code{\link[=transform_gitlog_to_hdsmj]{transform_gitlog_to_hdsmj()}},
\code{\link[=transform_temporal_gitlog_to_adsmj]{transform_temporal_gitlog_to_adsmj()}}
}
\concept{dv8}
