% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv8.R
\name{dv8_dsmb_to_dsmj}
\alias{dv8_dsmb_to_dsmj}
\title{Export a dsmb binary file as a separate json file}
\usage{
dv8_dsmb_to_dsmj(dv8_path, dsmb_path, dsmj_path)
}
\arguments{
\item{dv8_path}{path to DV8 binary}

\item{dsmb_path}{path to dsmb binary file created by \code{\link{dv8_gitnumstat_to_hdsmb}}}

\item{dsmj_path}{path to json file}
}
\value{
the dsmj_path
}
\description{
Takes a \*.dv8-dsm file and exports as \*.json file
}
