% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv8.R
\name{dv8_hsdsmb_to_decoupling_level}
\alias{dv8_hsdsmb_to_decoupling_level}
\title{Calculates the Decoupling Level from a hsdsm binary file}
\usage{
dv8_hsdsmb_to_decoupling_level(dv8_path, hsdsmb_path, json_path)
}
\arguments{
\item{dv8_path}{path to DV8 binary}

\item{hsdsmb_path}{path to hsdsm binary file created by \code{\link{dv8_gitnumstat_to_hdsmb}}}

\item{json_path}{name of output file}
}
\value{
path of output json file
}
\description{
Creates a *.json file with decoupling metrics from *.dv8-dsm file
}
\seealso{
\code{\link{parse_dv8_metrics_decoupling_level}} to parse decoupling level json file
}
