% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv8.R
\name{dv8_hsdsmb_to_flaws}
\alias{dv8_hsdsmb_to_flaws}
\title{Detect Architecture anti-patterns}
\usage{
dv8_hsdsmb_to_flaws(hsdsmb_path, flaws_path)
}
\arguments{
\item{hsdsmb_path}{path to the merged matrix input file dv8-dsm}

\item{flaws_path}{path to architecture folder}
}
\description{
Detect Architecture anti-patterns
}
