% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv8.R
\name{dv8_hsdsmb_to_hierclsxb}
\alias{dv8_hsdsmb_to_hierclsxb}
\title{Computes the design rule hierarchy for a hsdsm binary file}
\usage{
dv8_hsdsmb_to_hierclsxb(
  dv8_path,
  hsdsmb_path,
  hierclsxb_path,
  max_depth = "",
  modules = FALSE,
  recursive = FALSE
)
}
\arguments{
\item{dv8_path}{path to DV8 binary}

\item{hsdsmb_path}{path to hsdsm binary file created by \code{\link{dv8_gitnumstat_to_hdsmb}}}

\item{hierclsxb_path}{name of output file}

\item{max_depth}{recursion depth limit}

\item{modules}{merge modules following same design rules if switched on}

\item{recursive}{use recursive algorithm if switched on}
}
\value{
path of output file
}
\description{
Creates a *.dv8-clsx file from a *.dv8-dsm file
}
\seealso{
\code{\link{dv8_hsdsmb_drhier_to_excel}} for using *.dv8-clsx file when exporting hsdsmb binary file as an excel file
}
