% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv8.R
\name{dv8_mdsmb_drhier_to_excel}
\alias{dv8_mdsmb_drhier_to_excel}
\title{Export a mdsmb binary file into a spreadsheet}
\usage{
dv8_mdsmb_drhier_to_excel(
  dv8_path,
  mdsmb_path,
  excel_path,
  hierclsxm_path = "",
  detail = FALSE,
  drhier = TRUE,
  namespace = FALSE
)
}
\arguments{
\item{dv8_path}{path to DV8 binary}

\item{mdsmb_path}{path to mdsmb binary file created by \code{\link{dv8_gitnumstat_to_hdsmb}}}

\item{excel_path}{path to output excel file}

\item{hierclsxm_path}{path to clustering file used to order files in spreadsheet from \code{\link{dv8_mdsmb_to_hierclsxb}}}

\item{detail}{if TRUE, include details within each cell in spreadsheet, default FALSE}

\item{drhier}{if TRUE, use recursive drh clustering in spreadsheet, default FALSE}

\item{namespace}{if TRUE, use namespace clustering in spreadsheet, default FALSE}
}
\value{
the excel_path
}
\description{
Creates a \*.xlsx from a \*.dv8-dsm file
}
