% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv8.R
\name{dv8_mdsmb_to_decoupling_level}
\alias{dv8_mdsmb_to_decoupling_level}
\title{Computes the Decoupling Level from a mdsm binary file}
\usage{
dv8_mdsmb_to_decoupling_level(dv8_path, mdsmb_path, dl_path)
}
\arguments{
\item{dv8_path}{path to DV8 binary}

\item{mdsmb_path}{path to mdsm binary file created by \code{\link{dv8_gitnumstat_to_hdsmb}}}

\item{dl_path}{name of output file}
}
\value{
the dl_path
}
\description{
Creates a \*.json file with decoupling metrics from \*.dv8-dsm file
}
\seealso{
\code{\link{parse_dv8_metrics_decoupling_level}} to parse decoupling level JSON file `-dl.json`
}
