% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv8.R
\name{dv8_mdsmb_to_hierclsxb}
\alias{dv8_mdsmb_to_hierclsxb}
\title{Computes the design rule hierarchy for a mdsm binary file}
\usage{
dv8_mdsmb_to_hierclsxb(
  dv8_path,
  mdsmb_path,
  hierclsxb_path,
  max_depth = "",
  modules = FALSE,
  recursive = FALSE
)
}
\arguments{
\item{dv8_path}{path to DV8 binary}

\item{mdsmb_path}{path to mdsm binary file created by \code{\link{dv8_gitnumstat_to_hdsmb}}}

\item{hierclsxb_path}{name of output file}

\item{max_depth}{recursion depth limit}

\item{modules}{merge modules following same design rules if switched on}

\item{recursive}{use recursive algorithm if switched on}
}
\value{
the hierclsxb_path
}
\description{
Creates a \*.dv8-clsx file from a \*.dv8-dsm file
}
\seealso{
\code{\link{dv8_mdsmb_drhier_to_excel}} for exporting `*-clsx.dv8-clsx` to excel `-clsx.xlsx`, or
\code{\link{dv8_clsxb_to_clsxj}} and \code{\link{parse_dv8_clusters}} for conversion and parsing.
}
