% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_different_branches}
\alias{example_different_branches}
\title{Example Commit Different Branches}
\usage{
example_different_branches()
}
\value{
git_repo_path of newly created empty repo
}
\description{
One commit in two different with branches with 1 file each.
}
\details{
Checks if parse_gitlog includes both commits in gitlog.
}
