% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_different_files_commits}
\alias{example_different_files_commits}
\title{Example Different Files Commit}
\usage{
example_different_files_commits()
}
\value{
git_repo_path of newly created empty repo
}
\description{
Repo with 2 commits, first one contains 5 files modified, second contains only one
}
\details{
Tests if parse_gitlog includes both commits with different # of files in gitlog.
}
