% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_jira_issue_comments}
\alias{example_jira_issue_comments}
\title{Example Jira Issue Tracker With Comments}
\usage{
example_jira_issue_comments(folder_path = "/tmp", folder_name)
}
\arguments{
\item{folder_path}{The path where the folder will be created}

\item{folder_name}{The name of the folder}
}
\value{
the JSON path of the newly created issue issue tracker
}
\description{
Create fake jira issue tracker with one issue with 2 comments
}
\keyword{internal}
