% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_jira_issue_components}
\alias{example_jira_issue_components}
\title{Create one No-Comment Issue with Two Components}
\usage{
example_jira_issue_components(folder_path = "/tmp", folder_name)
}
\arguments{
\item{folder_path}{The path where the folder will be created}

\item{folder_name}{The name of the folder}
}
\value{
the JSON path of the newly created issue issue tracker
}
\description{
This example can be used to evaluate the parser does not replicate
new components on new issues, which would severely bias metrics
associated to issues such as bugs (see #244).
}
\keyword{internal}
