% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_jira_issue_resolution_datetimetz}
\alias{example_jira_issue_resolution_datetimetz}
\title{Example Jira Issue No Assignee}
\usage{
example_jira_issue_resolution_datetimetz(folder_path = "/tmp")
}
\arguments{
\item{folder_path}{The path where the folder will be created}
}
\value{
git_repo_path of newly created empty repo
}
\description{
create fake jira issue with no assignee, calls parse_jira
}
