% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_jira_issue_tracker}
\alias{example_jira_issue_tracker}
\title{Example Jira Issue Tracker No Comments}
\usage{
example_jira_issue_tracker(folder_path = "/tmp")
}
\arguments{
\item{folder_path}{The path where the folder will be created}
}
\value{
git_repo_path of newly created empty repo
}
\description{
Create fake jira issue tracker with 2 issues, no comments
}
