% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_jira_issue_tracker_comments}
\alias{example_jira_issue_tracker_comments}
\title{Example Jira Issue Tracker With Comments}
\usage{
example_jira_issue_tracker_comments(folder_path = "/tmp")
}
\arguments{
\item{folder_path}{The path where the folder will be created}
}
\value{
git_repo_path of newly created empty repo
}
\description{
Create fake jira issue with comments
}
