% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_jira_two_issues}
\alias{example_jira_two_issues}
\title{Example JIRA Issue Tracker No Comments}
\usage{
example_jira_two_issues(folder_path = "/tmp", folder_name)
}
\arguments{
\item{folder_path}{The path where the folder will be created}

\item{folder_name}{The name of the folder}
}
\value{
the JSON path of the newly created issue issue tracker
}
\description{
Create fake JIRA issue tracker with 2 issues, no comments
}
\keyword{internal}
