% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_large_sized_commits}
\alias{example_large_sized_commits}
\title{Example Different Files Commit}
\usage{
example_large_sized_commits(folder_path = "/tmp", folder_name)
}
\arguments{
\item{folder_path}{The path where the folder will be created}

\item{folder_name}{The name of the folder}
}
\value{
git_repo_path of newly created empty repo
}
\description{
Repo with 2 commits. The first commit contains 5 files modified, and
second commit contains only one file modified.
}
\details{
Useful to test unbalanced sized commits and filters.
}
\keyword{internal}
