% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_mailing_list_two_threads}
\alias{example_mailing_list_two_threads}
\title{Two Thread and Three Replies Mailing List}
\usage{
example_mailing_list_two_threads(folder_path = "/tmp", folder_name, file_name)
}
\arguments{
\item{folder_path}{Default folder path set to "/tmp"}

\item{folder_name}{Name of the example folder}

\item{file_name}{Name of the file where .mbox will be stored}
}
\value{
Folder path of .mbox sample file that was created
}
\description{
Create a mailing list of two e-mail threads, with
two and one reply respectively by two developers.
}
\keyword{internal}
