% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_mbox_normal}
\alias{example_mbox_normal}
\title{Example fake Mbox data with all parameters}
\usage{
example_mbox_normal(folder_path = "/tmp", folder_name)
}
\arguments{
\item{folder_path}{Default folder path set to "/tmp"}

\item{folder_name}{Name of folder where .mbox file will be stored}
}
\value{
Folder path of .mbox sample file that was created
}
\description{
Creating a fake mbox with all parameters filled in, to ensure the function generates mbox properly
}
