% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_mbox_three_replies}
\alias{example_mbox_three_replies}
\title{Create sample mbox with three replies}
\usage{
example_mbox_three_replies(folder_path = "/tmp", folder_name)
}
\arguments{
\item{folder_path}{Default folder path set to "/tmp"}

\item{folder_name}{Name of folder where .mbox file will be stored}
}
\value{
Folder path of .mbox sample file that was created
}
\description{
Creating a fake mbox with three replies. The first two fake mbox replies have the same email subject while the
third reply has a different subject.
}
