% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_notebook_alternating_function_in_files}
\alias{example_notebook_alternating_function_in_files}
\title{Example Alternating Undecided Developers}
\usage{
example_notebook_alternating_function_in_files(
  folder_path = "/tmp",
  folder_name
)
}
\arguments{
\item{folder_path}{The path where the folder will be created}

\item{folder_name}{The name of the folder}
}
\value{
git_repo_path of newly created empty repo
}
\description{
Developers keep alternating the lines changes.
}
\details{
Useful to check how git log entity behaves with files it should
recognize.
}
\keyword{internal}
