% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_notebook_function_in_code_blocks}
\alias{example_notebook_function_in_code_blocks}
\title{Example Commit of R Notebooks}
\usage{
example_notebook_function_in_code_blocks(folder_path = "/tmp", folder_name)
}
\arguments{
\item{folder_path}{The path where the folder will be created}

\item{folder_name}{The name of the folder}
}
\value{
git_repo_path of newly created empty repo
}
\description{
One commit that defines the function by Dev 1, then
one commit that modifies the same function by Dev 2.
}
\details{
Useful to check how git log entity behaves with files it does
not recognize.
}
\keyword{internal}
