% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_renamed_file}
\alias{example_renamed_file}
\title{Example Renamed File Repo}
\usage{
example_renamed_file(folder_path = "/tmp", folder_name)
}
\arguments{
\item{folder_path}{The path where the folder will be created}

\item{folder_name}{The name of the folder}
}
\value{
git_repo of newly created empty repo
}
\description{
A repo with 3 commits. The first adds hello.R, , the second
renames the file to hi.R. and the third adds a second file bye.R.
}
\details{
This example can be used to test how parsers trace file renaming.
}
\keyword{internal}
