% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_repo_prefix}
\alias{example_repo_prefix}
\title{example_repo_prefix function filters files with certain prefixes}
\usage{
example_repo_prefix(folder_path = "/tmp", folder_name)
}
\arguments{
\item{folder_path}{The path where the folder will be created}

\item{folder_name}{The name of the folder}
}
\value{
git_repo of newly created empty repo
}
\description{
A repo with 3 commits, where 1 file has as prefix _test.R,
1 file has the suffix example_*.R, and 1 file hello.R.
The second renames the file to hi.R. The third adds a second file bye.R
}
