% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_repo_three_commits}
\alias{example_repo_three_commits}
\title{A repo with 3 commits. The first adds hello.R.
The second renames the file to hi.R. The third adds a second file bye.R}
\usage{
example_repo_three_commits()
}
\value{
git_repo of newly created empty repo
}
\description{
A repo with 3 commits. The first adds hello.R.
The second renames the file to hi.R. The third adds a second file bye.R
}
