% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_test_example_src_repo}
\alias{example_test_example_src_repo}
\title{Example Unit Test and Examples Repository}
\usage{
example_test_example_src_repo(folder_path = "/tmp", folder_name)
}
\arguments{
\item{folder_path}{The path where the folder will be created}

\item{folder_name}{The name of the folder}
}
\value{
git_repo of newly created empty repo
}
\description{
A repository which contains test, example and
source files. Can be useful to test filter functions.
}
\details{
The repo contains 3 commits, where 1 file has as prefix
 \_test.R, 1 file has the suffix example\_*.R, and 1 file
 hello.R. The second renames the file to hi.R.
 The third adds a second file bye.R
}
\keyword{internal}
