% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{extension}
\alias{extension}
\title{Find the extension of a file}
\usage{
extension(filename)
}
\arguments{
\item{filename}{The path or the name of a file}
}
\value{
The extension of a file
}
\description{
Find the extension of a file
}
