% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_by_last_files_change}
\alias{filter_by_last_files_change}
\title{Filter Git Log by all last files changes in snapshot}
\usage{
filter_by_last_files_change(git_log, p_commit_hash)
}
\arguments{
\item{git_log}{A parsed git project by \code{parse_gitlog}.}

\item{p_commit_hash}{A commit hash from git_log}
}
\description{
Filter a git log to show the last commit change made
to a file since the start of the project.
}
