% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_by_reply_body_substring}
\alias{filter_by_reply_body_substring}
\title{Filter Replies by Body Substring}
\usage{
filter_by_reply_body_substring(reply_dt, substrings, case_insensitive)
}
\arguments{
\item{reply_dt}{A data.table from \code{\link{parse_jira_replies}}, \code{\link{parse_github_replies}} or \code{\link{parse_mbox}}.}

\item{substrings}{A character vector of substrings to filter.}

\item{case_insensitive}{Logical. Must explicitly pass TRUE or FALSE.}
}
\value{
A filtered data.table without matching bodies.
}
\description{
Removes rows where the message body contains certain substrings.
}
\seealso{
Other filters: 
\code{\link{filter_by_commit_interval}()},
\code{\link{filter_by_commit_size}()},
\code{\link{filter_by_file_extension}()},
\code{\link{filter_by_filepath_substring}()},
\code{\link{filter_by_reply_author_substring}()},
\code{\link{filter_by_reply_subject_substring}()},
\code{\link{replace_token_regex_with}()}
}
\concept{filters}
