% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_text_fenced_code_block}
\alias{filter_text_fenced_code_block}
\title{Remove fenced code blocks}
\usage{
filter_text_fenced_code_block(text_vector)
}
\arguments{
\item{text_vector}{A character vector containing text to clean.}
}
\value{
A character vector with fenced code blocks removed.
}
\description{
Removes fenced code blocks marked with triple backticks ``` from text.
}
