% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_text_newlines}
\alias{filter_text_newlines}
\title{Remove carriage returns and newlines}
\usage{
filter_text_newlines(text_vector)
}
\arguments{
\item{text_vector}{A character vector to clean.}
}
\value{
A character vector with \r and \n replaced by space.
}
\description{
Converts all \r and \n characters to a space, without collapsing the text fully.
}
