% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_text_punctuation}
\alias{filter_text_punctuation}
\title{Remove punctuation and symbols from text}
\usage{
filter_text_punctuation(text_vector)
}
\arguments{
\item{text_vector}{A character vector to be cleaned.}
}
\value{
A character vector with all punctuation and symbols removed.
}
\description{
Cleans a character vector by removing all Unicode punctuation and symbol characters.
}
