% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_text_quoted_lines}
\alias{filter_text_quoted_lines}
\title{Remove quoted lines}
\usage{
filter_text_quoted_lines(text_vector)
}
\arguments{
\item{text_vector}{A character vector containing text to clean.}
}
\value{
A character vector with quoted lines removed.
}
\description{
Removes lines starting with '>', which are typically quoted text.
}
