% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_text_whitespace}
\alias{filter_text_whitespace}
\title{Trim leading and trailing whitespace}
\usage{
filter_text_whitespace(text_vector)
}
\arguments{
\item{text_vector}{A character vector containing text to clean.}
}
\value{
A character vector with whitespace trimmed.
}
\description{
Removes leading and trailing whitespace from each element of a character vector.
}
