% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identity.R
\name{format_name_email}
\alias{format_name_email}
\title{Format Name and Email}
\usage{
format_name_email(name_email)
}
\arguments{
\item{name_email}{A string containing name and email}
}
\description{
Format Name and Email
}
