% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_bugzilla_project_key}
\alias{get_bugzilla_project_key}
\title{Get Bugzilla Project Key}
\usage{
get_bugzilla_project_key(project_name)
}
\arguments{
\item{project_name}{the name of the project config file (e.g. "kaiaulu" or "geronimo")}
}
\value{
the list of Jira issues comments paths
}
\description{
Gets the project key for the Bugzilla project
}
