% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_code_language}
\alias{get_code_language}
\title{Get Code Language}
\usage{
get_code_language(config_file)
}
\arguments{
\item{config_file}{The parsed config file}
}
\value{
the code language
}
\description{
Gets the code language of the project
}
