% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_conda_path}
\alias{get_conda_path}
\title{Returns the path to conda}
\usage{
get_conda_path(config_file)
}
\arguments{
\item{config_file}{A parsed configuration file obtained from \code{\link{parse_config}}.}
}
\value{
The file path to the PySenti script as a string.
}
\description{
This function retrieves the path to conda so system calls can be
used from R to tools in other conda environments when necessary.
}
