% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_conf}
\alias{get_conf}
\title{Get Parsed Config File}
\usage{
get_conf(config_path)
}
\arguments{
\item{config_path}{The path of the config file from the kaiaulu directory (e.g. "conf/kaiaulu.yml")}
}
\value{
the parsed config file whose path is specified by `config_path`
}
\description{
Gets the parsed configuration file for any path-specified project
}
