% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_cveid_regex}
\alias{get_cveid_regex}
\title{Get CVE ID Regex}
\usage{
get_cveid_regex(config_file)
}
\arguments{
\item{config_file}{The parsed config file}
}
\value{
the commit message CVE
}
\description{
Gets the commit message CVE
}
