% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{get_date_from_commit_hash}
\alias{get_date_from_commit_hash}
\title{Get datetime from commit hash in a git log}
\usage{
get_date_from_commit_hash(git_log, p_commit_hash)
}
\arguments{
\item{git_log}{a data.table where the key is commit+file}

\item{p_commit_hash}{a string character with the commit hash we want to obtain the datetime from
commit interval on the form `commitsha1-commitsha2`}
}
\value{
a numeric vector of `metric_function` values for each commit interval specified in `dt_range`
}
\description{
Locates the commit hash in `git_log` and obtains the timestamp. Assumes `git_log` contains a `data.AuthorDate``
column.
}
