% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_dv8_flaws_params}
\alias{get_dv8_flaws_params}
\title{Get Flaws Params}
\usage{
get_dv8_flaws_params(project_name)
}
\arguments{
\item{project_name}{the name of the project config file (e.g. "kaiaulu" or "geronimo")}
}
\value{
the flaws params
}
\description{
Gets the architectural flaws thresholds
}
