% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_enumeration_commits}
\alias{get_enumeration_commits}
\title{Get Enumeration Commits}
\usage{
get_enumeration_commits(project_name)
}
\arguments{
\item{project_name}{the name of the project config file (e.g. "kaiaulu" or "geronimo")}
}
\value{
the commit intervals as a list
}
\description{
Gets the list of enumerated commit intervals for analysis.
}
