% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_file_extensions}
\alias{get_file_extensions}
\title{Get File Extensions}
\usage{
get_file_extensions(project_name)
}
\arguments{
\item{project_name}{the name of the project config file (e.g. "kaiaulu" or "geronimo")}
}
\value{
the file extensions
}
\description{
Gets the file extensions for filtering
}
