% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_filter_by_reply_body_substring}
\alias{get_filter_by_reply_body_substring}
\title{Returns the body substrings used for filtering reply tables.}
\usage{
get_filter_by_reply_body_substring(config_file)
}
\arguments{
\item{config_file}{The parsed configuration file.}
}
\value{
A character vector of body substrings.
}
\description{
This function returns the list of body substrings used to filter replies
in the configuration file. The input `config_file` must be a parsed YAML configuration
file obtained from \code{\link{parse_config}}. The function will warn the user if
the attribute does not exist in the configuration file.
}
