% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_filter_commit_size}
\alias{get_filter_commit_size}
\title{Get Filter Commit Size}
\usage{
get_filter_commit_size(project_name)
}
\arguments{
\item{project_name}{the name of the project config file (e.g. "kaiaulu" or "geronimo")}
}
\value{
the commit size to filter out
}
\description{
Gets the commit size to filter out
}
