% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_git_branches}
\alias{get_git_branches}
\title{Get Project Git Branches}
\usage{
get_git_branches(project_name)
}
\arguments{
\item{project_name}{the name of the project config file (e.g. "kaiaulu" or "geronimo")}
}
\value{
the list of git branches
}
\description{
Gets the list of git branches of the project
}
