% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_git_repo_path}
\alias{get_git_repo_path}
\title{Get Project Git Repo Path}
\usage{
get_git_repo_path(project_name)
}
\arguments{
\item{project_name}{the name of the project config file (e.g. "kaiaulu" or "geronimo")}
}
\value{
the local git repo path
}
\description{
Gets the local project git repo path from the config file
}
