% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_github_commit_comment_path}
\alias{get_github_commit_comment_path}
\title{Returns the local folder path for GitHub commit comments for a specific project key.}
\usage{
get_github_commit_comment_path(config_file, project_key_index)
}
\arguments{
\item{config_file}{The parsed configuration file obtained from \code{\link{parse_config}}.}

\item{project_key_index}{The name of the index of the project key (e.g. "project_key_1" or "project_key_2").}
}
\value{
The local folder path for GitHub commit comments for project specified by key `project_key_index`.
}
\description{
This function returns the local folder path for GitHub commit comments
for a specific project key, that is specified in the input
parameter `config_file`. The input, `config_file` must be a parsed
configuration file. The function will inform the user if the local folder
path for the commit comments exists in the parsed configuration file,
`config_file`.
}
