% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_github_issue_or_pr_comment_path}
\alias{get_github_issue_or_pr_comment_path}
\title{Get GitHub Issue Or PR Comment Save Path}
\usage{
get_github_issue_or_pr_comment_path(config_file, project_key_index)
}
\arguments{
\item{config_file}{The parsed config file}

\item{project_key_index}{The name of the index of the project key (e.g. "project_key_1" or "project_key_2")}
}
\value{
the local save path for Github Issues or PR comments for project specified by key index `project_key_index`
}
\description{
Gets the local save path for GitHub Issues or PR Comments for a specific project key index
}
