% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_github_issue_search_path}
\alias{get_github_issue_search_path}
\title{Get GitHub Issue Search Save Path}
\usage{
get_github_issue_search_path(project_name, project_key_index)
}
\arguments{
\item{project_name}{the name of the project config file (e.g. "kaiaulu" or "geronimo")}

\item{project_key_index}{the name of the index of the project key (e.g. "project_key_1" or "project_key_2")}
}
\value{
the local save path for GitHub issue search
}
\description{
Gets the local save path for GitHub issue search for a specific project key index
}
