% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_github_owner}
\alias{get_github_owner}
\title{Get GitHub Owner}
\usage{
get_github_owner(config_file, project_key_index)
}
\arguments{
\item{config_file}{The parsed config file}

\item{project_key_index}{The name of the index of the project key (e.g. "project_key_1" or "project_key_2")}
}
\value{
the GitHub project owner name for project specified by key index `project_key_index`
}
\description{
Gets the owner of the GitHub repository for a specific project key index
}
