% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_issue_id_regex}
\alias{get_issue_id_regex}
\title{Get Issue Id Regex}
\usage{
get_issue_id_regex(project_name)
}
\arguments{
\item{project_name}{the name of the project config file (e.g. "kaiaulu" or "geronimo")}
}
\value{
the issue id regex
}
\description{
Gets the issue ID regex on commit messages
}
