% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_jira_issues_path}
\alias{get_jira_issues_path}
\title{Get Jira Issues Path}
\usage{
get_jira_issues_path(project_name, project_key_index)
}
\arguments{
\item{project_name}{the name of the project config file (e.g. "kaiaulu" or "geronimo")}

\item{project_key_index}{the name of the index of the project key (e.g. "project_key_1" or "project_key_2")}
}
\value{
the Jira issue path
}
\description{
Gets the local path to store Jira issues for a specific project key index
}
