% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_jira_key_indexes}
\alias{get_jira_key_indexes}
\title{Get Jira Issue Tracker Project Key Indexes}
\usage{
get_jira_key_indexes(project_name)
}
\arguments{
\item{project_name}{the name of the project config file (e.g. "kaiaulu" or "geronimo")}
}
\value{
the list of issue tracker project keys
}
\description{
Gets the list of Jira issue tracker project key indexes
}
