% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_jira_keys}
\alias{get_jira_keys}
\title{Returns the list of Jira issue tracker project keys.}
\usage{
get_jira_keys(config_file)
}
\arguments{
\item{config_file}{The parsed configuration file obtained from \code{\link{parse_config}}.}
}
\value{
The list of Jira issue tracker project keys.
}
\description{
This function returns the list of Jira issue tracker project
keys, that is specified in the input parameter `config_file`. The input,
`config_file` must be a parsed configuration file. The function will inform
the user if the project keys exist in the parsed configuration
file, `config_file`.
}
