% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_jira_project_key_name}
\alias{get_jira_project_key_name}
\title{Get Jira Project Key Name}
\usage{
get_jira_project_key_name(project_name, project_key_index)
}
\arguments{
\item{project_name}{the name of the project config file (e.g. "kaiaulu" or "geronimo")}

\item{project_key_index}{the name of the index of the project key (e.g. "project_key_1" or "project_key_2")}
}
\value{
the project key name
}
\description{
Gets the project key name for a specific project key index
}
