% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_keep_dependencies_type}
\alias{get_keep_dependencies_type}
\title{Get Keep Dependencies Type}
\usage{
get_keep_dependencies_type(project_name)
}
\arguments{
\item{project_name}{the name of the project config file (e.g. "kaiaulu" or "geronimo")}
}
\value{
the types of dependencies to keep
}
\description{
Gets the type of dependencies to keep
}
