% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_mbox_key_indexes}
\alias{get_mbox_key_indexes}
\title{Get Mailing List Mod Mbox Key Indexes}
\usage{
get_mbox_key_indexes(config_file)
}
\arguments{
\item{config_file}{The parsed config file}
}
\value{
the list of mod Mbox mail key indexes
}
\description{
Gets the list of mod Mbox mail key indexes
}
