% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_mbox_mailing_list}
\alias{get_mbox_mailing_list}
\title{Get Mod Mbox Mailing List}
\usage{
get_mbox_mailing_list(config_file, project_key_index)

get_mbox_mailing_list(config_file, project_key_index)
}
\arguments{
\item{config_file}{The parsed config file}

\item{project_key_index}{The name of the index of the project key (e.g. "project_key_1" or "project_key_2")}
}
\value{
the Mbox mailing list for project specified by key index `project_key_index`

the Mbox archive type for project specified by key index `project_key_index`
}
\description{
Gets the Mbox mailing list for a specific project key index

Gets the Mbox archive type for a specific project key index
}
