% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_nvdfeed_folder_path}
\alias{get_nvdfeed_folder_path}
\title{Get Nvd Feed Folder Path}
\usage{
get_nvdfeed_folder_path(project_name)
}
\arguments{
\item{project_name}{the name of the project config file (e.g. "kaiaulu" or "geronimo")}
}
\value{
the folder path with nvd cve feeds
}
\description{
Gets the folder path with nvd cve feeds
}
