% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_pattern4_filepath}
\alias{get_pattern4_filepath}
\title{Get Pattern4 Output Filepath}
\usage{
get_pattern4_filepath(project_name)
}
\arguments{
\item{project_name}{the name of the project config file (e.g. "kaiaulu" or "geronimo")}
}
\value{
the output filepath
}
\description{
Gets the Pattern4 filepath where the output is stored
}
