% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_pattern4_folder_path}
\alias{get_pattern4_folder_path}
\title{Get Pattern4 Class Folder Path}
\usage{
get_pattern4_folder_path(config_file)
}
\arguments{
\item{config_file}{The parsed config file}
}
\value{
the Pattern4 class folder path
}
\description{
Gets the Pattern4 class folder path where the output is stored
}
