% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_prediction_dataset_filepath}
\alias{get_prediction_dataset_filepath}
\title{Returns sentiment classifier prediction dataset filepath}
\usage{
get_prediction_dataset_filepath(config_file)
}
\arguments{
\item{config_file}{A parsed configuration file obtained from \code{\link{parse_config}}.}
}
\value{
The folder path for sentiment predictions as a string.
}
\description{
This function retrieves the filepath where the prediction will be stored
for the sentiment classifier.
}
