% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_pysenti_conda_env_name}
\alias{get_pysenti_conda_env_name}
\title{Returns the path to conda}
\usage{
get_pysenti_conda_env_name(config_file)
}
\arguments{
\item{config_file}{A parsed configuration file obtained from \code{\link{parse_config}}.}
}
\value{
The file path to the PySenti script as a string.
}
\description{
This function retrieves the path to pysenti conda environment. Used
in conjunction to \code{\link{get_conda_path}} to invoke pysenti from outside the environment.
}
