% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_pysenti_path}
\alias{get_pysenti_path}
\title{Returns the path to the PySenti Python script}
\usage{
get_pysenti_path(config_file)
}
\arguments{
\item{config_file}{A parsed configuration file obtained from \code{\link{parse_config}}.}
}
\value{
The file path to the PySenti script as a string.
}
\description{
This function retrieves the path to the PySenti Python script 
from a parsed configuration file. The input `config_file` must be a parsed 
configuration file (from \code{\link{parse_config}}). If the attribute does 
not exist, a warning is issued.
}
