% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_replace_token_regex_with}
\alias{get_replace_token_regex_with}
\title{Returns the token regex replacements for text processing.}
\usage{
get_replace_token_regex_with(config_file)
}
\arguments{
\item{config_file}{The parsed configuration file.}
}
\value{
A named list of regex patterns for token replacement.
}
\description{
This function returns a named list of token regex patterns used to replace
certain patterns in text. The input `config_file` must be a parsed YAML configuration
file obtained from \code{\link{parse_config}}. The function will warn the user if
the attribute does not exist in the configuration file.
}
