% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_sentiment_prediction_folder_path}
\alias{get_sentiment_prediction_folder_path}
\title{Returns the folder path for storing sentiment predictions}
\usage{
get_sentiment_prediction_folder_path(config_file)
}
\arguments{
\item{config_file}{A parsed configuration file obtained from \code{\link{parse_config}}.}
}
\value{
The folder path for sentiment predictions as a string.
}
\description{
This function retrieves the path to the folder where sentiment 
predictions are saved, from a parsed configuration file. The input `config_file` 
must be a parsed configuration file. If the attribute does not exist, a warning 
is issued.
}
