% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_sentiment_training_dataset_filepath}
\alias{get_sentiment_training_dataset_filepath}
\title{Returns sentiment classifier training dataset filepath}
\usage{
get_sentiment_training_dataset_filepath(config_file)
}
\arguments{
\item{config_file}{A parsed configuration file obtained from \code{\link{parse_config}}.}
}
\value{
The folder path for sentiment predictions as a string.
}
\description{

}
