% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_src_folder}
\alias{get_src_folder}
\title{Returns the The path to the source code zip, folder or file of to be annotated with srcML}
\usage{
get_src_folder(config_file)
}
\arguments{
\item{config_file}{The parsed configuration file obtained from \code{\link{parse_config}}.}
}
\value{
The file path for the project to be annotated by srcML.
}
\description{
This function returns the file path to the
project to be annotated with srcML that is specified in the
input parameter `config_file`. The input, `config_file` must be a parsed
configuration file. The function will inform the user if the file path
exists in the parsed configuration file, `config_file`.
}
