% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_srcml_filepath}
\alias{get_srcml_filepath}
\title{Returns the file path for the output of the srcML analysis for the project.}
\usage{
get_srcml_filepath(config_file)
}
\arguments{
\item{config_file}{The parsed configuration file.}
}
\value{
The output file path for srcML analysis.
}
\description{
This function returns the file path to be used to store the
output of the srcML analysis for the project, that is specified in the
input parameter `config_file`. The input, `config_file` must be a parsed
configuration file. The function will inform the user if the file path
exists in the parsed configuration file, `config_file`.
}
