% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_substring_filepath}
\alias{get_substring_filepath}
\title{Get Substring Filepath}
\usage{
get_substring_filepath(project_name)
}
\arguments{
\item{project_name}{the name of the project config file (e.g. "kaiaulu" or "geronimo")}
}
\value{
the substring filepaths
}
\description{
Gets the substring filepaths for filtering
}
