% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_tools}
\alias{get_tools}
\title{Get Parsed Tools File}
\usage{
get_tools(tool_path = "tools.yml")
}
\arguments{
\item{tool_path}{The path to the tools file, defaulted at tools.yml}
}
\value{
the parsed tools file
}
\description{
Get Parsed Tools File
}
