% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_topics}
\alias{get_topics}
\title{Get Topics}
\usage{
get_topics(project_name)
}
\arguments{
\item{project_name}{the name of the project config file (e.g. "kaiaulu" or "geronimo")}
}
\value{
the topics
}
\description{
Gets the topics and keywords for analysis
}
