% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_trained_model_path}
\alias{get_trained_model_path}
\title{Returns the trained sentiment model classifier filepath}
\usage{
get_trained_model_path(config_file)
}
\arguments{
\item{config_file}{A parsed configuration file obtained from \code{\link{parse_config}}.}
}
\value{
The folder path for sentiment models as a string.
}
\description{

}
