% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_uctags_line_types}
\alias{get_uctags_line_types}
\title{Returns the types to keep to to be considered for analysis.}
\usage{
get_uctags_line_types(config_file)
}
\arguments{
\item{config_file}{The parsed configuration file obtained from \code{\link{parse_config}}.}
}
\value{
The lines type to keep for analysis.
}
\description{
This function returns the types of file-file dependencies that
should be considered, that are specified in the input parameter
`config_file`. The input, `config_file` must be a parsed configuration file.
The function will inform the user if the lines type to keep exists in the
parsed configuration file, `config_file`.
}
