% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_understand_code_language}
\alias{get_understand_code_language}
\title{Returns the understand code language for analysis.}
\usage{
get_understand_code_language(config_file)
}
\arguments{
\item{config_file}{The parsed configuration file obtained from \code{\link{parse_config}}.}
}
\value{
The code language for parsing with the understand tool.
}
\description{
This function returns the specified code language that should
be used to parse dependencies with the understand tool, that
is specified in the input parameter `config_file`. The input, `config_file`
must be a parsed configuration file. The function will inform the user if
the understand code language exists in the parsed configuration file, `config_file`.
}
