% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_window_end_commit}
\alias{get_window_end_commit}
\title{Get Window End Commit}
\usage{
get_window_end_commit(config_file)
}
\arguments{
\item{config_file}{The parsed config file}
}
\value{
the analysis window end commit
}
\description{
Gets the analysis window end commit
}
