% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_window_size}
\alias{get_window_size}
\title{Get Analysis Window Size}
\usage{
get_window_size(project_name)
}
\arguments{
\item{project_name}{the name of the project config file (e.g. "kaiaulu" or "geronimo")}
}
\value{
the analysis window size
}
\description{
Gets the analysis window size
}
