% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_add}
\alias{git_add}
\title{Git adds a file to the gitlog sample}
\usage{
git_add(git_repo, folder_path, filepath)
}
\arguments{
\item{git_repo}{The git repo path}

\item{folder_path}{The worktree path}

\item{filepath}{The filepath we want to add}
}
\value{
The path to the sample .git file.
}
\description{
This is a function that git adds the filepath selected
}
