% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_blame}
\alias{git_blame}
\title{Call git blame from cmd line and store the output in gitlog_path}
\usage{
git_blame(file_path)
}
\arguments{
\item{file_path}{The path of a file}
}
\value{
the blame file
}
\description{
Call git blame from cmd line and store the output in gitlog_path
}
