% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_blame_all}
\alias{git_blame_all}
\title{Call git blame from cmd line and store the output in gitlog_path}
\usage{
git_blame_all(file_path, line, since, until)
}
\arguments{
\item{file_path}{The path of a file}
}
\value{
the blame of a single line with all the commits that touched the lines
}
\description{
Call git blame from cmd line and store the output in gitlog_path
}
