% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_checkout}
\alias{git_checkout}
\title{Performs a git checkout on specified repo}
\usage{
git_checkout(commit_hash, git_repo_path, new_branch = FALSE)
}
\arguments{
\item{commit_hash}{The commit hash the repo should be checkout}

\item{git_repo_path}{The git repo path}

\item{new_branch}{Boolean indicating if we want to create a new branch (default is FALSE)}
}
\value{
Any error message generated by git
}
\description{
Performs a git checkout on specified repo
}
