% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_commit}
\alias{git_commit}
\title{Git commit}
\usage{
git_commit(git_repo, folder_path, commit_msg, author, email)
}
\arguments{
\item{git_repo}{The git repo path}

\item{folder_path}{The worktree path}

\item{filepath}{The filepath we want to add}
}
\value{
The path to the sample .git file.
}
\description{
Git commits a file
}
