% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_create_sample_log}
\alias{git_create_sample_log}
\title{Creates a sample git log with one commit}
\usage{
git_create_sample_log(folder_path = "/tmp")
}
\arguments{
\item{folder_path}{An optional path to where the sample .git should be created.}
}
\value{
The path to the sample .git file.
}
\description{
This is a SetUp helper function for Kaiaulu unit tests
that manipulates git logs.
}
\details{
A folder kaiaulu_sample is created in /tmp by default. A file,
hello.R with a single print is then added to the folder.
Git init is performed, the file is git add, and commit to
the git log.
}
\seealso{
Other unittest:
\code{\link[=git_delete_sample_log]{git_delete_sample_log()}}
}
\concept{unittest}
