% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_init}
\alias{git_init}
\title{This function initializes a new Git repository in the specified folder.}
\usage{
git_init(folder_path)
}
\arguments{
\item{folder_path}{The path to the folder where the Git repository should be initialized.}
}
\value{
The path to the newly created Git repository.
}
\description{
This function initializes a new Git repository in the specified folder.
}
