% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_log}
\alias{git_log}
\title{Saves gitlog to a path}
\usage{
git_log(git_repo_path, flags, save_path)
}
\arguments{
\item{git_repo_path}{The git repo path}

\item{flags}{Optional flags for git log command}

\item{save_path}{the filepath to save the file}
}
\description{
Saves the `.git` of a github repository as a gitlog at the specified path
}
