% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_rename}
\alias{git_rename}
\title{This function renames a folder or file}
\usage{
git_rename(git_repo, folder_path, old_name, new_name)
}
\arguments{
\item{git_repo}{The git repo path}

\item{folder_path}{The worktree path}

\item{old_name}{The name of the file/folder that you are going to change or move}

\item{new_name}{The new name of the file/folder}
}
\description{
This function renames a folder or file
}
