% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_api_discussions}
\alias{github_api_discussions}
\title{Download Discussions}
\usage{
github_api_discussions(token, owner, repo)
}
\arguments{
\item{token}{Your Github API token}

\item{owner}{Github's repository owner (e.g. sailuh)}

\item{repo}{Github's repository name (e.g. kaiaulu)}
}
\description{
Download Discussions from GraphQL endpoint.
}
