% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_api_issue_pr_comments}
\alias{github_api_issue_pr_comments}
\title{Download Project Issue's or Pull Request's Comments by specific pull request number.}
\usage{
github_api_issue_pr_comments(owner, repo, issue_number, token)
}
\arguments{
\item{owner}{GitHub's repository owner (e.g. sailuh)}

\item{repo}{GitHub's repository name (e.g. kaiaulu)}

\item{issue_number}{The number that identifies the issue}

\item{token}{Your GitHub API token}
}
\description{
Download Issues' and Pull Request's Comments from "GET /repos/{owner}/{repo}/issues/comments" endpoint.
All Pull Requests are counted as Issues, but not all Issues are Pull Requests
Optional parameter since is used to download comments updated after the specified date.
If the value of since is NULL, it is not passed to the API call and all comments are downloaded.
}
\references{
For details, see \url{https://docs.github.com/en/rest/issues/comments?apiVersion=2022-11-28#list-issue-comments} and
\url{https://docs.github.com/en/rest/guides/working-with-comments#pull-request-comments}.
}
