% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_api_page_prev}
\alias{github_api_page_prev}
\title{Obtain the previous GitHub response page.}
\usage{
github_api_page_prev(gh_response)
}
\arguments{
\item{gh_response}{A response returned by any GitHub endpoint which is paginated (e.g. \code{\link{github_api_project_commits}}).}
}
\description{
Obtain the previous GitHub response page.
}
\keyword{internal}
