% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_api_pr_commits}
\alias{github_api_pr_commits}
\title{Download Pull Request Commits}
\usage{
github_api_pr_commits(owner, repo, pull_number, token)
}
\arguments{
\item{owner}{GitHub's repository owner (e.g. sailuh)}

\item{repo}{GitHub's repository name (e.g. kaiaulu)}

\item{token}{Your GitHub API token}
}
\description{
Download Pull Request Commits from "GET /repos/{owner}/{repo}/pulls/{pull_number}/commits" endpoint.
}
\references{
For details, see \url{https://docs.github.com/en/free-pro-team@latest/rest/reference/repos#list-commits-on-a-pull-request}.
}
