% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_api_pr_reviewers}
\alias{github_api_pr_reviewers}
\title{Download Pull Request Requested Reviewers}
\usage{
github_api_pr_reviewers(owner, repo, pull_number, token)
}
\arguments{
\item{owner}{GitHub's repository owner (e.g. sailuh)}

\item{repo}{GitHub's repository name (e.g. kaiaulu)}

\item{pull_number}{The number that identifies the pull request.}

\item{token}{Your GitHub API token}
}
\description{
Download Pull Request Requested Reviewers from "GET /repos/{owner}/{repo}/pulls/{pull_number}/requested_reviewers" endpoint.
}
\references{
For details, see \url{https://docs.github.com/en/free-pro-team@latest/rest/reference/repos#get-all-requested-reviewers-for-a-pull-request}.
}
