% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_api_pr_reviews}
\alias{github_api_pr_reviews}
\title{Download Pull Request Reviews}
\usage{
github_api_pr_reviews(owner, repo, pull_number, token)
}
\arguments{
\item{owner}{GitHub's repository owner (e.g. sailuh)}

\item{repo}{GitHub's repository name (e.g. kaiaulu)}

\item{pull_number}{Required pull request number to make API request}

\item{token}{Your GitHub API token}
}
\description{
Download Pull Request Reviews from "GET /repos/{owner}/{repo}/pulls/{pull_number}/reviews" endpoint.
}
\references{
For details, see \url{https://docs.github.com/en/free-pro-team@latest/rest/reference/repos#list-reviews-for-a-pull-request}.
}
