% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_api_project_commits}
\alias{github_api_project_commits}
\title{Download Project Commits}
\usage{
github_api_project_commits(owner, repo, token)
}
\arguments{
\item{owner}{GitHub's repository owner (e.g. sailuh)}

\item{repo}{GitHub's repository name (e.g. kaiaulu)}

\item{token}{Your GitHub API token}
}
\description{
Download Commits from "GET /repos/{owner}/{repo}/commits" endpoint.
Differently from parsing commits by git cloning the repository, this JSON provides
the GitHub user id, which allows for linking file changes and issue events by the
same author without relying on identity matching heuristics.
}
\references{
For details, see \url{https://docs.github.com/en/rest/reference/repos#commits}.
}
