% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_api_project_contributors}
\alias{github_api_project_contributors}
\title{Download Project Contributors}
\usage{
github_api_project_contributors(owner, repo, token)
}
\arguments{
\item{owner}{GitHub's repository owner (e.g. sailuh)}

\item{repo}{GitHub's repository name (e.g. kaiaulu)}

\item{token}{Your GitHub API token}
}
\description{
Download project contributors from GET /repos/{owner}/{repo}/contributors" endpoint.
}
\references{
For more details see \url{https://docs.github.com/en/free-pro-team@latest/rest/reference/repos#list-repository-contributors}.
}
