% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_api_project_issue_or_pr_comment_refresh}
\alias{github_api_project_issue_or_pr_comment_refresh}
\title{Download Project issues or pr comments after certain date}
\usage{
github_api_project_issue_or_pr_comment_refresh(owner, repo, token, created)
}
\arguments{
\item{owner}{GitHub's repository owner (e.g. sailuh)}

\item{repo}{GitHub's repository name (e.g. kaiaulu)}

\item{token}{Your GitHub API token}

\item{created}{Only issues created after this date are fetched}
}
\description{
Download  Issues from "GET /repos/{owner}/{repo}/issues" endpoint.
}
\references{
For details, see \url{https://docs.github.com/en/rest/reference/issues#list-repository-issues}.
}
