% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_api_project_pull_request}
\alias{github_api_project_pull_request}
\title{Download Project Pull Requests}
\usage{
github_api_project_pull_request(owner, repo, token)
}
\arguments{
\item{owner}{GitHub's repository owner (e.g. sailuh)}

\item{repo}{GitHub's repository name (e.g. kaiaulu)}

\item{token}{Your GitHub API token}
}
\description{
Download  Pull Requests from "GET /repos/{owner}/{repo}/pulls" endpoint.
}
\references{
For details, see \url{https://docs.github.com/en/rest/reference/pulls#list-pull-requests}.
}
