% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_api_project_pull_request_review_refresh}
\alias{github_api_project_pull_request_review_refresh}
\title{Download Project Pull Request Reviews Refresh}
\usage{
github_api_project_pull_request_review_refresh(
  owner,
  repo,
  token,
  save_path_pull_request,
  save_path_pr_reviews,
  verbose = TRUE
)
}
\arguments{
\item{owner}{GitHub's repository owner (e.g. sailuh)}

\item{repo}{GitHub's repository name (e.g. kaiaulu)}

\item{token}{Your GitHub API token}

\item{save_path_pull_request}{save path from data obtained by \code{\link{github_api_project_pull_request_refresh}}.}

\item{save_path_pr_reviews}{the save path for the pull request reviews folder to be downloaded}

\item{verbose}{True if messages should be displayed during execution.}
}
\description{
Download Pull Request Reviews based on downloaded pull requests downloaded by \code{\link{github_api_project_pull_request_refresh}}.
The data dependency is due to GitHub API requiring a specific pull request id to complete the API Request.
}
\references{
For details, see For details, see \url{https://docs.github.com/en/rest/pulls/comments?apiVersion=2022-11-28#about-pull-request-review-comments}.
}
\seealso{
\code{link{github_api_pull_request_reviews}} to download pull request review data.

\code{link{github_api_project_pull_request}} to download pull request data and retrieve pull numbers.
}
