% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_api_pull_pr_comments}
\alias{github_api_pull_pr_comments}
\title{Download Project's Pull Request Comments by Specific Pull Request Number}
\usage{
github_api_pull_pr_comments(owner, repo, pull_number, token)
}
\arguments{
\item{owner}{GitHub's repository owner (e.g. sailuh)}

\item{repo}{GitHub's repository name (e.g. kaiaulu)}

\item{token}{Your GitHub API token}

\item{since}{Optional parameter to specify pulling only comments updated after this date}
}
\description{
Downloads the Pull Request Comments from `GET /repos/{owner}/{repo}/pulls/comments` endpoint.
Optional parameter since is used to download comments updated after the specified date.
If the value of since is NULL, it is not passed to the API call and all comments are downloaded.
NOTE: This function is different from the `github_api_project_issue_or_pr_comments`.
}
\references{
For details, see \url{https://docs.github.com/en/rest/pulls/comments?apiVersion=2022-11-28#about-pull-request-review-comments}
}
