% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_api_pull_request}
\alias{github_api_pull_request}
\title{Download Specific Pull Requests}
\usage{
github_api_pull_request(owner, repo, pull_number, token)
}
\arguments{
\item{owner}{GitHub's repository owner (e.g. sailuh)}

\item{repo}{GitHub's repository name (e.g. kaiaulu)}

\item{pull_number}{The number that identifies the pull request.}

\item{token}{Your GitHub API token}
}
\description{
Download  Pull Requests from "GET /repos/{owner}/{repo}/pulls/{pull_number}" endpoint.
}
\references{
For details, see \url{https://docs.github.com/en/rest/reference/pulls#list-pull-requests}.
}
