% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_api_rate_limit}
\alias{github_api_rate_limit}
\title{Returns token remaining available requests.}
\usage{
github_api_rate_limit(token)
}
\arguments{
\item{token}{Your GitHub API token}
}
\description{
Returns token remaining available requests.
}
\references{
For more details see \url{https://docs.github.com/en/free-pro-team@latest/rest/overview/resources-in-the-rest-api#rate-limiting}.
}
