% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_parse_discussion_comments}
\alias{github_parse_discussion_comments}
\title{Parse Discussions JSON to Table}
\usage{
github_parse_discussion_comments(api_response)
}
\arguments{
\item{api_response}{API response obtained from `github_api_discussions` function.}
}
\value{
The parsed table of Discussion Comments.
}
\description{
This function parses through the JSON of Github Discussions
downloaded from the `github_api_discussions` function, and turns it into
a table. This function only parses through the Discussion comments within
the JSON, and does not parse the discussion themselves.
}
