% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_parse_discussions}
\alias{github_parse_discussions}
\title{Parse Discussions JSON to Table}
\usage{
github_parse_discussions(api_response)
}
\arguments{
\item{api_responses}{API response obtained from `github_api_discussions` function.}
}
\value{
The parsed table on the discussions.
}
\description{
This function parses through the JSON of Github Discussions
downloaded from the `github_api_discussions` function, and turns it into
a table. This function does not parse the data for the discussion comments.
}
