% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_parse_project_commit_comments}
\alias{github_parse_project_commit_comments}
\title{Parse Commit Comments JSON to Table}
\usage{
github_parse_project_commit_comments(api_responses)
}
\arguments{
\item{api_responses}{API response obtained from \code{\link{github_api_project_commit_comments}}.}
}
\description{
Parses the response from \code{\link{github_api_project_commit_comments}} into a data.table.
Note not all columns available in the downloaded json are parsed.
}
