% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_parse_project_pr_merge}
\alias{github_parse_project_pr_merge}
\title{Parse Pull Requests' Merge Status JSON to Table}
\usage{
github_parse_project_pr_merge(api_responses)
}
\arguments{
\item{api_responses}{API response obtained from github_api_* function.}
}
\description{
Note not all columns available in the downloaded json are parsed.
This function only checks for the merge status for a pull request.
}
