% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_parse_project_pull_request_reviews}
\alias{github_parse_project_pull_request_reviews}
\title{Parse Pull Requests' Reviews JSON to Table}
\usage{
github_parse_project_pull_request_reviews(api_responses)
}
\arguments{
\item{api_responses}{API response obtained from \code{\link{github_api_project_pull_request_review_refresh}}.}
}
\description{
Parse pull request data obtained by \code{\link{github_api_project_pull_request_review_refresh}}.
}
\details{
Note not all columns available in the downloaded json are parsed.
}
