% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{io_create_folder}
\alias{io_create_folder}
\title{This function creates a new folder}
\usage{
io_create_folder(folder_name, folder_path = "/tmp")
}
\arguments{
\item{folder_name}{Name of newly created folder}

\item{folder_path}{Path of folder where new folder will be, default set to /tmp}
}
\value{
The new folder_path with its updated contents
}
\description{
This function creates a new folder
}
