% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{io_delete_folder}
\alias{io_delete_folder}
\title{Delete Folder}
\usage{
io_delete_folder(folder_path = "/tmp", folder_name)
}
\arguments{
\item{folder_path}{Path of folder where the folder to be deleted is. Default set to /tmp}

\item{folder_name}{Name of folder to be deleted}
}
\value{
The deleted folder_path
}
\description{
Deletes a folder on disk.
}
