% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{io_make_file}
\alias{io_make_file}
\title{Make File}
\usage{
io_make_file(file_path, body)
}
\arguments{
\item{file_path}{The path of the file that is being saved}

\item{body}{The contents of the file}
}
\value{
file with specified path and body
}
\description{
This function creates a local file on disk with
the specified content.
}
