% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{io_make_folder}
\alias{io_make_folder}
\title{Make Folder}
\usage{
io_make_folder(folder_path = "/tmp", folder_name)
}
\arguments{
\item{folder_path}{Path of folder where new folder will be, default set to /tmp}

\item{folder_name}{Name of newly created folder}
}
\value{
The new folder_path with its updated contents
}
\description{
Creates a new folder on disk.
}
