% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identity.R
\name{is_same_identity}
\alias{is_same_identity}
\title{Compares identities in a list by index}
\usage{
is_same_identity(i, j, parsed_name_email, use_name_only = FALSE)
}
\arguments{
\item{i}{Index of the first identity}

\item{j}{Index of the second identity}

\item{parsed_name_email}{The list of name and email identities}

\item{use_name_only}{Uses only the name field (i.e. disregard the e-mail field)}
}
\description{
Compares identities in a list by index
}
