% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{jira_create_sample_log_issues}
\alias{jira_create_sample_log_issues}
\title{Creates a sample jira issue without comment and two components}
\usage{
jira_create_sample_log_issues(folder_path = "/tmp")
}
\arguments{
\item{folder_path}{An optional path to where the sample .json should be created.}
}
\value{
The path to the sample .json file.
}
\description{
This example replicates Kaiaulu issue #244
}
\seealso{
Other {unittest}: 
\code{\link{git_create_sample_log}()},
\code{\link{git_delete_sample_log}()},
\code{\link{jira_create_sample_log_issues_comments}()},
\code{\link{jira_delete_sample_log}()}
}
\concept{{unittest}}
