% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{jira_create_sample_log_issues_comments}
\alias{jira_create_sample_log_issues_comments}
\title{Create Jira Sample Log Issues with comments}
\usage{
jira_create_sample_log_issues_comments(folder_path = "/tmp")
}
\arguments{
\item{folder_path}{An optional path to where the sample .json should be created.}
}
\value{
The path to the sample .json file.
}
\description{
Creates a sample jira issue with two components WITH comments
}
\seealso{
Other {unittest}: 
\code{\link{git_create_sample_log}()},
\code{\link{git_delete_sample_log}()},
\code{\link{jira_create_sample_log_issues}()},
\code{\link{jira_delete_sample_log}()}
}
\concept{{unittest}}
