% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{list_files_in_repo}
\alias{list_files_in_repo}
\title{Stores the path of each file in a repository in a text file in /tmp/files.txt}
\usage{
list_files_in_repo(git_repo_path)
}
\arguments{
\item{git_repo_path}{The git repo path}
}
\value{
A list containing all the files
}
\description{
Stores the path of each file in a repository in a text file in /tmp/files.txt
}
