% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{make_jira_issue_tracker}
\alias{make_jira_issue_tracker}
\title{Make Jira Issue Tracker}
\usage{
make_jira_issue_tracker(issues)
}
\arguments{
\item{issues}{list of issues that will make up the issue tracker}
}
\value{
A list named 'status' containing status's information
}
\description{
Create a full JIRA Issue Tracker via
}
