% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\name{make_mbox_mailing_list}
\alias{make_mbox_mailing_list}
\title{Takes in mbox replies and creates a .mbox file}
\usage{
make_mbox_mailing_list(replies, folder_path = "/tmp", file_name)
}
\arguments{
\item{replies}{An array of replies that have been created with \code{\link{make_mbox_reply}}}

\item{folder_path}{Folder path for the .mbox file being created. Defaulted at /tmp}

\item{file_name}{Name of the file that will store the .mbox file}
}
\value{
the path of the .mbox file that was created
}
\description{
Takes a list of mbox replies generated with create_fake_mbox_replies function and compiles them all into a single
fake .mbox file
}
