% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{make_temporary_file}
\alias{make_temporary_file}
\title{Create Temporary File}
\usage{
make_temporary_file(content, extension)
}
\arguments{
\item{content}{Textual content to be added to the file}

\item{extension}{File extension (e.g. .c, .txt, .md, etc)}
}
\description{
Create Temporary File
}
