% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{motif_factory_square}
\alias{motif_factory_square}
\title{Square Motif}
\usage{
motif_factory_square()
}
\description{
Square motifs represent two developers who communicate when
changing two different files that have a dependency.
This is seeing as a "positive" motif: We expect
developers will communicate when changing different files which
have a dependency. To obtain
the project's network for this motif, refer to
\code{\link{transform_gitlog_to_bipartite_network}},
\code{\link{transform_reply_to_bipartite_network}}, and
\code{\link{transform_r_dependencies_to_network}} or
\code{\link{transform_dependencies_to_network}}.
}
\references{
W. Mauerer, M. Joblin, D. A. Tamburri, C. Paradis, R. Kazman and S. Apel,
"In Search of Socio-Technical Congruence: A Large-Scale Longitudinal Study,"
in IEEE Transactions on Software Engineering, vol. 48, no. 8, pp. 3159-3184,
1 Aug. 2022, doi: 10.1109/TSE.2021.3082074.
}
\seealso{
Other motif:
\code{\link[=motif_factory]{motif_factory()}},
\code{\link[=motif_factory_anti_square]{motif_factory_anti_square()}},
\code{\link[=motif_factory_anti_triangle]{motif_factory_anti_triangle()}},
\code{\link[=motif_factory_triangle]{motif_factory_triangle()}}
}
\concept{motif}
