% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{motif_factory_triangle}
\alias{motif_factory_triangle}
\title{Triangle Motif}
\usage{
motif_factory_triangle()
}
\description{
Triangle motifs represent two developers who communicate when
changing a file. This is seeing as a "positive" motif: We expect
developers will communicate when changing the same file. To obtain
the project's network for this motif, refer to
\code{\link{transform_gitlog_to_bipartite_network}} and
\code{\link{transform_reply_to_bipartite_network}}.
}
\references{
W. Mauerer, M. Joblin, D. A. Tamburri, C. Paradis, R. Kazman and S. Apel,
"In Search of Socio-Technical Congruence: A Large-Scale Longitudinal Study,"
in IEEE Transactions on Software Engineering, vol. 48, no. 8, pp. 3159-3184,
1 Aug. 2022, doi: 10.1109/TSE.2021.3082074.
}
\seealso{
Other motif:
\code{\link[=motif_factory]{motif_factory()}},
\code{\link[=motif_factory_anti_square]{motif_factory_anti_square()}},
\code{\link[=motif_factory_anti_triangle]{motif_factory_anti_triangle()}},
\code{\link[=motif_factory_square]{motif_factory_square()}}
}
\concept{motif}
