% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identity.R
\name{normalized_levenshtein}
\alias{normalized_levenshtein}
\title{Edit Distance of two words}
\usage{
normalized_levenshtein(a, b)
}
\arguments{
\item{a}{a string}

\item{b}{a string}
}
\description{
Edit Distance of two words
}
\keyword{internal}
