% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{openhub_api_iterate_pages}
\alias{openhub_api_iterate_pages}
\title{OpenHub Page Iterator}
\usage{
openhub_api_iterate_pages(
  token,
  openhub_api_function,
  save_folder_path,
  openhub_api_function_parameter,
  max_pages = NULL
)
}
\arguments{
\item{token}{Your OpenHub API token.}

\item{openhub_api_function}{A function that downloads a page of a specific XML response file (e.g. \code{\link{openhub_api_portfolio_projects}}).}

\item{save_folder_path}{A folder path to save the downloaded XML pages "as-is".}

\item{openhub_api_function_parameter}{Required unique parameter for use in `openhub_api_function` (`org_name` in \code{\link{openhub_api_portfolio_projects}} or `project_name` in \code{\link{openhub_api_projects}}).}

\item{max_pages}{The maximum number of pages to download, if NULL, maximum number of pages will be used, and if max_pages exceeds the maximum number of pages, it will use the maximum number of pages (Default set to NULL).}
}
\description{
Ohloh API endpoints return data in pages, each containing a
set number of items. This iterator can be used to iterate over the pages in
the collection that corresponds to `openhub_api_function`. `max_pages` can
be used to define the maximum number of pages to iterate through, otherwise,
the maximum number of pages will be iterated.
}
