% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{openhub_api_organizations}
\alias{openhub_api_organizations}
\title{placeholder}
\usage{
openhub_api_organizations(token, openhub_api_parameters, page = 1)
}
\arguments{
\item{token}{placeholder}

\item{openhub_api_parameters}{placeholder}

\item{page}{placeholder}
}
\value{
placeholder
}
\description{
placeholder
}
