% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{openhub_api_portfolio_projects}
\alias{openhub_api_portfolio_projects}
\title{Downloads a page of Portfolio Projects XML Response File.}
\usage{
openhub_api_portfolio_projects(token, openhub_api_parameters, page = 1)
}
\arguments{
\item{token}{Your OpenHub API token.}

\item{openhub_api_parameters}{List of parameters; must contain "portfolio_project_site" key-value pair (e.g. "https://openhub.net/orgs/apache/projects.xml").}

\item{page}{The page number to index (Default set to 1) (e.g. 1).}
}
\value{
The response XML single page file that contains a list of portfolio project items.
}
\description{
Downloads an XML response file from the "portfolio_projects" collection
endpoint. The `openhub_api_parameters` list must contain a "portfolio_project_site"
key-value pair for use as part of the URL. The XML response file will return a page
of items which can be indexed by the numeric `page` parameter.
}
\references{
For portfolio projects collection details, see \url{https://github.com/blackducksoftware/ohloh_api/blob/main/reference/portfolio_projects.md}.
}
