% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{openhub_download}
\alias{openhub_download}
\title{OpenHub API Response Downloader}
\usage{
openhub_download(
  api_response,
  save_folder_path,
  timestamp,
  type,
  unique_information = NULL,
  page = NULL
)
}
\arguments{
\item{api_response}{A single XML response file obtained from an openhub_api_* function.}

\item{save_folder_path}{A folder path to save the downloaded XML response file "as-is".}

\item{type}{The type of XML response file ("organization", "portfolio", "project", or "analysis").}

\item{unique_information}{Identifying information about the type of XML response (Default set to NULL).}

\item{page}{The page number of the response file (Default set to NULL).}
}
\description{
Stores a XML response file `api_response` into a defined folder
path `save_folder_path`. This function sanitizes the `unique_information`
parameter, if not NULL, (`html_url_or_name` in
\code{\link{openhub_api_organizations}}, `org_name` in
\code{\link{openhub_api_portfolio_projects}}, `project_name` in
\code{\link{openhub_api_projects}}, and `project_name` in
\code{\link{openhub_api_analyses}}) for use in the file name
(e.g. "Apache (Java_Test)" becomes "apachejavatest").
}
