% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{openhub_retrieve}
\alias{openhub_retrieve}
\title{OpenHub Parser Helper}
\usage{
openhub_retrieve(folder_path)
}
\arguments{
\item{folder_path}{A folder path to retrieve XML files.}
}
\value{
A list of XML responses as files stored in `folder_path`.
}
\description{
A helper function for `openhub_parse_*` functions to retrieve
XML responses as files from a specified folder path `folder_path`.
}
