% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_bugzilla}
\alias{parse_bugzilla}
\title{Parse Bugzilla data obtained from Perceval traditional Bugzilla backend}
\usage{
parse_bugzilla(bugzilla_json, comments = FALSE)
}
\arguments{
\item{bugzilla_json}{json object containing Bugzilla data}

\item{comments}{if true, the comments are parsed along with the issues}
}
\value{
data table with parsed bugzilla data
}
\description{
Parse Bugzilla data obtained from Perceval traditional Bugzilla backend
}
\seealso{
\code{\link{download_bugzilla}} a downloader function to download bugzilla data with perceval

Other parsers: 
\code{\link{parse_bugzillarest}()},
\code{\link{parse_commit_message_id}()},
\code{\link{parse_dependencies}()},
\code{\link{parse_gitlog}()},
\code{\link{parse_jira}()},
\code{\link{parse_mbox}()},
\code{\link{parse_nvdfeed}()}
}
\concept{parsers}
