% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_bugzilla_crawler_comments}
\alias{parse_bugzilla_crawler_comments}
\title{Parse Bugzilla comments data obtained from json files from Bugzilla crawler \code{\link{download_bugzilla_comments_from_rest_api}}}
\usage{
parse_bugzilla_crawler_comments(comments_folder_path)
}
\arguments{
\item{comments_folder_path}{path to the comments folder that contains json file with Bugzilla data inside}
}
\description{
Parse Bugzilla comments data obtained from json files from Bugzilla crawler \code{\link{download_bugzilla_comments_from_rest_api}}
}
\seealso{
Other parsers: 
\code{\link{dv8_depends_to_sdsmj}()},
\code{\link{parse_bugzilla_crawler_issue}()},
\code{\link{parse_commit_message_id}()},
\code{\link{parse_dependencies}()},
\code{\link{parse_dv8_clusters}()},
\code{\link{parse_gitlog}()},
\code{\link{parse_jira}()},
\code{\link{parse_mbox}()},
\code{\link{parse_nvdfeed}()}
}
\concept{parsers}
