% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_commit_message_id_network}
\alias{parse_commit_message_id_network}
\title{Transform parsed git repo commit messages id and files into an edgelist}
\usage{
parse_commit_message_id_network(project_git, commit_message_id_regex)
}
\arguments{
\item{project_git}{A parsed git project by \code{parse_gitlog}.}

\item{commit_message_id_regex}{the regex to extract the id from the commit message}
}
\description{
Transform parsed git repo commit messages id and files into an edgelist
}
\seealso{
Other edgelists: 
\code{\link{parse_dependencies_network}()},
\code{\link{parse_gitlog_network}()},
\code{\link{parse_mbox_network}()}
}
\concept{edgelists}
