% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{parse_dependencies}
\alias{parse_dependencies}
\title{Parse dependencies from Depends}
\usage{
parse_dependencies(
  depends_jar_path,
  git_repo_path,
  language,
  output_dir = "/tmp/"
)
}
\arguments{
\item{depends_jar_path}{path to depends jar}

\item{git_repo_path}{path to git repo (ends in .git)}

\item{language}{the language of the .git repo (accepts cpp, java, ruby, python, pom)}

\item{output_dir}{path to output directory (formatted output_path/)}
}
\description{
Parse dependencies from Depends
}
\seealso{
Other parsers:
\code{\link[=build_understand_project]{build_understand_project()}},
\code{\link[=export_understand_dependencies]{export_understand_dependencies()}},
\code{\link[=parse_bugzilla_perceval_rest_issue_comments]{parse_bugzilla_perceval_rest_issue_comments()}},
\code{\link[=parse_bugzilla_perceval_traditional_issue_comments]{parse_bugzilla_perceval_traditional_issue_comments()}},
\code{\link[=parse_bugzilla_rest_comments]{parse_bugzilla_rest_comments()}},
\code{\link[=parse_bugzilla_rest_issues]{parse_bugzilla_rest_issues()}},
\code{\link[=parse_bugzilla_rest_issues_comments]{parse_bugzilla_rest_issues_comments()}},
\code{\link[=parse_commit_message_id]{parse_commit_message_id()}},
\code{\link[=parse_dv8_clusters]{parse_dv8_clusters()}},
\code{\link[=parse_github_latest_date]{parse_github_latest_date()}},
\code{\link[=parse_gitlog]{parse_gitlog()}},
\code{\link[=parse_jira]{parse_jira()}},
\code{\link[=parse_jira_latest_date]{parse_jira_latest_date()}},
\code{\link[=parse_jira_rss_xml]{parse_jira_rss_xml()}},
\code{\link[=parse_mbox]{parse_mbox()}},
\code{\link[=parse_mbox_latest_date]{parse_mbox_latest_date()}},
\code{\link[=parse_nvdfeed]{parse_nvdfeed()}},
\code{\link[=parse_understand_dependencies]{parse_understand_dependencies()}}
}
\concept{parsers}
