% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_dependencies_network}
\alias{parse_dependencies_network}
\title{Transform parsed dependencies into a network}
\usage{
parse_dependencies_network(depends_parsed, weight_types = NA)
}
\arguments{
\item{weight_types}{The weight types as defined in Depends.}

\item{project_mbox}{A parsed mbox by \code{parse_dependencies}.}
}
\description{
Transform parsed dependencies into a network
}
\seealso{
Other edgelists: 
\code{\link{parse_gitlog_network}()},
\code{\link{parse_mbox_network}()}
}
\concept{edgelists}
