% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv8.R
\name{parse_dv8_metrics_decoupling_level}
\alias{parse_dv8_metrics_decoupling_level}
\title{Parses the decoupling level JSON file}
\usage{
parse_dv8_metrics_decoupling_level(dl_path)
}
\arguments{
\item{dl_path}{path to decoupling metric json file created by \code{\link{dv8_mdsmb_to_decoupling_level}}}
}
\value{
data.table object with decoupling metrics
}
\description{
Creates a data.table object from a *.json file
}
