% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_git_blame}
\alias{parse_git_blame}
\title{Parse the git blame message of a file}
\usage{
parse_git_blame(git_repo_path, flags, commit_hash, file_path)
}
\arguments{
\item{git_repo_path}{git_repo_path path to git repo (ends in .git)}

\item{flags}{the parameters for invoking git blame (https://git-scm.com/docs/git-blame)}

\item{commit_hash}{a commit hash which indicates the specific version of the file (the commit must exist in `git_log`)}
}
\value{
a data.table which contains blame commits for each line of a file and metadata of the commits.
}
\description{
Create a data.table with the blame data of each line of a file in a specific commit.
}
