% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_github_replies}
\alias{parse_github_replies}
\title{Parse GitHub Issue and Pull Request Comments}
\usage{
parse_github_replies(github_replies_folder_path)
}
\arguments{
\item{issues_json_folder_path}{The folder path to the issues json obtained from \code{\link{github_api_project_issue}}.}

\item{pull_requests_json_folder_path}{The folder path to the pull requests json obtained from \code{\link{github_api_project_pull_request}}.}

\item{comments_json_folder_path}{The folder path to the comments json obtained from \code{\link{github_api_project_issue_or_pr_comments}}.}
}
\value{
A single reply table which combines the communication from the three jsons.
}
\description{
Parses Issue, Pull Request, and Comments Endpoints into a reply table.
See example usage in the download_github_comments.Rmd vignette.
}
