% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{parse_gitlog_entity_temporal_network}
\alias{parse_gitlog_entity_temporal_network}
\title{Create time-ordered contribution network}
\usage{
parse_gitlog_entity_temporal_network(
  project_git_entity,
  mode = c("author", "identity")
)
}
\arguments{
\item{project_git_entity}{A parsed git project by \code{parse_gitlog_entity}.}
}
\description{
Create a collaboration network as described by Joblin et al.
where an edge from developer A to developer B is created if A modifies a
file, and B modifies it chronologically immediately after. This implementation
matches the one defined by Joblin et al.
}
\references{
M. Joblin, W. Mauerer, S. Apel,
J. Siegmund and D. Riehle, "From Developer Networks
to Verified Communities: A Fine-Grained Approach,"
2015 IEEE/ACM 37th IEEE International Conference on
Software Engineering, Florence, 2015, pp. 563-573,
doi: 10.1109/ICSE.2015.73.
}
\seealso{
Other edgelists: 
\code{\link{parse_commit_message_id_network}()},
\code{\link{parse_cve_cwe_file_network}()},
\code{\link{parse_dependencies_network}()},
\code{\link{parse_gitlog_entity_network}()},
\code{\link{parse_gitlog_network}()},
\code{\link{parse_gitlog_temporal_network}()},
\code{\link{parse_mbox_network}()}
}
\concept{edgelists}
