% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{parse_gitlog_network}
\alias{parse_gitlog_network}
\title{Transform parsed git repo into an edgelist}
\usage{
parse_gitlog_network(
  project_git,
  mode = c("author", "commit", "author-committer")
)
}
\arguments{
\item{project_git}{A parsed git project by \code{parse_gitlog}.}

\item{mode}{The network of interest: author-file, commit-file, or author-comitter}
}
\description{
Transform parsed git repo into an edgelist
}
\seealso{
Other edgelists: 
\code{\link{parse_commit_message_id_network}()},
\code{\link{parse_cve_cwe_file_network}()},
\code{\link{parse_dependencies_network}()},
\code{\link{parse_mbox_network}()}
}
\concept{edgelists}
