% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{parse_gitlog_temporal_network}
\alias{parse_gitlog_temporal_network}
\title{Create time-ordered contribution network}
\usage{
parse_gitlog_temporal_network(project_git, mode = c("author", "identity"))
}
\arguments{
\item{project_git}{A parsed git project by \code{parse_gitlog}.}
}
\description{
Create a collaboration network as described by Joblin et al.
where an edge from developer A to developer B is created if A modifies a
file, and B modifies it chronologically immediately after. Note contrary
to the paper this definition is for files, not functions, and the weight
of the edges is the number of changes to a file, not churn.
}
\references{
M. Joblin, W. Mauerer, S. Apel,
J. Siegmund and D. Riehle, "From Developer Networks
to Verified Communities: A Fine-Grained Approach,"
2015 IEEE/ACM 37th IEEE International Conference on
Software Engineering, Florence, 2015, pp. 563-573,
doi: 10.1109/ICSE.2015.73.
}
\seealso{
Other edgelists: 
\code{\link{parse_commit_message_id_network}()},
\code{\link{parse_cve_cwe_file_network}()},
\code{\link{parse_dependencies_network}()},
\code{\link{parse_gitlog_network}()},
\code{\link{parse_mbox_network}()}
}
\concept{edgelists}
