% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{parse_gof_patterns}
\alias{parse_gof_patterns}
\title{Parse GoF Patterns}
\usage{
parse_gof_patterns(output_filepath = "/tmp/gof.xml")
}
\arguments{
\item{output_filepath}{Optional path to read the XML generated by \code{\link{write_gof_patterns}}.  If not
specified, it will be assumed saved to the temporary folder path `/tmp/gof.xml`.}
}
\value{
A data.table containing the parsed gof patterns per class.
}
\description{
Parses GoF patterns generated by \code{\link{write_gof_patterns}} into a a table.
\url{https://www.srcml.org/documentation.html}.
Pattern4.jar is available on
[Tsantalis' homepage](https://users.encs.concordia.ca/~nikolaos/pattern_detection.html)).
}
\references{
N. Tsantalis, A. Chatzigeorgiou, G. Stephanides, S. T. Halkidis,
"Design Pattern Detection Using Similarity Scoring",
IEEE Transactions on Software Engineering,
vol. 32, no. 11, pp. 896-909, November, 2006.
}
