% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{parse_java_code_refactoring_json}
\alias{parse_java_code_refactoring_json}
\title{Parse Java Code Refactorings}
\usage{
parse_java_code_refactoring_json(
  rminer_path,
  git_repo_path,
  start_commit,
  end_commit
)
}
\arguments{
\item{rminer_path}{The path to RMiner binary.
See \url{https://github.com/tsantalis/RefactoringMiner#running-refactoringminer-from-the-command-line}}

\item{git_repo_path}{path to git repo (ends in .git)}

\item{start_commit}{the start commit hash}

\item{end_commit}{the end commit hash}
}
\description{
Parse Java Code Refactorings
}
\references{
Nikolaos Tsantalis, Matin Mansouri, Laleh Eshkevari,
Davood Mazinanian, and Danny Dig, "Accurate and Efficient Refactoring
Detection in Commit History," 40th
International Conference on Software Engineering (ICSE 2018),
Gothenburg, Sweden, May 27 - June 3, 2018.
}
\keyword{internal}
