% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_jira_comments}
\alias{parse_jira_comments}
\title{Parse jira comments}
\usage{
parse_jira_comments(jira_comments_csv_path)
}
\arguments{
\item{jira_comments_csv_path}{path to jira comments obtained using `download_jira_data.Rmd`.}
}
\description{
Parse jira comments
}
\seealso{
Other parsers: 
\code{\link{parse_commit_message_id}()},
\code{\link{parse_dependencies}()},
\code{\link{parse_gitlog}()},
\code{\link{parse_mbox}()},
\code{\link{parse_nvdfeed}()}
}
\concept{parsers}
