% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{parse_jira_replies}
\alias{parse_jira_replies}
\title{Format Parsed Jira to Replies}
\usage{
parse_jira_replies(parsed_jira)
}
\arguments{
\item{parsed_jira}{A project's jira including issues and comments from \code{\link{parse_jira}}.}
}
\value{
A reply table.
}
\description{
Combines the JIRA issue author and description to the comments author and
description into a reply table suitable for communication analysis.
}
