% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{parse_line_metrics}
\alias{parse_line_metrics}
\title{Parse File Line Metrics}
\usage{
parse_line_metrics(scc_path, git_repo_path)
}
\arguments{
\item{scc_path}{The path to scc binary.
See \url{https://github.com/boyter/scc}}

\item{git_repo_path}{path to git repo (ends in .git)}
}
\description{
Parse File Line Metrics
}
