% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_line_type}
\alias{parse_line_type}
\title{Parse File Line Type}
\usage{
parse_line_type(utags_path, git_repo_path)
}
\arguments{
\item{utags_path}{The path to scc binary.
See \url{https://github.com/boyter/scc}}

\item{git_repo_path}{path to git repo (ends in .git)}
}
\description{
Parse File Line Type
}
