% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_line_type_file}
\alias{parse_line_type_file}
\title{Parse File Line Type}
\usage{
parse_line_type_file(utags_path, filepath, kinds)
}
\arguments{
\item{utags_path}{The path to scc binary.
See \url{https://github.com/universal-ctags/ctags}}

\item{filepath}{path to file}
}
\description{
Parse File Line Type
}
