% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\name{parse_mbox}
\alias{parse_mbox}
\title{Parse Mbox}
\usage{
parse_mbox(perceval_path, mbox_file_path)
}
\arguments{
\item{perceval_path}{path to perceval binary}

\item{mbox_file_path}{path to mbox archive file (ends in .mbox)}
}
\description{
Parses an mbox file, which consists of emails in a mailbox, using the Perceval library.
Note .mbox files do not have a consistent number of fields (e.g. Reply Cc.). Due to that,
the resulting table of parse mbox may have a different number of columns depending on the
data used. This function only ensures if columns of interest are available, then they are
consistently renamed for clarity.
}
\seealso{
Other parsers:
\code{\link[=build_understand_project]{build_understand_project()}},
\code{\link[=export_understand_dependencies]{export_understand_dependencies()}},
\code{\link[=parse_bugzilla_perceval_rest_issue_comments]{parse_bugzilla_perceval_rest_issue_comments()}},
\code{\link[=parse_bugzilla_perceval_traditional_issue_comments]{parse_bugzilla_perceval_traditional_issue_comments()}},
\code{\link[=parse_bugzilla_rest_comments]{parse_bugzilla_rest_comments()}},
\code{\link[=parse_bugzilla_rest_issues]{parse_bugzilla_rest_issues()}},
\code{\link[=parse_bugzilla_rest_issues_comments]{parse_bugzilla_rest_issues_comments()}},
\code{\link[=parse_commit_message_id]{parse_commit_message_id()}},
\code{\link[=parse_dependencies]{parse_dependencies()}},
\code{\link[=parse_dv8_clusters]{parse_dv8_clusters()}},
\code{\link[=parse_github_latest_date]{parse_github_latest_date()}},
\code{\link[=parse_gitlog]{parse_gitlog()}},
\code{\link[=parse_jira]{parse_jira()}},
\code{\link[=parse_jira_latest_date]{parse_jira_latest_date()}},
\code{\link[=parse_jira_rss_xml]{parse_jira_rss_xml()}},
\code{\link[=parse_mbox_latest_date]{parse_mbox_latest_date()}},
\code{\link[=parse_nvdfeed]{parse_nvdfeed()}},
\code{\link[=parse_understand_dependencies]{parse_understand_dependencies()}}
}
\concept{parsers}
