% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_mbox_network}
\alias{parse_mbox_network}
\title{Transform parsed mbox into a network}
\usage{
parse_mbox_network(project_mbox)
}
\arguments{
\item{project_mbox}{A parsed mbox by \code{parse_mbox}.}
}
\description{
Transform parsed mbox into a network
}
\seealso{
Other edgelists: 
\code{\link{parse_dependencies_network}()},
\code{\link{parse_gitlog_network}()}
}
\concept{edgelists}
