% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{parse_rfile_ast}
\alias{parse_rfile_ast}
\title{Parse R Abstract Syntax Tree}
\usage{
parse_rfile_ast(filepath)
}
\arguments{
\item{filepath}{The filepath of an R file}
}
\description{
Parse R Abstract Syntax Tree
}
\keyword{internal}
