% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\name{process_gz_to_mbox_in_folder}
\alias{process_gz_to_mbox_in_folder}
\title{Process .gz files in a folder, unzip and convert them to .mbox
Checks a folder for any .gz files, unzips them, and renames them
to .mbox format. The original .gz files are deleted after unzipping. If a .mbox
file with the same name already exists, it will be overwritten.}
\usage{
process_gz_to_mbox_in_folder(folder_path)
}
\arguments{
\item{folder_path}{The path to the folder containing both .gz and .mbox files.}
}
\value{
A list of the .mbox files that were created or updated.
}
\description{
Process .gz files in a folder, unzip and convert them to .mbox
Checks a folder for any .gz files, unzips them, and renames them
to .mbox format. The original .gz files are deleted after unzipping. If a .mbox
file with the same name already exists, it will be overwritten.
}
