% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{productivity_developer_churn}
\alias{productivity_developer_churn}
\title{Productivity Developer Churn}
\usage{
productivity_developer_churn(project_git_log)
}
\arguments{
\item{project_git_log}{a parsed git log obtained from \code{\link{parse_gitlog}}}
}
\value{
a four column data.table of the form author_name_email | lines_added | lines_removed | developer_churn
}
\description{
Calculates the churn per developer in the git log.
}
\seealso{
\code{\link{parse_gitlog}} to obtain additions and deletions from gitlog

Other metrics: 
\code{\link{commit_message_id_coverage}()},
\code{\link{metric_churn}()},
\code{\link{metric_churn_per_commit_interval}()},
\code{\link{metric_churn_per_commit_per_file}()},
\code{\link{metric_file_bug_churn}()},
\code{\link{metric_file_bug_frequency}()},
\code{\link{metric_file_churn}()},
\code{\link{metric_file_non_bug_churn}()},
\code{\link{metric_file_non_bug_frequency}()},
\code{\link{productivity_developer_commits}()}
}
\concept{metrics}
