% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{productivity_developer_commits}
\alias{productivity_developer_commits}
\title{Productivity Developer Commits}
\usage{
productivity_developer_commits(project_git_log)
}
\arguments{
\item{project_git_log}{a parsed git log obtained from \code{\link{parse_gitlog}}}
}
\value{
a two column data.table of the form author_name_email | total_commits
}
\description{
Counts the unique number of commits per developer in the git log.
}
\seealso{
\code{\link{parse_gitlog}}

Other metrics: 
\code{\link{commit_message_id_coverage}()},
\code{\link{metric_churn}()},
\code{\link{metric_churn_per_commit_interval}()},
\code{\link{metric_churn_per_commit_per_file}()},
\code{\link{metric_file_bug_churn}()},
\code{\link{metric_file_bug_frequency}()},
\code{\link{metric_file_churn}()},
\code{\link{metric_file_non_bug_churn}()},
\code{\link{metric_file_non_bug_frequency}()},
\code{\link{productivity_developer_churn}()}
}
\concept{metrics}
