% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment.R
\name{pysenti_predict}
\alias{pysenti_predict}
\title{Predict sentiment using a trained PySenti model}
\usage{
pysenti_predict(
  pysenti_path,
  reply_dt,
  model_save_path,
  prediction_save_path,
  model_name,
  timestamp
)
}
\arguments{
\item{pysenti_path}{Path to the PySenti Python script.}

\item{reply_dt}{A data.table containing a `text` column.}

\item{model_save_path}{Path to the trained model folder.}

\item{timestamp}{Character string representing the current time, formatted as "YYYYMMDD_HHMMSS", used to append to the prediction file name for uniqueness.}

\item{prediction_path}{Directory where the prediction CSV will be written.}

\item{model}{Model architecture name; `"bert-base-cased"`, `"xlnet-base-cased", etc.}
}
\value{
A data.table containing predicted sentiment labels.
}
\description{
Applies a trained PySenti model to a reply data.table, writes predictions
to a CSV file, and returns them as a data.table.
}
