% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment.R
\name{pysenti_train_model}
\alias{pysenti_train_model}
\title{Train a sentiment model using PySenti}
\usage{
pysenti_train_model(pysenti_path, train_data_path, model_save_path, model_name)
}
\arguments{
\item{pysenti_path}{Path to the PySenti Python script.}

\item{train_data_path}{Path to the model train .csv with at least `text` and `polarity` columns.}

\item{model_save_path}{Directory where the trained model will be stored.}

\item{model_name}{Model architecture name; `"bert-base-cased"`, `"xlnet-base-cased", etc.}
}
\value{
Character vector containing stdout/stderr from the Python process.
}
\description{
Trains a transformer-based sentiment model (e.g., BERT, XLNet, RoBERTa,
ALBERT) by passing a labeled reply data.table to the external PySenti
Python script.
}
