% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{query_src_text}
\alias{query_src_text}
\title{Query srcML XML}
\usage{
query_src_text(srcml_path, xpath_query, srcml_filepath)
}
\arguments{
\item{srcml_path}{The path to srcML binary}

\item{xpath_query}{The XPath query to be performed on the .xml}

\item{srcml_filepath}{The path to the srcML file to be queried
(see \code{\link{annotate_src_text}}).}
}
\value{
The path where the output xml was saved (i.e. srcml_filepath)
}
\description{
Queries srcML XML for code units (e.g. function names, declarations, etc.).
For a list of code units and languages supported by srcML XML see:
\url{https://www.srcml.org/documentation.html}.
}
\details{
Note a query, unless explicitly specified to be parsed (e.g. by using the
string() scrML function), is a srcML XML itself, which can be also queried.
}
\references{
For details, see \url{https://www.srcml.org/tutorials/xpath-query.html}.
}
