% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{query_src_text_class_docs}
\alias{query_src_text_class_docs}
\title{Query srcML Class Documentation Comments}
\usage{
query_src_text_class_docs(srcml_path, srcml_filepath, exclude_license = TRUE)
}
\arguments{
\item{srcml_path}{The path to the srcML binary.}

\item{srcml_filepath}{The path to the srcML file to be queried.}

\item{exclude_license}{If TRUE, excludes license-related comments. Defaults to TRUE.}
}
\value{
A data.table containing the file path and the class-level documentation comment.
}
\description{
This is a convenience function to parse class-level Javadoc comments immediately preceding class definitions, excluding the license block.
}
