% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{query_src_text_file_docs}
\alias{query_src_text_file_docs}
\title{Query srcML File-Level Documentation Comments}
\usage{
query_src_text_file_docs(srcml_path, srcml_filepath)
}
\arguments{
\item{srcml_path}{The path to the srcML binary.}

\item{srcml_filepath}{The path to the srcML file to be queried.}
}
\value{
A data.table containing the file path and the file-level documentation comment.
}
\description{
This is a convenience function to parse file-level documentation comments, excluding the license block.
}
