% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{query_src_text_functions}
\alias{query_src_text_functions}
\title{Query srcML Function Declarations}
\usage{
query_src_text_functions(srcml_path, srcml_filepath, include_parameters = TRUE)
}
\arguments{
\item{srcml_path}{The path to the srcML binary.}

\item{srcml_filepath}{The path to the srcML file to be queried (see \code{\link{annotate_src_text}}).}

\item{include_parameters}{If TRUE, includes function parameters. Defaults to TRUE.}
}
\value{
A data.table containing filepath, function_name, and optionally parameters.
}
\description{
This function extracts function names and optionally their parameters from the srcML XML file.
}
