% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{query_src_text_imports}
\alias{query_src_text_imports}
\title{Query srcML Imports}
\usage{
query_src_text_imports(srcml_path, srcml_filepath)
}
\arguments{
\item{srcml_path}{The path to the srcML binary.}

\item{srcml_filepath}{The path to the srcML file to be queried (see \code{\link{annotate_src_text}}).}
}
\value{
A data.table containing filepath and imports.
}
\description{
This function extracts import statements and their associated file paths from the srcML XML file.
}
