% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{query_src_text_namespace}
\alias{query_src_text_namespace}
\title{Query srcML Namespace}
\usage{
query_src_text_namespace(srcml_path, srcml_filepath)
}
\arguments{
\item{srcml_path}{The path to srcML binary}

\item{srcml_filepath}{The path to the srcML file to be queried
(see \code{\link{annotate_src_text}}).}
}
\value{
A data.table containing Namespace.
}
\description{
This is a convenience function to parse namespace names out of a project.
\url{https://www.srcml.org/documentation.html}.
}
\references{
For details, see \url{https://www.srcml.org/documentation.html}.
}
