% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{query_src_text_packages}
\alias{query_src_text_packages}
\title{Query srcML Package Names}
\usage{
query_src_text_packages(srcml_path, srcml_filepath)
}
\arguments{
\item{srcml_path}{The path to the srcML binary.}

\item{srcml_filepath}{The path to the srcML file to be queried.}
}
\value{
A data.table containing filepath and package names.
}
\description{
This function extracts package names from the srcML XML file.
}
