% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{query_src_text_variables}
\alias{query_src_text_variables}
\title{Query srcML Variable Declarations}
\usage{
query_src_text_variables(srcml_path, srcml_filepath, var_type = TRUE)
}
\arguments{
\item{srcml_path}{The path to the srcML binary.}

\item{srcml_filepath}{The path to the srcML file to be queried (see \code{\link{annotate_src_text}}).}

\item{var_type}{If TRUE, includes variable types. Defaults to TRUE.}
}
\value{
A data.table containing filepath, variable_name, and optionally variable_type.
}
\description{
This function extracts variable names and their types from the srcML XML file.
}
