% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_temporary_file}
\alias{read_temporary_file}
\title{Read Temporary File}
\usage{
read_temporary_file(temp)
}
\arguments{
\item{temp}{A temporary file created by \code{make_temporary_file}}
}
\description{
Read a temporary file.
}
