% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugzilla.R
\name{refresh_bugzilla_issues_comments}
\alias{refresh_bugzilla_issues_comments}
\title{Refresh Bugzilla Issues or Issues Comments}
\usage{
refresh_bugzilla_issues_comments(
  bugzilla_site,
  start_timestamp = "1970-01-01T:00:00:00Z",
  save_folder_path,
  project_key,
  limit_upperbound = 20,
  comments = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{bugzilla_site}{URL to specific bugzilla site}

\item{start_timestamp}{When to start bug retrieval (ex. 2023-01-01T00:14:57Z). When not specified, the start timestamp
will begin from the first issue of the issue tracker closest to 1970.}

\item{save_folder_path}{the full folder path where the bugzilla issues will be stored}

\item{project_key}{the project key of the project which can be found in the respective config file}

\item{limit_upperbound}{the number of issues saved in each page file. Some bugzilla sites have limits set on how many bugs
can be retrieved in one GET request, in which case, the limit set by the bugzilla site will be used in place of
limit_upperbound to ensure full bug retrieval.}

\item{comments}{set true to download issues with comments, leave as false to download only issues (without comments)}

\item{verbose}{set true to print execution details}
}
\description{
Downloads Bugzilla issues or issues with comments starting from the most recent file in the folder.
If the directory is empty, then all issues will be downloaded for the Bugzilla project.
}
\seealso{
\code{\link{parse_bugzilla_rest_issues_comments}} a parser function to parse Bugzilla issues and comments data
}
