% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{refresh_jira_issues_comments_by_issuekey}
\alias{refresh_jira_issues_comments_by_issuekey}
\title{Define function to extract greatest issuekey from a file and pass it to download_jira_issues_comments_by_issuekey}
\usage{
refresh_jira_issues_comments_by_issuekey(
  domain,
  credentials,
  jql_query,
  fields,
  save_path_issue_tracker_issues,
  maxresults,
  verbose,
  maxdownloads,
  file_name,
  unaltered_file_path
)
}
\arguments{
\item{domain}{Custom JIRA domain URL set in config file}

\item{credentials}{a path to text file containing your username/api token}

\item{jql_query}{Specific query string to specify criteria for fetching}

\item{fields}{List of fields that are downloaded in each issue}

\item{save_path_issue_tracker_issues}{Path that files will be save along}

\item{maxresults}{(optional) the maximum number of results to download per page.
Default is 50}

\item{verbose}{boolean flag to specify printing operational
messages or not}

\item{maxdownloads}{Maximum downloads per function call}

\item{file_name}{The filename that contains the greatest issuekey. Returned by a parser}

\item{unaltered_file_path}{the unaltered filepath set in the config file}
}
\description{
Download issue data from "rest/api/2/search" endpoint
}
\seealso{
\code{link{download_jira_issues_comments_by_issuekey}} to download JIRA data using issuekey as a search criteria,

\code{link{download_jira_issues_comments}} to download all JIRA issues data,

\code{link{download_jira_issues_comments_by_date}} to download JIRA data using created as a search criteria,

\code{link{parse_jira}} to parse jira issues along a file-path,

\code{link{parse_jira_latest_date}} to retrieve the file_name to be passed to this function

Other downloaders: 
\code{\link{download_jira_issues_comments_by_date}()},
\code{\link{download_jira_issues_comments_by_issuekey}()},
\code{\link{download_jira_issues_comments}()}

Other jira: 
\code{\link{download_jira_issues_comments_by_date}()},
\code{\link{download_jira_issues_comments_by_issuekey}()},
\code{\link{download_jira_issues_comments}()}
}
\concept{downloaders}
\concept{jira}
