% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\name{refresh_mod_mbox}
\alias{refresh_mod_mbox}
\title{Refresh Mod_Mbox}
\usage{
refresh_mod_mbox(
  mailing_list,
  start_year_month,
  save_folder_path,
  verbose = TRUE
)
}
\arguments{
\item{mailing_list}{The URL of the mailing list being downloaded (e.g., \url{https://lists.apache.org/list.html?announce@apache.org})}

\item{start_year_month}{The year and month of the first file to be downloaded (format: 'YYYYMM').}

\item{save_folder_path}{The folder path in which all the downloaded mod_mbox files will be stored.}

\item{verbose}{if TRUE, prints diagnostic messages.}
}
\value{
Returns `downloaded_files`, a vector of the newly downloaded files in the current working directory.
}
\description{
This function refreshes the mailing list files by checking the contents of a specified folder.
If the folder is empty, it calls \code{\link{download_mod_mbox}} to download all mod_mbox files from start_year_month to the current month.
If the folder contains already-downloaded mbox files, it identifies the most recent month, deletes that file, and redownloads it
along with all future months up to the current real-life month.
}
\details{
The naming convention of files is `YYYYMM.mbox`, and the function uses this pattern to identify the most recent month.
After deleting the most recent file, the function ensures that the month is redownloaded, along with all subsequent months up to the current month.
Redownloading the most recent file ensures any files added in that month after the latest refresh are included.
}
