% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\name{refresh_pipermail}
\alias{refresh_pipermail}
\title{Refresh mbox files downloaded via pipermail
Uses the adopted file name convention by \code{\link{download_pipermail}} to identify
the latest downloaded mbox. It deletes this file, then redownloads it along with all future months
up to the current real-life month.
If the directory is empty, then it downloads all pipermail files (as mbox files) via \code{\link{download_pipermail}}}
\usage{
refresh_pipermail(mailing_list, start_year_month, save_folder_path)
}
\arguments{
\item{mailing_list}{The name of the mailing list being downloaded}

\item{start_year_month}{The year and month of the first file to be downloaded}

\item{save_folder_path}{The folder path in which all the downloaded pipermail files will be stored}
}
\value{
Returns `downloaded_files`, a vector of the downloaded files in the current working directory
}
\description{
Refresh mbox files downloaded via pipermail
Uses the adopted file name convention by \code{\link{download_pipermail}} to identify
the latest downloaded mbox. It deletes this file, then redownloads it along with all future months
up to the current real-life month.
If the directory is empty, then it downloads all pipermail files (as mbox files) via \code{\link{download_pipermail}}
}
