% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{replace_token_regex_with}
\alias{replace_token_regex_with}
\title{Replace Tokens in specified columns.}
\usage{
replace_token_regex_with(dt_file, regex_to_replace_key_with, file_column_name)
}
\arguments{
\item{dt_file}{A data.table containing the column to process.}

\item{regex_to_replace_key_with}{A named list of token names and their corresponding regex patterns.}

\item{file_column_name}{The name of the column to perform replacements on.}
}
\value{
The data.table with regex matches replaced by token names.
}
\description{
Replaces patterns in a specified column of a data.table with token placeholders.
}
\seealso{
Other filters: 
\code{\link{filter_by_commit_interval}()},
\code{\link{filter_by_commit_size}()},
\code{\link{filter_by_file_extension}()},
\code{\link{filter_by_filepath_substring}()},
\code{\link{filter_by_reply_author_substring}()},
\code{\link{filter_by_reply_body_substring}()},
\code{\link{filter_by_reply_subject_substring}()}
}
\concept{filters}
