% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smells.R
\name{smell_detect_ragequit}
\alias{smell_detect_ragequit}
\title{Detect Ragequit Social Smell.}
\usage{
smell_detect_ragequit(
  timestamp,
  author,
  message,
  sentiment_value,
  quit_lag = 90,
  current_time = Sys.time()
)
}
\arguments{
\item{timestamp}{A data table column indicating timestamp of an author's commit message}

\item{author}{A data table column indicating the author of a commit message}

\item{message}{A data table column containing the commit message text}

\item{sentiment_value}{A data table column containing the sentiment value of the commit message 
(-1 for negative, 0 for neutral, 1 for positive)}

\item{quit_lag}{The number of days since a developer's last commit message}

\item{current_time}{The current time to compare against the last commit timestamp (default is the system's current time)}
}
\description{
Detect potential ragequits by identifying cases where, 
after a period of inactivity, a developer's final commit message expressed negative sentiment.
}
\references{
Wouter Mulder (2025). Am I finished yet? A discovery of burnout and
ragequits within open-source projects. (Master thesis, Jheronimus Academy of Data Science).
}
