% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smells.R
\name{smell_missing_links}
\alias{smell_missing_links}
\title{Missing link Social Smell}
\usage{
smell_missing_links(mail.graph, code.graph, precomputed.silo = NA)
}
\arguments{
\item{mail.graph}{A uni-modal network G = (V,E), where V are
developers (authors or contributors) and E is a collaboration between developers
in the mailing list}

\item{code.graph}{A uni-modal network G = (V,E), where V are
developers (authors or contributors) and E is a collaboration between developers
in the git log.}

\item{precomputed.silo}{Optional. The return of \code{smell_organizational_silo}}
}
\description{
Given the communication and collaboration graphs (mailing lists and VCS)
return the collaboration edges that do not have a communication counterpart.
It is possible to pass precomputed organisational silo community smell. For conceptual
details please see references.
}
\references{
D. A. Tamburri, F. Palomba and R. Kazman,
"Exploring Community Smells in Open-Source: An Automated Approach"
in IEEE Transactions on Software Engineering, vol. 47, no. 3,
pp. 630-652, 1 March 2021, doi: 10.1109/TSE.2019.2901490.

Simone Magnoni (2016). An approach to measure Community
Smells in software development communities. (Doctoral dissertation, Politecnico Milano).
}
