% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smells.R
\name{smell_organizational_silo}
\alias{smell_organizational_silo}
\title{Organizational Silo Social Smell}
\usage{
smell_organizational_silo(mail.graph, code.graph)
}
\arguments{
\item{mail.graph}{A uni-modal network G = (V,E), where V are
developers (authors or contributors) and E is a collaboration between developers
in the mailing list}

\item{code.graph}{A uni-modal network G = (V,E), where V are
developers (authors or contributors) and E is a collaboration between developers
in the git log.}
}
\description{
Given the communication and collaboration graphs (mailing lists and VCS)
return the collaboration edges, analyzed in groups of two devs, in
which one of co-committing devs do not communicate at all. For conceptual
details please see references.
}
\references{
D. A. Tamburri, F. Palomba and R. Kazman,
"Exploring Community Smells in Open-Source: An Automated Approach"
in IEEE Transactions on Software Engineering, vol. 47, no. 3,
pp. 630-652, 1 March 2021, doi: 10.1109/TSE.2019.2901490.

Simone Magnoni (2016). An approach to measure Community
Smells in software development communities. (Doctoral dissertation, Politecnico Milano).
}
