% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smells.R
\name{smell_radio_silence}
\alias{smell_radio_silence}
\title{Radio Silence Social Smell.}
\usage{
smell_radio_silence(mail.graph, clusters)
}
\arguments{
\item{mail.graph}{A uni-modal network G = (V,E), where V are
developers (authors or contributors) and E is a collaboration between developers
in the mailing list}

\item{clusters}{Clusters obtained via community detection of mail.graph network.}
}
\description{
Given the communication graph and its sub-communities,
return the ids of unique boundary spanners towards another
sub-community, generation this smell. For conceptual
details please see references.
}
\references{
D. A. Tamburri, F. Palomba and R. Kazman,
"Exploring Community Smells in Open-Source: An Automated Approach"
in IEEE Transactions on Software Engineering, vol. 47, no. 3,
pp. 630-652, 1 March 2021, doi: 10.1109/TSE.2019.2901490.

Simone Magnoni (2016). An approach to measure Community
Smells in software development communities. (Doctoral dissertation, Politecnico Milano).
}
