% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identity.R
\name{split_name_email}
\alias{split_name_email}
\title{Split Name and Email}
\usage{
split_name_email(name_email)
}
\arguments{
\item{name_email}{A formatted name and email.}
}
\description{
Split Name and Email
}
