% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{split_path}
\alias{split_path}
\title{splith_path splits the file path into root adress and file name}
\usage{
split_path(file_path)
}
\arguments{
\item{file_path}{The path of a file}
}
\value{
List: first element:root folder, second element:file name
}
\description{
splith_path splits the file path into root adress and file name
}
