% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{transform_dependencies_to_network}
\alias{transform_dependencies_to_network}
\title{Transform parsed dependencies into a network}
\usage{
transform_dependencies_to_network(depends_parsed, weight_types = NA)
}
\arguments{
\item{depends_parsed}{A parsed mbox by \code{\link{parse_dependencies}}.}

\item{weight_types}{The weight types as defined in Depends.}
}
\description{
Transform parsed dependencies into a network
}
\seealso{
Other edgelists:
\code{\link[=transform_commit_message_id_to_network]{transform_commit_message_id_to_network()}},
\code{\link[=transform_cve_cwe_file_to_network]{transform_cve_cwe_file_to_network()}},
\code{\link[=transform_dependencies_to_sdsmj]{transform_dependencies_to_sdsmj()}},
\code{\link[=transform_gitlog_to_bipartite_network]{transform_gitlog_to_bipartite_network()}},
\code{\link[=transform_gitlog_to_entity_bipartite_network]{transform_gitlog_to_entity_bipartite_network()}},
\code{\link[=transform_gitlog_to_entity_temporal_network]{transform_gitlog_to_entity_temporal_network()}},
\code{\link[=transform_gitlog_to_hdsmj]{transform_gitlog_to_hdsmj()}},
\code{\link[=transform_gitlog_to_temporal_network]{transform_gitlog_to_temporal_network()}},
\code{\link[=transform_reply_to_bipartite_network]{transform_reply_to_bipartite_network()}},
\code{\link[=transform_temporal_gitlog_to_adsmj]{transform_temporal_gitlog_to_adsmj()}},
\code{\link[=transform_understand_dependencies_to_network]{transform_understand_dependencies_to_network()}}
}
\concept{edgelists}
