% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{transform_mbox_to_bipartite_network}
\alias{transform_mbox_to_bipartite_network}
\title{Transform parsed mbox into a network}
\usage{
transform_mbox_to_bipartite_network(project_mbox)
}
\arguments{
\item{project_mbox}{A parsed mbox by \code{parse_mbox}.}
}
\description{
Transform parsed mbox into a network
}
\seealso{
Other edgelists: 
\code{\link{transform_commit_message_id_to_network}()},
\code{\link{transform_cve_cwe_file_to_network}()},
\code{\link{transform_dependencies_to_network}()},
\code{\link{transform_gitlog_to_bipartite_network}()},
\code{\link{transform_gitlog_to_entity_bipartite_network}()},
\code{\link{transform_gitlog_to_entity_temporal_network}()},
\code{\link{transform_gitlog_to_temporal_network}()}
}
\concept{edgelists}
