% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{bipartite_graph_projection}
\alias{bipartite_graph_projection}
\title{Apply a bipartite graph projection}
\usage{
bipartite_graph_projection(graph, mode, is_intermediate_projection = FALSE)
}
\arguments{
\item{graph}{A bipartite network (see transform_*_bipartite functions).}

\item{mode}{Which of the two nodes the projection should be applied to. TRUE or FALSE}

\item{is_intermediate_projection}{If true, displays the eliminated nodes by the projection. TRUE or FALSE.}
}
\value{
A graph projection.
}
\description{
Apply a bipartite graph projection
}
