% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{build_understand_project}
\alias{build_understand_project}
\title{Build Understand DB}
\usage{
build_understand_project(scitools_path, project_path, language, output_dir)
}
\arguments{
\item{scitools_path}{path to the scitools binary `und`}

\item{project_path}{path to the project source code folder to create the Understand DB.}

\item{language}{the primary language of the project (language must be supported by Understand)}

\item{output_dir}{path to output directory (formatted output_path/)}
}
\value{
The created Scitools Understand DB path
}
\description{
Uses Scitools Understand to create a source code project Und Database.
}
\references{
See pg. 352 in https://documentation.scitools.com/pdf/understand.pdf Sept. 2024 Edition
}
\seealso{
Other parsers:
\code{\link[=export_understand_dependencies]{export_understand_dependencies()}},
\code{\link[=parse_bugzilla_perceval_rest_issue_comments]{parse_bugzilla_perceval_rest_issue_comments()}},
\code{\link[=parse_bugzilla_perceval_traditional_issue_comments]{parse_bugzilla_perceval_traditional_issue_comments()}},
\code{\link[=parse_bugzilla_rest_comments]{parse_bugzilla_rest_comments()}},
\code{\link[=parse_bugzilla_rest_issues]{parse_bugzilla_rest_issues()}},
\code{\link[=parse_bugzilla_rest_issues_comments]{parse_bugzilla_rest_issues_comments()}},
\code{\link[=parse_commit_message_id]{parse_commit_message_id()}},
\code{\link[=parse_dependencies]{parse_dependencies()}},
\code{\link[=parse_dv8_clusters]{parse_dv8_clusters()}},
\code{\link[=parse_github_latest_date]{parse_github_latest_date()}},
\code{\link[=parse_gitlog]{parse_gitlog()}},
\code{\link[=parse_jira]{parse_jira()}},
\code{\link[=parse_jira_latest_date]{parse_jira_latest_date()}},
\code{\link[=parse_jira_rss_xml]{parse_jira_rss_xml()}},
\code{\link[=parse_mbox]{parse_mbox()}},
\code{\link[=parse_mbox_latest_date]{parse_mbox_latest_date()}},
\code{\link[=parse_nvdfeed]{parse_nvdfeed()}},
\code{\link[=parse_understand_dependencies]{parse_understand_dependencies()}}
}
\concept{parsers}
