% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{build_understand_project}
\alias{build_understand_project}
\title{Build Understand Analysis Folder}
\usage{
build_understand_project(project_path, language, output_dir)
}
\arguments{
\item{project_path}{path to the project folder to analyze}

\item{language}{the primary language of the project (language must be supported by Understand)}

\item{output_dir}{path to output directory (formatted output_path/)}
}
\description{
This function builds the data files for Understand from the project_path folder, reading from files that are written in the target language into output_dir
}
\seealso{
Other parsers: 
\code{\link{parse_bugzilla_perceval_rest_issue_comments}()},
\code{\link{parse_bugzilla_perceval_traditional_issue_comments}()},
\code{\link{parse_bugzilla_rest_comments}()},
\code{\link{parse_bugzilla_rest_issues}()},
\code{\link{parse_bugzilla_rest_issues_comments}()},
\code{\link{parse_commit_message_id}()},
\code{\link{parse_dependencies}()},
\code{\link{parse_dv8_clusters}()},
\code{\link{parse_gitlog}()},
\code{\link{parse_jira}()},
\code{\link{parse_jira_latest_date}()},
\code{\link{parse_jira_rss_xml}()},
\code{\link{parse_mbox}()},
\code{\link{parse_nvdfeed}()},
\code{\link{parse_understand_dependencies}()}
}
\concept{parsers}
