% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{community_oslom}
\alias{community_oslom}
\alias{community_oslom.unimodal}
\alias{community_oslom.multimodal}
\title{OSLOM Community Detection (S3 generic)}
\usage{
community_oslom(oslom_bin_dir_undir_path, graph, seed, n_runs, ...)

\method{community_oslom}{unimodal}(oslom_bin_dir_undir_path, graph, seed, n_runs, ...)

\method{community_oslom}{multimodal}(oslom_bin_dir_undir_path, graph, seed, n_runs, ...)
}
\arguments{
\item{oslom_bin_dir_undir_path}{The path to oslom directed or undirected network binary}

\item{graph}{A graph returned by \code{\link{model_unimodal_graph}}.}

\item{seed}{An integer specifying the seed to replicate the result}

\item{n_runs}{the number of runs for the first hierarchical level}

\item{...}{Additional arguments passed to the method.}
}
\description{
Wrapper for OSLOM Community Detection \url{http://oslom.org/}
}
\references{
Finding statistically significant communities in networks
A. Lancichinetti, F. Radicchi, J.J. Ramasco and
S. Fortunato PLoS ONE 6, e18961 (2011).
}
\seealso{
Other community: 
\code{\link{recolor_network_by_community}()}
}
\concept{community}
