% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{create_base_info}
\alias{create_base_info}
\title{Create base_info_cell}
\usage{
create_base_info(jira_domain_url, issue_key)
}
\arguments{
\item{jira_domain_url}{URL of JIRA}

\item{issue_key}{key for Jira issue}
}
\value{
A list named 'base_info_cell' containing all the parameters
}
\description{
Creates and formats base_info_cell for fake Jira issue
}
