% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{create_base_info}
\alias{create_base_info}
\title{Create base_info_cell}
\usage{
create_base_info(jira_domain_url, issue_key)
}
\arguments{
\item{jira_domain_url}{URL of JIRA domain}

\item{issue_key}{key for JIRA issue}
}
\value{
A list named 'base_info_cell' containing all information for base cell in json file
}
\description{
Creates and formats base_info_cell list for \code{\link{make_jira_issue}}.
}
\keyword{internal}
