% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{create_creator}
\alias{create_creator}
\title{Create Creator}
\usage{
create_creator(jira_domain_url, creator_name)
}
\arguments{
\item{jira_domain_url}{URL of JIRA domain}

\item{creator_name}{name of creator}
}
\value{
A list named 'creator' that has creator's information
}
\description{
Creates the issue creator cell for \code{\link{make_jira_issue}}.
}
\keyword{internal}
