% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{create_ext_info}
\alias{create_ext_info}
\title{Create ext_info_cell for issues (without comments)}
\usage{
create_ext_info(
  jira_domain_url,
  issue_type_description,
  summary,
  components,
  description,
  reporter_name,
  assignee_name,
  status_description
)
}
\value{
ext info cell
}
\description{
Creates and formats ext_info_cell for fake Jira issue
}
