% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\name{create_fake_mbox}
\alias{create_fake_mbox}
\title{Creates fake mbox data}
\usage{
create_fake_mbox(
  mlist,
  reply_from_author,
  reply_from_email,
  reply_to_author,
  reply_to_email,
  reply_cc_author,
  reply_cc_email,
  reply_datetime,
  timezone,
  reply_subject,
  reply_body
)
}
\arguments{
\item{mlist}{specific mailing list associated with this mbox data}

\item{reply_from_author}{name of sender. Formatted as: FIRSTNAME LASTNAME}

\item{reply_from_email}{email of sender. eg: 'sender@gmail.com'}

\item{reply_to_author}{name of receiver. Formatted as: FIRSTNAME LASTNAME}

\item{reply_to_email}{email of receiver eg: 'receiver@gmail.com'}

\item{reply_cc_author}{name of carbon copy of email to additional recipients. Formatted as: FIRSTNAME LASTNAME}

\item{reply_cc_email}{email of carbon copy recipients}

\item{reply_datetime}{Date, time in the following format: 2023-02-11T09:30:00  Be sure to add the capital T to separate date and time}

\item{timezone}{Abbreviated timezone of the reply eg: HST}

\item{reply_subject}{subject of the email as a string}

\item{reply_body}{the body of the email as a string}
}
\value{
the content of the fake mbox you created
}
\description{
Used to create a fake mbox with the specified parameters and returns the output to the console
}
