% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{create_issue_type}
\alias{create_issue_type}
\title{Create Issue Type}
\usage{
create_issue_type(jira_domain_url, issue_type)
}
\arguments{
\item{jira_domain_url}{URL of JIRA domain}

\item{issue_type}{name of the issue type (e.g. New Feature)}
}
\value{
A list named 'issue_type' that represents the issue type of the JIRA issue
}
\description{
Create issue type cell for \code{\link{make_jira_issue}}.
}
\keyword{internal}
