% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{create_parent}
\alias{create_parent}
\title{Create Parent}
\usage{
create_parent(jira_domain_url, issue_key, status, priority, issue_type)
}
\arguments{
\item{jira_domain_url}{URL of JIRA domain}

\item{issue_key}{issue key of JIRA issue (e.g. "PROJECT-68" or "GERONIMO-6723)}

\item{status}{status of issue for development (e.g. "In Progress")}

\item{priority}{the name of the priority of the issue (Major, Minor, Trivial)}

\item{issue_type}{type of JIRA issue (e.g. "New Feature", "Task", "Bug")}
}
\value{
A list named 'parent' that contains information on a parent issue
}
\description{
Create a parent cell for \code{\link{make_jira_issue}}.
}
\keyword{internal}
