% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{create_project}
\alias{create_project}
\title{Create Project}
\usage{
create_project(jira_domain_url, project_key)
}
\arguments{
\item{jira_domain_url}{URL of JIRA domain}

\item{project_key}{key of the project that contains the JIRA issue (e.g. "SPARK" or "GERONIMO")}
}
\value{
A list named 'project' that contains the project's information
}
\description{
Create a project cell for \code{\link{make_jira_issue}}.
}
\keyword{internal}
