% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{create_reporter}
\alias{create_reporter}
\title{Create Reporter}
\usage{
create_reporter(jira_domain_url, reporter_name)
}
\arguments{
\item{jira_domain_url}{URL of JIRA domain}

\item{reporter_name}{name of reporter}
}
\value{
A list named 'reporter' which contains the reporter's information
}
\description{
Creates a reporter cell for \code{\link{make_jira_issue}}
}
\keyword{internal}
