% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{create_status}
\alias{create_status}
\title{Create Status}
\usage{
create_status(jira_domain_url, status)
}
\arguments{
\item{jira_domain_url}{URL of JIRA domain}

\item{status}{description of status}
}
\value{
A list named 'status' containing status's information
}
\description{
Creates a status cell for \code{\link{make_jira_issue}}.
}
\keyword{internal}
