% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv8.R
\name{dependencies_to_sdsmj}
\alias{dependencies_to_sdsmj}
\title{Convert dependencies table to a sdsm.json}
\usage{
dependencies_to_sdsmj(depends_table, project_name, is_sorted = FALSE)
}
\arguments{
\item{depends_table}{parsed dependencies table}

\item{project_name}{project name for the sdsm.json}

\item{is_sorted}{whether the json is sorted by src and dest}
}
\description{
Converts table of dependencies from \code{\link{parse_dependencies}} into an *-sdsm.json binary.
In the sdsm.json, the Variables are all files/methods or any variables under analysis
(rows/columns in dependency matrix) and the Cells (matrix cell) contain all the relations of
 variable (src & dest) pairs.
}
\seealso{
\code{\link{parse_dependencies}}

Other dv8: 
\code{\link{dv8_clsxb_to_clsxj}()},
\code{\link{gitlog_to_hdsmj}()},
\code{\link{parse_dv8_clusters}()}
}
\concept{dv8}
