% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_bugzilla}
\alias{download_bugzilla}
\title{Download Bugzilla issues and comments using Perceval}
\usage{
download_bugzilla(perceval_path, bugzilla_site, date, backend = "bugzilla")
}
\arguments{
\item{perceval_path}{path to perceval binary}

\item{bugzilla_site}{link to specific bugzilla site}

\item{date}{the date to start bug retrieval}

\item{backend}{either traditional "bugzilla" or "bugzillarest" for Bugzilla >= 5.0 servers}
}
\description{
Download Bugzilla issues and comments using Perceval
}
