% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\name{download_mod_mbox}
\alias{download_mod_mbox}
\title{Download all mod_mbox files in a mailing list as mbox files}
\usage{
download_mod_mbox(
  mailing_list,
  start_year_month,
  end_year_month,
  save_folder_path,
  verbose = TRUE
)
}
\arguments{
\item{mailing_list}{The URL of the Apache Pony Mail list from which mbox files are to be downloaded
(e.g., "https://lists.apache.org/list.html?announce@apache.org").}

\item{start_year_month}{The year and month of the first file to be downloaded (format: 'YYYYMM').}

\item{end_year_month}{The year and month of the last file to be downloaded (format: 'YYYYMM').}

\item{save_folder_path}{The folder path where all the downloaded mbox files will be stored.}

\item{verbose}{if TRUE, prints detailed messages during the download process.}
}
\value{
Returns `save_folder_path`, the folder path where the mbox files are stored.
}
\description{
This function downloads mod_mbox archives from a specified Apache Pony Mail mailing list as .mbox files.
It constructs the download URLs for each month based on the start and end date range and downloads the mbox files
in the format "YYYY-MM". The downloaded .mbox files are saved in the specified folder, with a naming convention
of kaiaulu_YYYYMM.mbox.

The function loops through each month in the range specified by `start_year_month` and `end_year_month`,
and constructs the appropriate URL to download each month's data. If any download fails, an error message is printed.
}
