% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\name{download_pipermail}
\alias{download_pipermail}
\title{Download all pipermail files in an archive}
\usage{
download_pipermail(url)
}
\arguments{
\item{url}{An url pointing to a pipermail archive}
}
\value{
Returns `destination`, a vector of the downloaded files in the current working directory
}
\description{
Download all pipermail files in an archive
}
