% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv8.R
\name{dv8_clsxj_to_clsxb}
\alias{dv8_clsxj_to_clsxb}
\title{Convert an input file project-hier.json to a clustering file (*.dv8-clsx)}
\usage{
dv8_clsxj_to_clsxb(input_path, output_path)
}
\arguments{
\item{input_path}{path to the input file project-hier.json}

\item{output_path}{path to clustering file (*.dv8-clsx)}
}
\description{
Convert an input file project-hier.json to a clustering file (*.dv8-clsx)
}
