% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv8.R
\name{dv8_clsxj_to_clsxb}
\alias{dv8_clsxj_to_clsxb}
\title{Convert a json cluster file, clsxj, to a binary cluster file, clsxb.
An input file *-hier.json is converted to *.dv8-clsx.}
\usage{
dv8_clsxj_to_clsxb(clsxj_path, clsxb_path)
}
\arguments{
\item{clsxj_path}{path to the input file project-hier.json}

\item{clsxb_path}{path to clustering file (*.dv8-clsx)}
}
\description{
Convert a json cluster file, clsxj, to a binary cluster file, clsxb.
An input file *-hier.json is converted to *.dv8-clsx.
}
