% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv8.R
\name{dv8_depends_to_sdsmj}
\alias{dv8_depends_to_sdsmj}
\title{Creates dependencies using Depends}
\usage{
dv8_depends_to_sdsmj(depends_jar_path, git_repo_path, language, sdsmj_path)
}
\arguments{
\item{depends_jar_path}{path to depends jar}

\item{git_repo_path}{path to git repo (ends in .git)}

\item{language}{the language of the .git repo (accepts cpp, java, ruby, python, pom)}

\item{dsmb_path}{the path to save the sdsm-dv8.dsm file.}
}
\description{
Creates dependencies using Depends
}
\seealso{
Other parsers: 
\code{\link{parse_bugzilla_crawler_comments}()},
\code{\link{parse_bugzilla_crawler_issue}()},
\code{\link{parse_commit_message_id}()},
\code{\link{parse_dependencies}()},
\code{\link{parse_dv8_clusters}()},
\code{\link{parse_gitlog}()},
\code{\link{parse_jira}()},
\code{\link{parse_mbox}()},
\code{\link{parse_nvdfeed}()}
}
\concept{parsers}
