% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv8.R
\name{dv8_dsmj_to_dsmb}
\alias{dv8_dsmj_to_dsmb}
\title{Convert from sdsm.json to a dependency matrix in sdsm.dv8-dsm file}
\usage{
dv8_dsmj_to_dsmb(dsmj_path, dsmb_path)
}
\arguments{
\item{dsmj_path}{path to sdsm.json}

\item{dsmb_path}{path to save dv8-dsm file}
}
\description{
Convert from sdsm.json to a dependency matrix in sdsm.dv8-dsm file
}
