% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv8.R
\name{dv8_gitlog_to_gitnumstat}
\alias{dv8_gitlog_to_gitnumstat}
\title{Transforms a git log to a git numstat file.}
\usage{
dv8_gitlog_to_gitnumstat(git_repo_path, git_numstat_path)
}
\arguments{
\item{git_repo_path}{path to local project git repository}

\item{git_numstat_path}{path to save gitlog numstat}
}
\value{
the git_numstat_path
}
\description{
Comverts a git folder \*.git, to a gitlog numstat file \*.txt
}
\seealso{
\code{\link{dv8_gitnumstat_to_hdsmb}} to covert to `*-hdsm.dv8-dsm` and
\code{\link{dv8_hdsmb_sdsmb_to_mdsmb}} to merge DSMs into `*-merge.dv8-dsm`.
}
