% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv8.R
\name{dv8_gitnumstat_to_hdsmb}
\alias{dv8_gitnumstat_to_hdsmb}
\title{Converts a git log text file into a hdsmb binary file}
\usage{
dv8_gitnumstat_to_hdsmb(
  dv8_path,
  git_numstat_path,
  hdsmb_path,
  start,
  stop,
  max_cochange_count = 1000,
  params_output_file = ""
)
}
\arguments{
\item{dv8_path}{path to DV8 binary}

\item{git_numstat_path}{path to local git log text file obtained by \code{\link{dv8_gitlog_to_gitnumstat}}}

\item{hdsmb_path}{name of output hdsmb binary file, defaults to name of input}

\item{start}{start of date/time range in ISO format ie. 2017-07-08T00:00:00Z}

\item{stop}{end of date/time range in ISO format ie. 2018-07-08T00:00:00Z}

\item{max_cochange_count}{maximum count of co-changed files per commit, default 1000}

\item{params_output_file}{output file used to record parameters}
}
\value{
path to hdsmb binary file
}
\description{
An input file *.txt is converted to *.dv8-dsm
}
\seealso{
\code{\link{dv8_hsdsmb_to_decoupling_level}} to calculate decoupling level from hsdsmb binary file

\code{\link{dv8_hsdsmb_to_hierclsxb}} to compute design rule hierarchy from hsdsmb binary file

\code{\link{dv8_hsdsmb_drhier_to_excel}} to export a hsdsmb binary file as an excel file

\code{\link{dv8_hsdsmb_dsmb_to_dsmj}} to export a hsdsmb binary file as a json file
}
