% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv8.R
\name{dv8_hdsmb_sdsmb_to_hsdsmb}
\alias{dv8_hdsmb_sdsmb_to_hsdsmb}
\title{Merge matrices from dv8-dsm file to one new matrix in hsdsm.dv8-dsm}
\usage{
dv8_hdsmb_sdsmb_to_hsdsmb(dv8_dsm_path_1, dv8_dsm_path_2, merged_dv8_dsm_path)
}
\arguments{
\item{dv8_dsm_path_1}{path to the matrix from sdsm.dv8-dsm file}

\item{dv8_dsm_path_2}{path to the matrix from sdsm.dv8-dsm file}

\item{merged_dv8_dsm_path}{path to save the merged matrices in hsdsm.dv8-dsm file}
}
\description{
Merge matrices from dv8-dsm file to one new matrix in hsdsm.dv8-dsm
}
