% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv8.R
\name{dv8_hdsmb_sdsmb_to_hsdsmb}
\alias{dv8_hdsmb_sdsmb_to_hsdsmb}
\title{Merge matrices from dv8-dsm file to one new matrix in hsdsm.dv8-dsm}
\usage{
dv8_hdsmb_sdsmb_to_hsdsmb(hdsmb_path, sdsmb_path, hsdsmb_path)
}
\arguments{
\item{hdsmb_path}{path to the matrix from hdsmb file}

\item{hsdsmb_path}{path to save the merged matrices in hsdsm.dv8-dsm file}

\item{sdsmb}{path to the matrix from sdsmb file}
}
\description{
Merge matrices from dv8-dsm file to one new matrix in hsdsm.dv8-dsm
}
