% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{engagement_sentiment}
\alias{engagement_sentiment}
\title{Engagement Sentiment Metric}
\usage{
engagement_sentiment(
  datetimetz,
  user_name_email,
  polarity,
  lag = 90,
  aggregate_func = mean
)
}
\arguments{
\item{datetimetz}{A POSIXct data table column indicating the timestamp of an author's message}

\item{user_name_email}{A data table column indicating the author of a message}

\item{polarity}{A data table column indicating the sentiment of a message}

\item{lag}{The number of days for each rolling window (90 day default)}

\item{aggregate_func}{The aggregate function to apply to the polarity values in the windows.
The default is mean, but it can be overridden}
}
\description{
Apply an aggregate function to the sentiment (polarity) for each 
rolling window (lag days) from the author (user_name_email).
Creates consecutive non-overlapping windows and only includes complete windows.
}
\references{
Wouter Mulder (2025). Am I finished yet? A discovery of burnout and
ragequits within open-source projects. (Master thesis, Jheronimus Academy of Data Science).
}
