% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_different_branches}
\alias{example_different_branches}
\title{Commit in two different branches}
\usage{
example_different_branches()
}
\value{
git_repo of newly created empty repo
}
\description{
One commit in two different with branches with 1 file each
}
