% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_different_branches}
\alias{example_different_branches}
\title{Example Commit Different Branches}
\usage{
example_different_branches(folder_path = "/tmp", folder_name)
}
\arguments{
\item{folder_path}{The path where the folder will be created}

\item{folder_name}{The name of the folder}
}
\value{
git_repo_path of newly created empty repo
}
\description{
One commit in two different with branches with 1 file each.
}
\details{
Useful to check parser includes commits from different branches.
}
\keyword{internal}
