% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_different_files_commits}
\alias{example_different_files_commits}
\title{Create repo with 2 commits, diff # of files}
\usage{
example_different_files_commits()
}
\value{
git_repo of newly created empty repo
}
\description{
Repo with 2 commits, first one contains 5 files modified, second contains only one
}
