% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_empty_repo}
\alias{example_empty_repo}
\title{Create Empty Repo}
\usage{
example_empty_repo()
}
\value{
git_repo_path of newly created empty repo
}
\description{
Checks if parse_gitlog works on an empty repo.
}
