% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_empty_repo}
\alias{example_empty_repo}
\title{Example Empty Repo}
\usage{
example_empty_repo(folder_path = "/tmp", folder_name)
}
\arguments{
\item{folder_path}{The path where the folder will be created}

\item{folder_name}{The name of the folder}
}
\value{
git_repo_path of newly created empty repo
}
\description{
Creates an empty git repo named "empty_repo".
}
\details{
Useful to test the behavior of git_log exporter and parse_gitlog
on repositories with no commits.
}
\keyword{internal}
