% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_by_filepath_substring}
\alias{filter_by_filepath_substring}
\title{Filter by filepath substring}
\usage{
filter_by_filepath_substring(dt_file, substring, file_column_name)
}
\arguments{
\item{dt_file}{any data.table containing filepaths}

\item{substring}{a character vector of substrings (e.g. c(py,java)) we wish to *filter*}

\item{file_column_name}{a string indicating the column name which contains filepaths}
}
\value{
a data.table which contains does *not* contain filepaths with the specified words
}
\description{
Filters a data.table with filepaths using the specified substring (e.g. remove
all filepaths which contain the word 'test' anywhere in it)
}
\seealso{
\code{\link{parse_gitlog}} and \code{\link{parse_dependencies}} to create dt_file

Other filters:
\code{\link[=filter_by_commit_interval]{filter_by_commit_interval()}},
\code{\link[=filter_by_commit_size]{filter_by_commit_size()}},
\code{\link[=filter_by_file_extension]{filter_by_file_extension()}},
\code{\link[=filter_by_reply_author_substring]{filter_by_reply_author_substring()}},
\code{\link[=filter_by_reply_body_substring]{filter_by_reply_body_substring()}},
\code{\link[=filter_by_reply_subject_substring]{filter_by_reply_subject_substring()}},
\code{\link[=replace_token_regex_with]{replace_token_regex_with()}}
}
\concept{filters}
