% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{format_created_at_from_file}
\alias{format_created_at_from_file}
\title{get the created_at field from a filename}
\usage{
format_created_at_from_file(filename)
}
\arguments{
\item{filename}{the path and the filename}
}
\description{
Function to read a JSON file along a path and return the 'created_at'
date of the first item
}
