% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{format_created_at_from_file}
\alias{format_created_at_from_file}
\title{Retrieve greatest 'created_at' value from file}
\usage{
format_created_at_from_file(file_name, item_path)
}
\arguments{
\item{file_name}{the path and the file name. For example:
../../rawdata/github/kaiaulu/issue_or_pr_comment/sailuh_kaiaulu_issue_or_pr_comment_1701216000_1701261374.json}

\item{item_path}{specifies the level of nesting to look for the created_at value. This was
implemented given that the results of the search endpoint are differently nested than others.}
}
\description{
Function to read a JSON file along a path and return the 'created_at'
date of the greatest value for the issue key. Note that the 'created_at'
value differs in how it is nested. This format is returned by the
issue endpoint currently, but is in level 2 of data returned by search endpoint.
So we allow the input of an item_path parameter to specify the level of nesting
}
\seealso{
\code{link{github_api_project_issue_or_pr_comment_refresh}} to refresh comment data

\code{link{github_api_project_issue_refresh}} to refresh issue data
}
