% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_bugzilla_project_key}
\alias{get_bugzilla_project_key}
\title{Get Bugzilla Project Key}
\usage{
get_bugzilla_project_key(config_file)
}
\arguments{
\item{config_file}{The parsed config file}
}
\value{
the project key name
}
\description{
Gets the project key name for the Bugzilla project
}
