% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_code_language}
\alias{get_code_language}
\title{Get Code Language}
\usage{
get_code_language(project_name)
}
\arguments{
\item{project_name}{the name of the project config file (e.g. "kaiaulu" or "geronimo")}
}
\value{
the code language
}
\description{
Gets the code language of the project
}
