% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_conf}
\alias{get_conf}
\title{Get config.yml File}
\usage{
get_conf(project_name)
}
\arguments{
\item{project_name}{the name of the project config file (e.g. "kaiaulu" or "geronimo")}
}
\value{
the config yml file
}
\description{
Gets the configuration file for any project
}
