% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_cveid_regex}
\alias{get_cveid_regex}
\title{Get CVE ID Regex}
\usage{
get_cveid_regex(project_name)
}
\arguments{
\item{project_name}{the name of the project config file (e.g. "kaiaulu" or "geronimo")}
}
\value{
the commit message CVE
}
\description{
Gets the commit message CVE
}
