% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_dv8_flaws_params}
\alias{get_dv8_flaws_params}
\title{Get Flaws Params}
\usage{
get_dv8_flaws_params(config_file)
}
\arguments{
\item{config_file}{The parsed config file}
}
\value{
the architectural flaws thresholds
}
\description{
Gets the architectural flaws thresholds
}
