% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_dv8_folder_path}
\alias{get_dv8_folder_path}
\title{Get DV8 Project Folder Path}
\usage{
get_dv8_folder_path(project_name)
}
\arguments{
\item{project_name}{the name of the project config file (e.g. "kaiaulu" or "geronimo")}
}
\value{
the dv8 project folder path
}
\description{
Gets the dv8 project folder path
}
