% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_enumeration_commits}
\alias{get_enumeration_commits}
\title{Get Enumeration Commits}
\usage{
get_enumeration_commits(config_file)
}
\arguments{
\item{config_file}{The parsed config file}
}
\value{
the commit intervals as a list
}
\description{
Gets the list of enumerated commit intervals for analysis.
}
