% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_file_extensions}
\alias{get_file_extensions}
\title{Get File Extensions}
\usage{
get_file_extensions(config_file)
}
\arguments{
\item{config_file}{The parsed config file}
}
\value{
the file extensions
}
\description{
Gets the file extensions for filtering
}
