% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_filter_commit_size}
\alias{get_filter_commit_size}
\title{Get Filter Commit Size}
\usage{
get_filter_commit_size(config_file)
}
\arguments{
\item{config_file}{The parsed config file}
}
\value{
the commit size to filter out
}
\description{
Gets the commit size to filter out
}
