% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_git_repo_path}
\alias{get_git_repo_path}
\title{Get Project Git Repo Path}
\usage{
get_git_repo_path(config_file)
}
\arguments{
\item{config_file}{The parsed config file}
}
\value{
the local git repository path specified in `config_file`
}
\description{
Gets the local project git repository path of the project
}
