% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_github_owner}
\alias{get_github_owner}
\title{Get GitHub Owner}
\usage{
get_github_owner(project_name, project_key_index)
}
\arguments{
\item{project_name}{the name of the project config file (e.g. "kaiaulu" or "geronimo")}

\item{project_key_index}{the name of the index of the project key (e.g. "project_key_1" or "project_key_2")}
}
\value{
the GitHub project owner name
}
\description{
Gets the owner of the GitHub repository for a specific project key index
}
