% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_issue_id_regex}
\alias{get_issue_id_regex}
\title{Returns the issue Id regular expression for commit messages.}
\usage{
get_issue_id_regex(config_file)
}
\arguments{
\item{config_file}{The parsed configuration file obtained from \code{\link{parse_config}}.}
}
\value{
The commit message issue Id regular expression.
}
\description{
This function returns the issue Id regular expression for commit
messages, that is specified in the input parameter `config_file`. The input,
`config_file` must be a parsed configuration file. The function will inform
the user if the issue Id regular expression for commit messages exists in the
parsed configuration file, `config_file`.
}
