% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_issue_id_regex}
\alias{get_issue_id_regex}
\title{Get Issue Id Regex}
\usage{
get_issue_id_regex(config_file)
}
\arguments{
\item{config_file}{The parsed config file}
}
\value{
the issue ID regex
}
\description{
Gets the issue ID regex on commit messages
}
