% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_jira_domain}
\alias{get_jira_domain}
\title{Get Jira Project Domain}
\usage{
get_jira_domain(project_name, project_key_index)
}
\arguments{
\item{project_name}{the name of the project config file (e.g. "kaiaulu" or "geronimo")}

\item{project_key_index}{the name of the index of the project key (e.g. "project_key_1" or "project_key_2")}
}
\value{
the Jira domain of the project key index
}
\description{
Gets the Jira domain for a specific project key index
}
