% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_jira_key_indexes}
\alias{get_jira_key_indexes}
\title{Get Jira Issue Tracker Project Key Indexes}
\usage{
get_jira_key_indexes(config_file)
}
\arguments{
\item{config_file}{The parsed config file}
}
\value{
the list of Jira issue tracker project key indexes
}
\description{
Gets the list of Jira issue tracker project key indexes
}
