% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_jira_project_key_name}
\alias{get_jira_project_key_name}
\title{Get Jira Project Key Name}
\usage{
get_jira_project_key_name(config_file, project_key_index)
}
\arguments{
\item{config_file}{The parsed config file}

\item{project_key_index}{The name of the index of the project key (e.g. "project_key_1" or "project_key_2")}
}
\value{
the Jira project key name for project specified by key index `project_key_index`
}
\description{
Gets the Jira project key name for a specific project key index
}
