% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_keep_dependencies_type}
\alias{get_keep_dependencies_type}
\title{Get Keep Dependencies Type}
\usage{
get_keep_dependencies_type(config_file)
}
\arguments{
\item{config_file}{The parsed config file}
}
\value{
the types of dependencies to keep
}
\description{
Gets the type of dependencies to keep
}
