% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_mbox_domain}
\alias{get_mbox_domain}
\title{Get Mbox Archive URL}
\usage{
get_mbox_domain(project_name, project_key_index)
}
\arguments{
\item{project_name}{the name of the project config file (e.g. "kaiaulu" or "geronimo")}

\item{project_key_index}{the name of the index of the project key (e.g. "project_key_1" or "project_key_2")}
}
\value{
the domain of the mailing list archive
}
\description{
Gets the Mbox archive URL for a specific project key index
}
