% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_mbox_key_indexes}
\alias{get_mbox_key_indexes}
\title{Get Mailing List Mod Mbox Key Indexes}
\usage{
get_mbox_key_indexes(project_name)
}
\arguments{
\item{project_name}{the name of the project config file (e.g. "kaiaulu" or "geronimo")}
}
\value{
the list of mod mbox mail key indexes
}
\description{
Gets the list of mod mbox mail key indexes
}
