% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_mbox_mailing_list}
\alias{get_mbox_mailing_list}
\title{Get Mod Mbox Mailing List}
\usage{
get_mbox_mailing_list(project_name, project_key_index)

get_mbox_mailing_list(project_name, project_key_index)
}
\arguments{
\item{project_name}{the name of the project config file (e.g. "kaiaulu" or "geronimo")}

\item{project_key_index}{the name of the index of the project key (e.g. "project_key_1" or "project_key_2")}
}
\value{
the mbox mailing list

the mbox mailing list
}
\description{
Gets the Mbox mailing list for a specific project key index

Gets the Mbox archive type for a specific project key index
}
