% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_mbox_path}
\alias{get_mbox_path}
\title{Get Mbox Save Path}
\usage{
get_mbox_path(config_file, project_key_index)
}
\arguments{
\item{config_file}{The parsed config file}

\item{project_key_index}{The name of the index of the project key (e.g. "project_key_1" or "project_key_2")}
}
\value{
the local Mbox path for project specified by key index `project_key_index`
}
\description{
Gets the local path to store Mbox data for a specific project key index
}
