% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_nvdfeed_folder_path}
\alias{get_nvdfeed_folder_path}
\title{Get Nvd Feed Folder Path}
\usage{
get_nvdfeed_folder_path(config_file)
}
\arguments{
\item{config_file}{The parsed config file}
}
\value{
the folder path with nvd feeds
}
\description{
Gets the folder path with nvd feeds
}
