% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_pattern4_filepath}
\alias{get_pattern4_filepath}
\title{Get Pattern4 Output Filepath}
\usage{
get_pattern4_filepath(config_file)
}
\arguments{
\item{config_file}{The parsed config file}
}
\value{
the Pattern4 output filepath
}
\description{
Gets the Pattern4 filepath where the output is stored
}
