% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_pattern4_filepath}
\alias{get_pattern4_filepath}
\title{Returns the folder path for the output of the pattern4 analysis.}
\usage{
get_pattern4_filepath(config_file)
}
\arguments{
\item{config_file}{The parsed configuration file obtained from \code{\link{parse_config}}.}
}
\value{
The Pattern4 output folder path.
}
\description{
This function returns the folder path that contains the
output of the pattern4 analysis for the project, that is specified in the
input parameter `config_file`. The input, `config_file` must be a parsed
configuration file. The function will inform the user if the folder path
exists in the parsed configuration file, `config_file`.
}
