% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_pattern4_folder_path}
\alias{get_pattern4_folder_path}
\title{Get Pattern4 Class Folder Path}
\usage{
get_pattern4_folder_path(project_name)
}
\arguments{
\item{project_name}{the name of the project config file (e.g. "kaiaulu" or "geronimo")}
}
\value{
the class folder path
}
\description{
Gets the Pattern4 class folder path where the output is stored
}
