% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_srcml_filepath}
\alias{get_srcml_filepath}
\title{Get SrcML Filepath}
\usage{
get_srcml_filepath(config_file)
}
\arguments{
\item{config_file}{The parsed config file}
}
\value{
the SrcML filepath
}
\description{
Gets the SrcML filepath where the output is stored
}
