% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_srcml_filepath}
\alias{get_srcml_filepath}
\title{Get SrcML Filepath}
\usage{
get_srcml_filepath(project_name)
}
\arguments{
\item{project_name}{the name of the project config file (e.g. "kaiaulu" or "geronimo")}
}
\value{
the srcml filepath
}
\description{
Gets the SrcML filepath where the output is stored
}
