% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_substring_filepath}
\alias{get_substring_filepath}
\title{Returns the list of file extensions used for filtering files to remove.}
\usage{
get_substring_filepath(config_file)
}
\arguments{
\item{config_file}{The parsed configuration file.}
}
\value{
The list of file extensions to remove.
}
\description{
This function returns the list of file extensions that will be
used for filtering files specified in the input parameter `config_file`. The
input, `config_file` must be a parsed configuration file. The function will
inform the user if the list of file extensions exists in the parsed
configuration file, `config_file`.
}
