% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_substring_filepath}
\alias{get_substring_filepath}
\title{Get Substring Filepaths}
\usage{
get_substring_filepath(config_file)
}
\arguments{
\item{config_file}{The parsed config file}
}
\value{
the substring filepaths
}
\description{
Gets the substring filepaths for filtering
}
