% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_tools_project}
\alias{get_tools_project}
\title{Get Tool Path}
\usage{
get_tools_project(tool_name)
}
\arguments{
\item{tool_name}{the name of the tool (e.g. "perceval" or "dv8")}
}
\value{
the tool path from tools.yml
}
\description{
Gets the tool path from the tools.yml file
}
