% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_topics}
\alias{get_topics}
\title{Get Topics}
\usage{
get_topics(config_file)
}
\arguments{
\item{config_file}{The parsed config file}
}
\value{
the topics
}
\description{
Gets the topics and keywords for analysis
}
