% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_topics}
\alias{get_topics}
\title{Returns the list of topics and keywords for analysis.}
\usage{
get_topics(config_file)
}
\arguments{
\item{config_file}{The parsed configuration file obtained from \code{\link{parse_config}}.}
}
\value{
The list of keywords and topics for analysis.
}
\description{
This function returns the list of keywords and topics for
analysis, that is specified in the input parameter `config_file`. The
input, `config_file` must be a parsed configuration file. The function will
inform the user if the list of keywords and topics exists in the parsed
configuration file, `config_file`.
}
