% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_uctags_line_types}
\alias{get_uctags_line_types}
\title{Get Uctags Line Types}
\usage{
get_uctags_line_types(config_file)
}
\arguments{
\item{config_file}{The parsed config file}
}
\value{
the uctags line type
}
\description{
Gets the Uctags keep lines type
}
