% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_uctags_line_types}
\alias{get_uctags_line_types}
\title{Get Uctags Line Types}
\usage{
get_uctags_line_types(project_name)
}
\arguments{
\item{project_name}{the name of the project config file (e.g. "kaiaulu" or "geronimo")}
}
\value{
the uctags line type
}
\description{
Gets the Uctags keep lines type
}
