% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_window_end_commit}
\alias{get_window_end_commit}
\title{Get Window End Commit}
\usage{
get_window_end_commit(project_name)
}
\arguments{
\item{project_name}{the name of the project config file (e.g. "kaiaulu" or "geronimo")}
}
\value{
the analysis window size
}
\description{
Gets the analysis window end commit
}
