% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_window_size}
\alias{get_window_size}
\title{Get Analysis Window Size}
\usage{
get_window_size(config_file)
}
\arguments{
\item{config_file}{The parsed config file}
}
\value{
the analysis window size
}
\description{
Gets the analysis window size
}
