% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_window_size}
\alias{get_window_size}
\title{Returns the size of a window for analysis.}
\usage{
get_window_size(config_file)
}
\arguments{
\item{config_file}{The parsed configuration file.}
}
\value{
The size of a window for analysis.
}
\description{
This function returns the size of a window, that is
specified in the input parameter `config_file`. The input, `config_file`
must be a parsed configuration file. The function will inform the user if
the window size exists in the parsed configuration file, `config_file`.
}
