% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_window_start_commit}
\alias{get_window_start_commit}
\title{Get Window Start Commit}
\usage{
get_window_start_commit(config_file)
}
\arguments{
\item{config_file}{The parsed config file}
}
\value{
the analysis window start commit
}
\description{
Gets the analysis window start commit
}
