% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_add}
\alias{git_add}
\title{Git Add}
\usage{
git_add(git_repo, folder_path, filepath)
}
\arguments{
\item{git_repo}{The git repo path}

\item{folder_path}{The worktree path}

\item{filepath}{The filepath we want to add}
}
\value{
The path to the sample .git file.
}
\description{
Performs the `git add` command, which stages a given
file in the repo.
}
