% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_blame}
\alias{git_blame}
\title{Git blame wrapper}
\usage{
git_blame(git_repo_path, flags, commit_hash, file_path)
}
\arguments{
\item{git_repo_path}{The git repo pat}

\item{flags}{Optional flags for git log command}

\item{commit_hash}{The commit hash of the file we will blame}

\item{file_path}{The file we will blame}
}
\description{
Git blame wrapper
}
