% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_commit}
\alias{git_commit}
\title{Writes a git commit to the git log - git --git-dir sample/.git --work-tree sample commit -m 'hello world commit'}
\usage{
git_commit(git_repo, folder_path, commit_msg, author, email)
}
\arguments{
\item{git_repo}{The git repo path}

\item{folder_path}{The worktree path}

\item{commit_msg}{The commit msg}

\item{author}{The author of the commit}

\item{email}{The email of whoever made the commit}
}
\value{
The path to the sample .git file.
}
\description{
Writes a git commit with author and email to the folder path
}
