% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_commit}
\alias{git_commit}
\title{Git Commit}
\usage{
git_commit(git_repo, folder_path, commit_msg, author, email)
}
\arguments{
\item{git_repo}{The git repo path.}

\item{folder_path}{The worktree path.}

\item{commit_msg}{The commit associated with the commit.}

\item{author}{The author associated with the commit.}

\item{email}{The email of the author associated with the commit.}
}
\value{
The path to the sample .git file.
}
\description{
Executes the `git commit` command on the target
git repository.
}
