% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_create_sample_log}
\alias{git_create_sample_log}
\title{Creates a sample git log with one commit}
\usage{
git_create_sample_log(folder_path = "/tmp")
}
\arguments{
\item{folder_path}{The worktree path}

\item{file_to_add}{The file we want to add}

\item{git_repo}{The git repo path}

\item{commit_msg}{The commit msg}

\item{author}{The author of the commit}

\item{email}{The email of whoever made the commit}
}
\value{
The path to the sample .git file.
}
\description{
This is a SetUp helper function for Kaiaulu unit tests
that manipulates git logs.
}
\details{
A folder kaiaulu_sample is created in /tmp by default. A file,
hello.R with a single print is then added to the folder.
Git init is performed, the file is git add, and commit to
the git log.
}
\seealso{
Other {unittest}: 
\code{\link{git_delete_sample_log}()}
}
\concept{{unittest}}
