% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_head}
\alias{git_head}
\title{Gets the current commit hash head of the git repo}
\usage{
git_head(git_repo_path)
}
\arguments{
\item{git_repo_path}{The git repo path}
}
\value{
A commit hash character
}
\description{
Gets the current commit hash head of the git repo
}
