% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_api_project_issue_or_pr_comment_refresh}
\alias{github_api_project_issue_or_pr_comment_refresh}
\title{Download Project issues or pr comments after certain date}
\usage{
github_api_project_issue_or_pr_comment_refresh(
  owner,
  repo,
  token,
  file_save_path,
  verbose = TRUE
)
}
\arguments{
\item{owner}{GitHub's repository owner (e.g. sailuh)}

\item{repo}{GitHub's repository name (e.g. kaiaulu)}

\item{token}{Your GitHub API token}

\item{file_save_path}{the save path for the issue comments folder}

\item{verbose}{boolean value. When set to true, it prints operational messages including
greatest dates and the file name that contains the greatest date.}
}
\description{
Uses the adopted file name convention by \code{\link{github_api_iterate_pages}} to identify
the latest downloaded Github created_at date among the directory(intended to be the comment folder).
It uses this date to construct a query and calls \code{\link{github_api_project_issue_or_pr_comments}}
}
\details{
If no files exist in the file_save_path,\code{link{github_api_project_issue_or_pr_comments}}
is called with no additional query and all comments are downloaded.
}
\references{
For details, see \url{https://docs.github.com/en/rest/reference/issues#list-repository-issues}.
}
\seealso{
\code{link{github_api_project_issue_or_pr_comments}} to download all comment data

\code{link{format_created_at_from_file}} for function that iterates through
a .json file and returns the greatest 'created_at' value

\code{link{github_api_iterate_pages}} to write data returned by this function to file as .json

\code{link{github_api_project_issue_or_pr_comments}} to call issue/comments endpoint
}
