% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_api_project_issue_or_pr_comments}
\alias{github_api_project_issue_or_pr_comments}
\title{Download Project Issue's or Pull Request's Comments}
\usage{
github_api_project_issue_or_pr_comments(owner, repo, token)
}
\arguments{
\item{owner}{GitHub's repository owner (e.g. sailuh)}

\item{repo}{GitHub's repository name (e.g. kaiaulu)}

\item{token}{Your GitHub API token}
}
\description{
Download Issues' or Pull Request's Comments from "GET /repos/{owner}/{repo}/issues/comments" endpoint.
}
\references{
For details, see \url{https://docs.github.com/en/rest/reference/issues#list-issue-comments-for-a-repository} and
\url{https://docs.github.com/en/rest/guides/working-with-comments#pull-request-comments}.
}
