% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_api_project_issue_refresh}
\alias{github_api_project_issue_refresh}
\title{Download Project Issues after a date}
\usage{
github_api_project_issue_refresh(owner, repo, token)
}
\arguments{
\item{owner}{GitHub's repository owner (e.g. sailuh)}

\item{repo}{GitHub's repository name (e.g. kaiaulu)}

\item{token}{Your GitHub API token}

\item{created}{Github's created at date}
}
\description{
Returns issue data that has not already been downloaded
Gets the name of the file with the most recent data along the designated save path.
Extracts the greatest 'created_at' date from that file
Calls search/issues endpoint to download issues created after that date
}
\references{
For details, see \url{https://docs.github.com/en/rest/reference/issues#list-repository-issues}.
}
