% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_api_project_pull_request_refresh}
\alias{github_api_project_pull_request_refresh}
\title{Download Project Pull Requests Description (First Comment) Refresh}
\usage{
github_api_project_pull_request_refresh(
  owner,
  repo,
  token,
  save_path_pull_request,
  verbose = TRUE
)
}
\arguments{
\item{owner}{GitHub's repository owner (e.g. sailuh)}

\item{repo}{GitHub's repository name (e.g. kaiaulu)}

\item{token}{Your GitHub API token}

\item{verbose}{True if messages should be displayed during execution.}

\item{file_save_path}{the save path for the pr folder}
}
\description{
Downloads the Pull Request First Comment (i.e. the Pull Request "Description"). This serves
as a convenience function to \code{\link{github_api_project_issue_refresh}} with parameter
issue_or_pr = "is:pull-request". See referenced function for details.
}
\references{
For details, see For details, see \url{https://docs.github.com/en/rest/pulls/comments?apiVersion=2022-11-28#about-pull-request-review-comments}.
}
\seealso{
\code{link{github_api_project_pull_request}} to download all pull request  data

\code{link{format_created_at_from_file}} for function that iterates through
a .json file and returns the greatest 'created_at' value

\code{link{github_api_iterate_pages}} to write data returned by this function to file as .json
}
