% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_api_project_pull_request_review_refresh}
\alias{github_api_project_pull_request_review_refresh}
\title{Download Project Pull Request Reviews Refresh}
\usage{
github_api_project_pull_request_review_refresh(
  owner,
  repo,
  token,
  save_path_pull_request,
  save_path_pr_reviews,
  verbose = TRUE
)
}
\arguments{
\item{owner}{GitHub's repository owner (e.g. sailuh)}

\item{repo}{GitHub's repository name (e.g. kaiaulu)}

\item{token}{Your GitHub API token}

\item{save_path_pull_request}{save path for the pull requests to retrieve pull numbers.}

\item{save_path_pr_reviews}{the save path for the pr reviews folder}

\item{verbose}{boolean value. When set to true, it prints operational messages including
greatest dates and the file name that contains the greatest date.}
}
\description{
Pull Request Reviews require a specific Pull Number to complete the API Request.
To download all pull request reviews we must iterate through all pull numbers and download one file per pull request.
Using the Pull Request endpoint to download the pull numbers, we now have a basis for iteration.
The naming convention for downloading each pull request would be sailuh_kaiaulu_330 (owner_repo_pullnumber).
}
\references{
For details, see For details, see \url{https://docs.github.com/en/rest/pulls/comments?apiVersion=2022-11-28#about-pull-request-review-comments}.
}
\seealso{
\code{link{github_api_pr_reviews}} to download pull request review data.

\code{link{github_api_project_pull_request}} to download pull request data and retrieve pull numbers.
}
