% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_parse_project_pull_request}
\alias{github_parse_project_pull_request}
\title{Parse Pull Requests JSON to Table}
\usage{
github_parse_project_pull_request(api_responses)
}
\arguments{
\item{api_responses}{API response obtained from github_api_* function.}
}
\description{
Note not all columns available in the downloaded json are parsed.
}
