% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv8.R
\name{gitlog_to_hdsmj}
\alias{gitlog_to_hdsmj}
\title{Convert gitlog table to a hdsm.json}
\usage{
gitlog_to_hdsmj(gitlog_table, project_name, is_sorted = FALSE)
}
\arguments{
\item{gitlog_table}{parsed gitlog table}

\item{project_name}{project name for the hdsm.json}

\item{is_sorted}{whether the json is sorted by src and dest}
}
\description{
Converts a gitlog table into an *-hdsm.json binary.
}
\seealso{
Other dv8: 
\code{\link{dependencies_to_sdsmj}()},
\code{\link{dv8_clsxb_to_clsxj}()},
\code{\link{parse_dv8_clusters}()}
}
\concept{dv8}
