% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_to_dsmj}
\alias{graph_to_dsmj}
\title{Create a dsm.json file given either a dependencies table from \code{\link{transform_dependencies_to_sdsmj}}
or a gitlog table from \code{\link{transform_gitlog_to_hdsmj}}.}
\usage{
graph_to_dsmj(graph, dsm_path, is_directed, is_sorted = FALSE)
}
\arguments{
\item{graph}{the table containing the nodes and edgelist of the graph of dependencies or the gitlog.}

\item{dsm_path}{path to save either the *-sdsm.json from \code{\link{transform_dependencies_to_sdsmj}} or the
*-hdsmj.json from \code{\link{transform_gitlog_to_hdsmj}}}

\item{is_directed}{whether the graph is directed. If false, cells will be doubled
with src/dest switched to represent an undirected graph as a directed graph.}

\item{is_sorted}{whether to sort the variables (filenames) in the dsm.json files}
}
\value{
either the path to a sdsm.json (from dependencies table) or a hdsmj.json (from gitlog table).
}
\description{
# Returns either a sdsm.json (from dependencies table) or a hdsmj.json (from gitlog table).
}
\seealso{
\code{\link{transform_dependencies_to_sdsmj}} and \code{\link{transform_gitlog_to_hdsmj}}

Other dv8: 
\code{\link{dependencies_to_sdsmj}()},
\code{\link{dv8_clsxb_to_clsxj}()},
\code{\link{dv8_clsxj_to_clsxb}()},
\code{\link{dv8_depends_to_sdsmj}()},
\code{\link{dv8_dsmj_to_dsmb}()},
\code{\link{dv8_hdsmb_sdsmb_to_mdsmb}()},
\code{\link{dv8_mdsmb_to_flaws}()},
\code{\link{gitlog_to_hdsmj}()},
\code{\link{parse_dv8_architectural_flaws}()},
\code{\link{parse_dv8_clusters}()},
\code{\link{transform_dependencies_to_sdsmj}()},
\code{\link{transform_gitlog_to_hdsmj}()}
}
\concept{dv8}
