% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{io_create_folder}
\alias{io_create_folder}
\title{This function creates a new folder}
\usage{
io_create_folder(folderName, body)
}
\arguments{
\item{folderName}{Name of the folder you are creating}

\item{body}{The contents of the folder, what files it will consist of}
}
\value{
The new folder with its updated contents
}
\description{
This function creates a new folder
}
