% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{io_make_sample_file}
\alias{io_make_sample_file}
\title{This function creates a sample file}
\usage{
io_make_sample_file(file_path, body)
}
\arguments{
\item{file_path}{The path of the file that is being saved}

\item{body}{The contents of the file}
}
\value{
file with specified path and body
}
\description{
Used to create sample file for fake data
}
