% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{jira_author_communication_count}
\alias{jira_author_communication_count}
\title{Jira Author Communication Count Metric}
\usage{
jira_author_communication_count(
  comments_dt,
  comment_created_col = "comment_created_datetimetz",
  comment_author_col = "comment_author_name",
  lag = 90
)
}
\arguments{
\item{comments_dt}{a parsed jira comments table obtained from \code{\link{parse_jira}}}

\item{comment_created_col}{the column name in `comments_dt` which contains the comment creation time}

\item{comment_author_col}{the column name in `comments_dt` which contains the comment author name}

\item{lag}{the number of days to look back for the rolling window (default is 90)}
}
\value{
A three column data.table of the form datetimetz | comment_author_name | comment_count
}
\description{
Calculates the number of messages/comments an author has sent in Jira within
a 90-day rolling window.
}
\seealso{
\code{\link{parse_jira}} to obtain the comments table

Other metrics: 
\code{\link{commit_message_id_coverage}()},
\code{\link{metric_churn}()},
\code{\link{metric_churn_per_commit_interval}()},
\code{\link{metric_churn_per_commit_per_file}()},
\code{\link{metric_file_bug_churn}()},
\code{\link{metric_file_bug_frequency}()},
\code{\link{metric_file_churn}()},
\code{\link{metric_file_non_bug_churn}()},
\code{\link{metric_file_non_bug_frequency}()}
}
\concept{metrics}
