% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{make_jira_issue}
\alias{make_jira_issue}
\title{Creates a sample jira issue without comment and two components}
\usage{
make_jira_issue(
  jira_domain_url,
  issue_key,
  issue_type,
  status,
  resolution,
  title,
  description,
  components,
  creator_name,
  reporter_name,
  assignee_name,
  comments = NULL
)
}
\arguments{
\item{jira_domain_url}{URL of JIRA domain (e.g. "https://project.org/jira")}

\item{issue_key}{issue key of JIRA issue (e.g. "PROJECT-68" or "GERONIMO-6723)}

\item{issue_type}{type of JIRA issue (e.g. "New Feature", "Task", "Bug")}

\item{status}{status of issue for development (e.g. "In Progress")}

\item{resolution}{name of resolution for issue (e.g. "Fixed")}

\item{title}{summary of the issue (e.g. "Site Keeps Crashing")}

\item{description}{more detailed description of issue (e.g. "The program keeps crashing because this reason")}

\item{components}{components of issue separate by ; (e.g. "x-core;x-spring")}

\item{creator_name}{name of creator of issue (e.g. "John Doe")}

\item{reporter_name}{name of reporter of issue (e.g. "Jane Doe")}

\item{assignee_name}{name of person the issue is being assigned to (e.g. "Joe Schmo")}

\item{comments}{a character vector where each element is a comment string (e.g. c("This is first comment", "This is second comment"))}
}
\value{
The path to the sample .json file.
}
\description{
This example replicates Kaiaulu issue #244
}
\seealso{
Other {unittest}: 
\code{\link{git_create_sample_log}()},
\code{\link{git_delete_sample_log}()},
\code{\link{jira_delete_sample_log}()}
}
\concept{{unittest}}
