% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{make_jira_issue}
\alias{make_jira_issue}
\title{Creates a sample jira issue without comment and two components}
\usage{
make_jira_issue(
  jira_domain_url,
  issue_key,
  issue_type_description,
  summary,
  components,
  description,
  reporter_name,
  assignee_name,
  status_description
)
}
\value{
The path to the sample .json file.
}
\description{
This example replicates Kaiaulu issue #244
}
\seealso{
Other {unittest}: 
\code{\link{git_create_sample_log}()},
\code{\link{git_delete_sample_log}()},
\code{\link{jira_delete_sample_log}()}
}
\concept{{unittest}}
