% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{make_jira_issue_tracker}
\alias{make_jira_issue_tracker}
\title{Make Jira Issue Tracker}
\usage{
make_jira_issue_tracker(issues, save_filepath)
}
\arguments{
\item{issues}{list of issues obtained from cell for \code{\link{make_jira_issue}}.}

\item{save_filepath}{the filepath where the JSON should be saved, including
file name and extension.}
}
\value{
The `save_filepath` specified.
}
\description{
Create a full JIRA Issue Tracker with multiple issues using
\code{\link{make_jira_issue}}.
}
