% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{metric_file_churn}
\alias{metric_file_churn}
\title{File Churn}
\usage{
metric_file_churn(project_git)
}
\arguments{
\item{project_git}{a parsed git log obtained from \code{\link{parse_gitlog}}}
}
\value{
a two column data.table of the form file_pathname | file_churn
}
\description{
The total churn of a file
}
\seealso{
Other metrics: 
\code{\link{commit_message_id_coverage}()},
\code{\link{engagement_communication}()},
\code{\link{metric_churn}()},
\code{\link{metric_churn_per_commit_interval}()},
\code{\link{metric_churn_per_commit_per_file}()},
\code{\link{metric_file_bug_churn}()},
\code{\link{metric_file_bug_frequency}()},
\code{\link{metric_file_non_bug_churn}()},
\code{\link{metric_file_non_bug_frequency}()},
\code{\link{productivity_author_churn}()},
\code{\link{productivity_author_commits}()}
}
\concept{metrics}
