% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{metric_file_non_bug_churn}
\alias{metric_file_non_bug_churn}
\title{File Non Bug Churn}
\usage{
metric_file_non_bug_churn(project_git, jira_issues)
}
\arguments{
\item{project_git}{a parsed git log obtained from \code{\link{parse_commit_message_id}}}

\item{jira_issues}{a parsed jira issue log obtained from \code{\link{parse_jira}}}
}
\value{
a two column data.table of the form file_pathname | file_non_bug_churn
}
\description{
The total churn sum of commits of all closed non-bug type issues the file was involved.
}
\seealso{
Other metrics:
\code{\link[=commit_message_id_coverage]{commit_message_id_coverage()}},
\code{\link[=metric_churn]{metric_churn()}},
\code{\link[=metric_churn_per_commit_interval]{metric_churn_per_commit_interval()}},
\code{\link[=metric_churn_per_commit_per_file]{metric_churn_per_commit_per_file()}},
\code{\link[=metric_file_bug_churn]{metric_file_bug_churn()}},
\code{\link[=metric_file_bug_frequency]{metric_file_bug_frequency()}},
\code{\link[=metric_file_churn]{metric_file_churn()}},
\code{\link[=metric_file_non_bug_frequency]{metric_file_non_bug_frequency()}}
}
\concept{metrics}
