% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{metric_file_non_bug_frequency}
\alias{metric_file_non_bug_frequency}
\title{File Non Bug Frequency}
\usage{
metric_file_non_bug_frequency(project_git, jira_issues)
}
\arguments{
\item{project_git}{a parsed git log obtained from \code{\link{parse_commit_message_id}}}

\item{jira_issues}{a parsed jira issue log obtained from \code{\link{parse_jira}}}
}
\value{
a two column data.table of the form file_pathname | non_file_bug_frequency
}
\description{
The total number of commits of all closed non-bug type issues the file was involved.
}
\seealso{
Other metrics: 
\code{\link{commit_message_id_coverage}()},
\code{\link{jira_author_communication_count}()},
\code{\link{metric_churn}()},
\code{\link{metric_churn_per_commit_interval}()},
\code{\link{metric_churn_per_commit_per_file}()},
\code{\link{metric_file_bug_churn}()},
\code{\link{metric_file_bug_frequency}()},
\code{\link{metric_file_churn}()},
\code{\link{metric_file_non_bug_churn}()},
\code{\link{productivity_author_churn}()},
\code{\link{productivity_author_commits}()}
}
\concept{metrics}
