% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{model_directed_graph}
\alias{model_directed_graph}
\title{Create a directed graph model}
\usage{
model_directed_graph(edgelist, is_bipartite, color, aggregate_duplicate = TRUE)
}
\arguments{
\item{edgelist}{a 2-column data.table containing named columns
`from` and `to` in any order.}

\item{is_bipartite}{boolean specifying if network is bipartite: TRUE or FALSE}

\item{color}{a character vector of length 1 or 2 specifying the hexacolor}

\item{aggregate_duplicate}{If duplicated rows are encoutered, define a weight column which counts them}
}
\value{
a named list list(nodes,edgelist).
}
\description{
Create a directed graph model
}
