% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{motif_factory}
\alias{motif_factory}
\title{Motif Factory}
\usage{
motif_factory(type)
}
\arguments{
\item{type}{The type of motif to construct.}
}
\description{
A simple factory function to generate available
motifs.
}
\seealso{
Other motif: 
\code{\link{motif_factory_anti_square}()},
\code{\link{motif_factory_anti_triangle}()},
\code{\link{motif_factory_square}()},
\code{\link{motif_factory_triangle}()}
}
\concept{motif}
