% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{openhub_api_analyses}
\alias{openhub_api_analyses}
\title{Downloads a page of Analysis XML Response File.}
\usage{
openhub_api_analyses(token, save_folder_path, project_id, project_name)
}
\arguments{
\item{token}{Your OpenHub API token.}

\item{save_folder_path}{A folder path to save the downloaded XML page "as-is".}

\item{project_id}{The unique ID of the project (e.g. 3562).}

\item{project_name}{The unique name of the project (e.g. "Apache Tomcat").}
}
\value{
The response XML single page file that contains an analysis item.
}
\description{
Downloads an XML response file from the "analysis" collection endpoint.
The `openhub_api_parameters` list must contain a "project_id" key-value pair
for use in the URL for the project's id. The XML response file will return
a single page containing the analysis item.
}
\references{
For analysis collection details, see \url{https://github.com/blackducksoftware/ohloh_api/blob/main/reference/analysis.md}.
}
