% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{openhub_api_analyses}
\alias{openhub_api_analyses}
\title{Downloads a single page of Analysis XML Response File.}
\usage{
openhub_api_analyses(token, save_folder_path, project_id, project_name)
}
\arguments{
\item{token}{Your OpenHub API token.}

\item{save_folder_path}{A folder path to save the downloaded XML page "as-is".}

\item{project_id}{The unique ID of the project (e.g. 3562).}

\item{project_name}{The unique name of the project (e.g. "Apache Tomcat").}
}
\description{
Downloads an XML response file from the "analysis" collection
endpoint. The XML response file will return a single page containing the
single analysis item. The XML response file is saved to disk using the
function \code{\link{openhub_download}} with the file name
"analysis_projectname_unixtimestamp.xml", where "projectname" is the
sanitized version of the name of the project from `project_name`, and
"unixtimestamp" is the integer unix timestamp derived from the most recent
time this analysis collection was updated for the project ("updated_at"
field in the analysis collection).
}
\references{
For analysis collection details, see \url{https://github.com/blackducksoftware/ohloh_api/blob/main/reference/analysis.md}.
}
