% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{openhub_api_iterate_pages}
\alias{openhub_api_iterate_pages}
\title{placeholder}
\usage{
openhub_api_iterate_pages(
  token,
  openhub_api_function,
  openhub_api_parameters,
  max_pages = NULL
)
}
\arguments{
\item{token}{placeholder}

\item{openhub_api_function}{placeholder}

\item{openhub_api_parameters}{placeholder}

\item{max_pages}{placeholder}
}
\value{
placeholder
}
\description{
placeholder
}
