% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{openhub_api_organizations}
\alias{openhub_api_organizations}
\title{Download a page of Organization XML Response File.}
\usage{
openhub_api_organizations(token, save_folder_path, html_url_or_name)
}
\arguments{
\item{token}{Your OpenHub API token.}

\item{save_folder_path}{A folder path to save the downloaded XML page "as-is".}

\item{html_url_or_name}{Either the URL for the organization page on OpenHub (e.g. "https://openhub.net/orgs/apache") or the short, unique, handle for the organization (e.g. "apache").}
}
\value{
The response XML single page file that contains a single organization item.
}
\description{
Downloads an XML response file from "organization" collection
endpoint. The XML response file will return a single page containing the
organization item.
}
\references{
For organization collection details, see \url{https://github.com/blackducksoftware/ohloh_api/blob/main/reference/organization.md}.
}
