% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{openhub_api_portfolio_projects}
\alias{openhub_api_portfolio_projects}
\title{Downloads a page of Portfolio Projects XML Response File.}
\usage{
openhub_api_portfolio_projects(
  token,
  save_folder_path,
  org_name,
  iterating = FALSE,
  page = 1
)
}
\arguments{
\item{token}{Your OpenHub API token.}

\item{save_folder_path}{A folder path to save the downloaded XML page "as-is".}

\item{org_name}{The short, unique, handle for the organization taken from the organization's URL (e.g. "apache" from "https://openhub.net/orgs/apache").}

\item{iterating}{If TRUE, downloading will disabled and it is assumed to be handled by \code{\link{openhub_api_iterate_pages}} to synchronize timestamps, if FALSE, downloading will occur (Default set to FALSE).}

\item{page}{The page number to index (Default set to 1) (e.g. 1).}
}
\value{
The response XML single page file that contains a list of portfolio project items.
}
\description{
Downloads an XML response file from the "portfolio_projects" collection
endpoint. The `openhub_api_parameters` list must contain a "portfolio_project_site"
key-value pair for use as part of the URL. The XML response file will return a page
of items which can be indexed by the numeric `page` parameter.
}
\references{
For portfolio projects collection details, see \url{https://github.com/blackducksoftware/ohloh_api/blob/main/reference/portfolio_projects.md}.
}
