% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{openhub_api_projects}
\alias{openhub_api_projects}
\title{Downloads a page of Project XML Response File.}
\usage{
openhub_api_projects(
  token,
  save_folder_path,
  project_name,
  iterating = FALSE,
  page = 1
)
}
\arguments{
\item{token}{Your OpenHub API token.}

\item{save_folder_path}{A folder path to save the downloaded XML page "as-is".}

\item{project_name}{The unique name of the project (e.g. "Apache Tomcat").}

\item{iterating}{If TRUE, downloading will disabled and it is assumed to be handled by \code{\link{openhub_api_iterate_pages}} to synchronize timestamps, if FALSE, downloading will occur (Default set to FALSE).}

\item{page}{The page number to index (Default set to 1) (e.g. 1).}
}
\value{
The response XML single page file that contains a list of project items.
}
\description{
Downloads an XML response file from the "project" collection endpoint.
The `openhub_api_parameters` list must contain a "project_name" key-value pair
to query the project's name. The XML response file will return a page
of items which can be indexed by the numeric `page` parameter.
}
\references{
For project collection details, see \url{https://github.com/blackducksoftware/ohloh_api/blob/main/reference/project.md}.
}
