% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{openhub_download}
\alias{openhub_download}
\title{OpenHub API Response Downloader}
\usage{
openhub_download(save_folder_path, unique_information, api_responses)
}
\arguments{
\item{save_folder_path}{A folder path to save the downloaded XML pages "as-is".}

\item{unique_information}{Identifying information about the type of XML responses.}

\item{api_responses}{A list of XML responses obtained from an openhub_api_* function.}
}
\description{
Stores a list of XML response files `api_responses` into a
defined folder path `save_folder_path`.
}
