% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{openhub_parse_projects}
\alias{openhub_parse_projects}
\title{placeholder}
\usage{
openhub_parse_projects(api_responses, openhub_api_parameters)
}
\arguments{
\item{api_responses}{placeholder}

\item{openhub_api_parameters}{placeholder}
}
\value{
placeholder
}
\description{
placeholder
}
