% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugzilla.R
\name{parse_bugzilla_perceval_traditional_issue_comments}
\alias{parse_bugzilla_perceval_traditional_issue_comments}
\title{Parse Bugzilla data obtained from Perceval traditional Bugzilla backend}
\usage{
parse_bugzilla_perceval_traditional_issue_comments(
  bugzilla_json_path,
  comments = FALSE
)
}
\arguments{
\item{bugzilla_json_path}{json path for downloaded bugzilla JSON from \code{\link{download_bugzilla_perceval_traditional_issue_comments}}}

\item{comments}{if true, the comments are parsed along with the issues}
}
\value{
data table with parsed bugzilla data
}
\description{
Parse Bugzilla data obtained from Perceval traditional Bugzilla backend
}
\seealso{
\code{\link{download_bugzilla_perceval_traditional_issue_comments}} a downloader function to download bugzilla data with perceval

Other parsers: 
\code{\link{build_understand_project}()},
\code{\link{export_understand_dependencies}()},
\code{\link{parse_bugzilla_perceval_rest_issue_comments}()},
\code{\link{parse_bugzilla_rest_comments}()},
\code{\link{parse_bugzilla_rest_issues}()},
\code{\link{parse_bugzilla_rest_issues_comments}()},
\code{\link{parse_commit_message_id}()},
\code{\link{parse_dependencies}()},
\code{\link{parse_dv8_clusters}()},
\code{\link{parse_github_latest_date}()},
\code{\link{parse_gitlog}()},
\code{\link{parse_jira}()},
\code{\link{parse_jira_latest_date}()},
\code{\link{parse_jira_rss_xml}()},
\code{\link{parse_mbox}()},
\code{\link{parse_mbox_latest_date}()},
\code{\link{parse_nvdfeed}()},
\code{\link{parse_understand_dependencies}()}
}
\concept{parsers}
