% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_bugzilla_rest_comments}
\alias{parse_bugzilla_rest_comments}
\title{Parse Bugzilla comments data obtained from json files from Bugzilla crawler \code{\link{parse_bugzilla_rest_comments}}}
\usage{
parse_bugzilla_rest_comments(comments_folder_path)
}
\arguments{
\item{comments_folder_path}{path to the comments folder that contains json file with Bugzilla data inside}
}
\value{
data table with parsed bugzilla comments data
}
\description{
Parse Bugzilla comments data obtained from json files from Bugzilla crawler \code{\link{parse_bugzilla_rest_comments}}
}
\seealso{
\code{\link{parse_bugzilla_rest_comments}} a downloader function to download bugzilla data with perceval

Other parsers: 
\code{\link{parse_bugzilla_rest_issues}()},
\code{\link{parse_commit_message_id}()},
\code{\link{parse_dependencies}()},
\code{\link{parse_dv8_clusters}()},
\code{\link{parse_gitlog}()},
\code{\link{parse_jira}()},
\code{\link{parse_mbox}()},
\code{\link{parse_nvdfeed}()}
}
\concept{parsers}
