% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugzilla.R
\name{parse_bugzilla_rest_comments}
\alias{parse_bugzilla_rest_comments}
\title{Parse Bugzilla comments data obtained from json files from Bugzilla crawler \code{\link{parse_bugzilla_rest_comments}}}
\usage{
parse_bugzilla_rest_comments(comments_folder_path)
}
\arguments{
\item{comments_folder_path}{path to the comments folder that contains json file with Bugzilla data inside}
}
\value{
data table with parsed Bugzilla comments data
}
\description{
Parse Bugzilla comments data obtained from json files from Bugzilla crawler \code{\link{parse_bugzilla_rest_comments}}
}
\seealso{
Other parsers: 
\code{\link{parse_bugzilla_perceval_rest_issue_comments}()},
\code{\link{parse_bugzilla_perceval_traditional_issue_comments}()},
\code{\link{parse_bugzilla_rest_issues_comments}()},
\code{\link{parse_bugzilla_rest_issues}()},
\code{\link{parse_commit_message_id}()},
\code{\link{parse_dependencies}()},
\code{\link{parse_dv8_clusters}()},
\code{\link{parse_gitlog}()},
\code{\link{parse_jira_rss_xml}()},
\code{\link{parse_jira}()},
\code{\link{parse_mbox}()},
\code{\link{parse_nvdfeed}()}
}
\concept{parsers}
