% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugzilla.R
\name{parse_bugzilla_rest_issues}
\alias{parse_bugzilla_rest_issues}
\title{Parse Bugzilla issues data obtained from json files from Bugzilla crawler}
\usage{
parse_bugzilla_rest_issues(issues_folder_path)
}
\arguments{
\item{issues_folder_path}{path to the issue folder that contains json file with Bugzilla data inside}
}
\value{
data table with parsed Bugzilla issues data
}
\description{
Parse Bugzilla issues data obtained from json files from Bugzilla crawler
}
\seealso{
\code{\link{download_bugzilla_rest_issues_comments}} a downloader function to parse Bugzilla issues and comments data

Other parsers:
\code{\link[=build_understand_project]{build_understand_project()}},
\code{\link[=export_understand_dependencies]{export_understand_dependencies()}},
\code{\link[=parse_bugzilla_perceval_rest_issue_comments]{parse_bugzilla_perceval_rest_issue_comments()}},
\code{\link[=parse_bugzilla_perceval_traditional_issue_comments]{parse_bugzilla_perceval_traditional_issue_comments()}},
\code{\link[=parse_bugzilla_rest_comments]{parse_bugzilla_rest_comments()}},
\code{\link[=parse_bugzilla_rest_issues_comments]{parse_bugzilla_rest_issues_comments()}},
\code{\link[=parse_commit_message_id]{parse_commit_message_id()}},
\code{\link[=parse_dependencies]{parse_dependencies()}},
\code{\link[=parse_dv8_clusters]{parse_dv8_clusters()}},
\code{\link[=parse_github_latest_date]{parse_github_latest_date()}},
\code{\link[=parse_gitlog]{parse_gitlog()}},
\code{\link[=parse_jira]{parse_jira()}},
\code{\link[=parse_jira_latest_date]{parse_jira_latest_date()}},
\code{\link[=parse_jira_rss_xml]{parse_jira_rss_xml()}},
\code{\link[=parse_mbox]{parse_mbox()}},
\code{\link[=parse_mbox_latest_date]{parse_mbox_latest_date()}},
\code{\link[=parse_nvdfeed]{parse_nvdfeed()}},
\code{\link[=parse_understand_dependencies]{parse_understand_dependencies()}}
}
\concept{parsers}
