% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_bugzilla_rest_issues}
\alias{parse_bugzilla_rest_issues}
\title{Parse Bugzilla issues data obtained from json files from Bugzilla crawler \code{\link{download_bugzilla_rest_issues}}}
\usage{
parse_bugzilla_rest_issues(issues_folder_path)
}
\arguments{
\item{issues_folder_path}{path to the issue folder that contains json file with Bugzilla data inside}
}
\value{
data table with parsed bugzilla issues data
}
\description{
Parse Bugzilla issues data obtained from json files from Bugzilla crawler \code{\link{download_bugzilla_rest_issues}}
}
\seealso{
\code{\link{download_bugzilla_rest_issues}} a downloader function to download bugzilla issues data with REST API

Other parsers: 
\code{\link{parse_bugzilla_rest_comments}()},
\code{\link{parse_commit_message_id}()},
\code{\link{parse_dependencies}()},
\code{\link{parse_dv8_clusters}()},
\code{\link{parse_gitlog}()},
\code{\link{parse_jira}()},
\code{\link{parse_mbox}()},
\code{\link{parse_nvdfeed}()}
}
\concept{parsers}
