% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugzilla.R
\name{parse_bugzilla_rest_issues}
\alias{parse_bugzilla_rest_issues}
\title{Parse Bugzilla issues data obtained from json files from Bugzilla crawler}
\usage{
parse_bugzilla_rest_issues(issues_folder_path)
}
\arguments{
\item{issues_folder_path}{path to the issue folder that contains json file with Bugzilla data inside}
}
\value{
data table with parsed Bugzilla issues data
}
\description{
Parse Bugzilla issues data obtained from json files from Bugzilla crawler
}
\seealso{
\code{\link{download_bugzilla_rest_issues_comments}} a downloader function to parse Bugzilla issues and comments data

Other parsers: 
\code{\link{parse_bugzilla_perceval_rest_issue_comments}()},
\code{\link{parse_bugzilla_perceval_traditional_issue_comments}()},
\code{\link{parse_bugzilla_rest_comments}()},
\code{\link{parse_bugzilla_rest_issues_comments}()},
\code{\link{parse_commit_message_id}()},
\code{\link{parse_dependencies}()},
\code{\link{parse_dv8_clusters}()},
\code{\link{parse_gitlog}()},
\code{\link{parse_jira_rss_xml}()},
\code{\link{parse_jira}()},
\code{\link{parse_mbox}()},
\code{\link{parse_nvdfeed}()}
}
\concept{parsers}
