% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_bugzilla_rest_issues_comments}
\alias{parse_bugzilla_rest_issues_comments}
\title{Combine Bugzilla issues data table with Bugzilla comments data table}
\usage{
parse_bugzilla_rest_issues_comments(bugzilla_issues, bugzilla_comments)
}
\arguments{
\item{bugzilla_issues}{a data table with parsed Bugzilla issues data}

\item{bugzilla_comments}{a data table with parsed Bugzilla comments data}
}
\value{
data table combined Bugzilla issue data and Bugzilla comments data
}
\description{
Combine Bugzilla issues data table with Bugzilla comments data table
}
\seealso{
\code{\link{parse_bugzilla_rest_issues}} a parser function to parse Bugzilla issues data

\code{\link{parse_bugzilla_rest_comments}} a parser function to parse Bugzilla comments data

Other parsers: 
\code{\link{parse_bugzilla_rest_comments}()},
\code{\link{parse_bugzilla_rest_issues}()},
\code{\link{parse_commit_message_id}()},
\code{\link{parse_dependencies}()},
\code{\link{parse_dv8_clusters}()},
\code{\link{parse_gitlog}()},
\code{\link{parse_jira}()},
\code{\link{parse_mbox}()},
\code{\link{parse_nvdfeed}()}
}
\concept{parsers}
