% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_bugzillarest}
\alias{parse_bugzillarest}
\title{Parse Bugzilla data obtained from Perceval REST API Bugzilla backend}
\usage{
parse_bugzillarest(bugzilla_json, comments = FALSE)
}
\arguments{
\item{bugzilla_json}{json object containing Bugzilla data}

\item{comments}{if true, the comments are parsed along with the issues}
}
\value{
data table
}
\description{
Parse Bugzilla data obtained from Perceval REST API Bugzilla backend
}
\seealso{
\code{\link{download_bugzilla}} a donwoloader function download bugzilla data with perceval

Other parsers: 
\code{\link{dv8_depends_to_sdsmj}()},
\code{\link{parse_bugzilla_crawler_comments}()},
\code{\link{parse_bugzilla_crawler_issue}()},
\code{\link{parse_bugzilla}()},
\code{\link{parse_commit_message_id}()},
\code{\link{parse_dependencies}()},
\code{\link{parse_dv8_clusters}()},
\code{\link{parse_gitlog}()},
\code{\link{parse_jira}()},
\code{\link{parse_mbox}()},
\code{\link{parse_nvdfeed}()}
}
\concept{parsers}
