% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{parse_commit_message_id}
\alias{parse_commit_message_id}
\title{Adds a column commit_message_id containing the parsed commit message}
\usage{
parse_commit_message_id(project_git, commit_message_id_regex)
}
\arguments{
\item{project_git}{A parsed git project by \code{parse_gitlog}.}

\item{commit_message_id_regex}{the regex to extract the id from the commit message}
}
\description{
Adds a column commit_message_id containing the parsed commit message
}
\seealso{
Other parsers: 
\code{\link{build_understand_project}()},
\code{\link{export_understand_dependencies}()},
\code{\link{parse_bugzilla_perceval_rest_issue_comments}()},
\code{\link{parse_bugzilla_perceval_traditional_issue_comments}()},
\code{\link{parse_bugzilla_rest_comments}()},
\code{\link{parse_bugzilla_rest_issues}()},
\code{\link{parse_bugzilla_rest_issues_comments}()},
\code{\link{parse_dependencies}()},
\code{\link{parse_dv8_clusters}()},
\code{\link{parse_github_latest_date}()},
\code{\link{parse_gitlog}()},
\code{\link{parse_jira}()},
\code{\link{parse_jira_latest_date}()},
\code{\link{parse_jira_rss_xml}()},
\code{\link{parse_mbox}()},
\code{\link{parse_mbox_latest_date}()},
\code{\link{parse_nvdfeed}()},
\code{\link{parse_understand_dependencies}()}
}
\concept{parsers}
