% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_cve_cwe_file_network}
\alias{parse_cve_cwe_file_network}
\title{Transform parsed cveid and nvdfeed into a network}
\usage{
parse_cve_cwe_file_network(project_cve, nvd_feed)
}
\arguments{
\item{project_cve}{A parsed cve edgelist by \link{\code{parse_commit_message_id_network}}.}

\item{nvdfeed}{Parsed  nvdfeed by \link{\code{parse_nvdfeed}}.}
}
\description{
Transform parsed cveid and nvdfeed into a network
}
\seealso{
Other edgelists: 
\code{\link{parse_commit_message_id_network}()},
\code{\link{parse_dependencies_network}()},
\code{\link{parse_gitlog_network}()},
\code{\link{parse_mbox_network}()}
}
\concept{edgelists}
