% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_dependencies}
\alias{parse_dependencies}
\title{Parse dependencies from Depends}
\usage{
parse_dependencies(depends_jar_path, git_repo_path, language)
}
\arguments{
\item{depends_jar_path}{path to depends jar}

\item{git_repo_path}{path to git repo (ends in .git)}

\item{language}{the language of the .git repo (accepts cpp, java, ruby, python, pom)}
}
\description{
Parse dependencies from Depends
}
