% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{parse_dependencies_network}
\alias{parse_dependencies_network}
\title{Transform parsed dependencies into a network}
\usage{
parse_dependencies_network(depends_parsed, weight_types = NA)
}
\arguments{
\item{depends_parsed}{A parsed mbox by \code{parse_dependencies}.}

\item{weight_types}{The weight types as defined in Depends.}
}
\description{
Transform parsed dependencies into a network
}
\seealso{
Other edgelists: 
\code{\link{parse_commit_message_id_network}()},
\code{\link{parse_cve_cwe_file_network}()},
\code{\link{parse_gitlog_network}()},
\code{\link{parse_mbox_network}()}
}
\concept{edgelists}
