% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv8.R
\name{parse_dv8_architectural_flaws}
\alias{parse_dv8_architectural_flaws}
\title{Parse Architecture Flaws}
\usage{
parse_dv8_architectural_flaws(file_measure_report_path, flaws_path)
}
\arguments{
\item{file_measure_report_path}{path to file-measure-report.csv}

\item{flaws_path}{path to architecture folder}
}
\description{
Parse Architecture Flaws
}
