% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv8.R
\name{parse_dv8_architectural_flaws}
\alias{parse_dv8_architectural_flaws}
\title{Parse Architecture Flaws}
\usage{
parse_dv8_architectural_flaws(
  dv8_path,
  flaws_path,
  keep_intermediate_files = FALSE
)
}
\arguments{
\item{dv8_path}{path to dv8 binary}

\item{flaws_path}{path to architecture folder}

\item{keep_intermediate_files}{TRUE if the user wishes to keep the intermediate files generated by this function, FALSE otherwise}
}
\description{
Parse Architecture Flaws
}
\seealso{
Other dv8: 
\code{\link{dv8_clsxj_to_clsxb}()},
\code{\link{dv8_dsmj_to_dsmb}()},
\code{\link{dv8_hdsmb_sdsmb_to_hsdsmb}()},
\code{\link{dv8_hsdsmb_to_flaws}()}
}
\concept{dv8}
