% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv8.R
\name{parse_dv8_clusters}
\alias{parse_dv8_clusters}
\title{Parse a DV8 json cluster file to R data.table.}
\usage{
parse_dv8_clusters(clsxj_path)
}
\arguments{
\item{clsxj_path}{path to clustering json (obtained from dv8_clsxb_to_clsxj)}
}
\description{
Parses a *-clsx.json into a data table.
}
\seealso{
Other parsers: 
\code{\link{parse_commit_message_id}()},
\code{\link{parse_dependencies}()},
\code{\link{parse_gitlog}()},
\code{\link{parse_jira}()},
\code{\link{parse_mbox}()},
\code{\link{parse_nvdfeed}()}

Other dv8: 
\code{\link{dependencies_to_sdsmj}()},
\code{\link{dv8_clsxb_to_clsxj}()},
\code{\link{gitlog_to_hdsmj}()}
}
\concept{dv8}
\concept{parsers}
