% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{parse_git_blame}
\alias{parse_git_blame}
\title{Creates a dataframe where each line of each file has the last commit that touched the line}
\usage{
parse_git_blame(git_repo_path, all = FALSE, since, until)
}
\arguments{
\item{git_repo_path}{The git repo path}

\item{all}{If FALSE returns only the last commit for each line, if TRUE all co}
}
\value{
Dataframe with the last commit that touched each line of all files
}
\description{
Creates a dataframe where each line of each file has the last commit that touched the line
}
