% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{parse_github_replies}
\alias{parse_github_replies}
\title{Parse GitHub Issue and Pull Request Comments}
\usage{
parse_github_replies(github_replies_folder_path)
}
\arguments{
\item{github_replies_folder_path}{The folder path to where the github api json files have been downloaded.}
}
\value{
A single reply table which combines the communication from the three jsons.
}
\description{
Parses Issue, Pull Request, and Comments Endpoints into a reply table.
See example usage in the download_github_comments.Rmd vignette.
}
