% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_gitlog}
\alias{parse_gitlog}
\title{Parse gitlog from Perceval}
\usage{
parse_gitlog(perceval_path, git_repo_path, save_path = NA, perl_regex = NA)
}
\arguments{
\item{perceval_path}{path to perceval binary}

\item{git_repo_path}{path to git repo (ends in .git)}

\item{save_path}{optional save path for .rds object}

\item{from_date}{a string of the form "YYYY-MM-DD"}

\item{to_date}{a string of the form "YYYY-MM-DD"}
}
\description{
Parse gitlog from Perceval
}
\seealso{
Other parsers: 
\code{\link{parse_commit_message_id}()},
\code{\link{parse_dependencies}()},
\code{\link{parse_mbox}()},
\code{\link{parse_nvdfeed}()}
}
\concept{parsers}
