% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_gitlog_network}
\alias{parse_gitlog_network}
\title{Transform parsed git repo into an edgelist}
\usage{
parse_gitlog_network(project_git, mode = c("author", "commit"))
}
\arguments{
\item{project_git}{A parsed git project by \code{parse_gitlog}.}

\item{mode}{The network of interest: author-file, or commit-file}
}
\description{
Transform parsed git repo into an edgelist
}
\seealso{
Other edgelists: 
\code{\link{parse_dependencies_network}()},
\code{\link{parse_mbox_network}()}
}
\concept{edgelists}
