% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{parse_jira}
\alias{parse_jira}
\title{Parse Jira issue and comments}
\usage{
parse_jira(json_path)
}
\arguments{
\item{json_path}{is a folder path containing a set of jira_issues as json files. The general folder structure is as follows:
"../../rawdata/issue_tracker/geronimo/issues/" or "../../rawdata/issue_tracker/geronimo/issue_comments/". The issues folder only
contains jira issue json files that don't include comments. The issue_comments folder contains jira issue json files that can
include comments.}
}
\value{
A named list of two named elements ("issues", and "comments"), each containing a data.table.
}
\description{
Note the comments element will be empty if the downloaded json only contain issues.
the jira json naming convention is as follows: "(ProjectKey)_issues_(UNIXTIMElowerbound)_(UNITITMEupperbound).json"
or "(ProjectKey)_issue_comments_(UNIXTIMElowerbound)_(UNIXTIMEupperbound).json".
For example: "KAIAULU_issues_1231234_2312413.json". Notice how the ProjectKey portion is in all caps.
}
\seealso{
Other parsers: 
\code{\link{parse_bugzilla_perceval_rest_issue_comments}()},
\code{\link{parse_bugzilla_perceval_traditional_issue_comments}()},
\code{\link{parse_bugzilla_rest_comments}()},
\code{\link{parse_bugzilla_rest_issues_comments}()},
\code{\link{parse_bugzilla_rest_issues}()},
\code{\link{parse_commit_message_id}()},
\code{\link{parse_dependencies}()},
\code{\link{parse_dv8_clusters}()},
\code{\link{parse_gitlog}()},
\code{\link{parse_jira_latest_date}()},
\code{\link{parse_jira_rss_xml}()},
\code{\link{parse_mbox}()},
\code{\link{parse_nvdfeed}()}
}
\concept{parsers}
