% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{parse_jira}
\alias{parse_jira}
\title{Parse Jira issue and comments}
\usage{
parse_jira(json_path)
}
\arguments{
\item{json_path}{path to jira json (issues or issues with comments) obtained using `download_jira_data.Rmd`.}
}
\value{
A named list of two named elements ("issues", and "comments"), each containing a data.table.
Note the comments element will be empty if the downloaded json only contain issues.
}
\description{
Parse Jira issue and comments
}
\seealso{
Other parsers: 
\code{\link{parse_bugzilla_perceval_rest_issue_comments}()},
\code{\link{parse_bugzilla_perceval_traditional_issue_comments}()},
\code{\link{parse_bugzilla_rest_comments}()},
\code{\link{parse_bugzilla_rest_issues_comments}()},
\code{\link{parse_bugzilla_rest_issues}()},
\code{\link{parse_commit_message_id}()},
\code{\link{parse_dependencies}()},
\code{\link{parse_dv8_clusters}()},
\code{\link{parse_gitlog}()},
\code{\link{parse_jira_rss_xml}()},
\code{\link{parse_mbox}()},
\code{\link{parse_nvdfeed}()}
}
\concept{parsers}
