% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{parse_jira_latest_date}
\alias{parse_jira_latest_date}
\title{parse latest date}
\usage{
parse_jira_latest_date(json_path)
}
\arguments{
\item{json_path}{the path}
}
\description{
Download  Issues from "GET /repos/{owner}/{repo}/issues" endpoint.
}
