% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{parse_jira_latest_date}
\alias{parse_jira_latest_date}
\title{parse latest date}
\usage{
parse_jira_latest_date(json_path)
}
\arguments{
\item{json_path}{the path with folders to read}
}
\description{
Takes a filepath and returns a filename of the .json file that contains the
most recent 'created_at' value
}
