% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira.R
\name{parse_jira_latest_date}
\alias{parse_jira_latest_date}
\title{Parse Jira latest dates}
\usage{
parse_jira_latest_date(json_path)
}
\arguments{
\item{json_path}{path to save folder containing JIRA issue and/or comments json files.}
}
\value{
The name of the jira issue file with the latest created date that was created/downloaded for
use by the Jira Downloader refresher
}
\description{
This is a function that returns the json file containing the Jira issue with the latest created date.
It is based on the name of the file, which is as follows: "(ProjectKey)_issues_(UNIXTIMElowerbound)_(UNITITMEupperbound).json"
or "(ProjectKey)_issue_comments_(UNIXTIMElowerbound)_(UNIXTIMEupperbound).json"
For example: "KAIAULU_issues_1231234_2312413.json". Notice how the ProjectKey portion is in all caps.
}
\seealso{
Other parsers: 
\code{\link{parse_bugzilla_perceval_rest_issue_comments}()},
\code{\link{parse_bugzilla_perceval_traditional_issue_comments}()},
\code{\link{parse_bugzilla_rest_comments}()},
\code{\link{parse_bugzilla_rest_issues_comments}()},
\code{\link{parse_bugzilla_rest_issues}()},
\code{\link{parse_commit_message_id}()},
\code{\link{parse_dependencies}()},
\code{\link{parse_dv8_clusters}()},
\code{\link{parse_gitlog}()},
\code{\link{parse_jira_rss_xml}()},
\code{\link{parse_jira}()},
\code{\link{parse_mbox}()},
\code{\link{parse_nvdfeed}()}
}
\concept{parsers}
