% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{parse_line_type_file}
\alias{parse_line_type_file}
\title{Parse File Line Type}
\usage{
parse_line_type_file(utags_path, filepath, kinds)
}
\arguments{
\item{utags_path}{The path to utags binary.
See \url{https://github.com/universal-ctags/ctags}}

\item{filepath}{path to file}

\item{kinds}{the entity kinds utags should identify per line.}
}
\description{
Parse File Line Type
}
\keyword{internal}
