% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_mbox}
\alias{parse_mbox}
\title{Parse mbox from Perceval}
\usage{
parse_mbox(perceval_path, mbox_path)
}
\arguments{
\item{perceval_path}{path to perceval binary}

\item{mbox_path}{path to mbox archive file (ends in .mbox)}
}
\description{
Parse mbox from Perceval
}
\seealso{
Other parsers: 
\code{\link{parse_commit_message_id}()},
\code{\link{parse_dependencies}()},
\code{\link{parse_gitlog}()},
\code{\link{parse_nvdfeed}()}
}
\concept{parsers}
