% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\name{parse_mbox_latest_date}
\alias{parse_mbox_latest_date}
\title{Parse mbox latest date}
\usage{
parse_mbox_latest_date(mbox_path)
}
\arguments{
\item{mbox_path}{path to mbox archive file (ends in .mbox)}
}
\value{
Returns the name of the latest mod_mbox file
}
\description{
Returns the name of the latest mod_mbox file downloaded in the specified folder
}
\details{
The folder assumes the following convention: "(mailing_list)_(archive_type)_yearmonth.mbox"
For example: "geronimo-dev_apache_202401.mbox". This nomenclature is defined by \code{\link{download_mod_mbox_per_month}}
}
\seealso{
Other parsers: 
\code{\link{parse_bugzilla_perceval_rest_issue_comments}()},
\code{\link{parse_bugzilla_perceval_traditional_issue_comments}()},
\code{\link{parse_bugzilla_rest_comments}()},
\code{\link{parse_bugzilla_rest_issues_comments}()},
\code{\link{parse_bugzilla_rest_issues}()},
\code{\link{parse_commit_message_id}()},
\code{\link{parse_dependencies}()},
\code{\link{parse_dv8_clusters}()},
\code{\link{parse_gitlog}()},
\code{\link{parse_jira_latest_date}()},
\code{\link{parse_jira_rss_xml}()},
\code{\link{parse_jira}()},
\code{\link{parse_mbox}()},
\code{\link{parse_nvdfeed}()}
}
\concept{parsers}
