% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_nvdfeed}
\alias{parse_nvdfeed}
\title{Parse NVD Feed CVEs, descriptions and CWE ids}
\usage{
parse_nvdfeed(nvdfeed_folder_path)
}
\arguments{
\item{nvdfeed_folder_path}{Folderpath for nvd feed files
under schema 1.1 (e.g. nvdcve-1.1-2018.json)}
}
\description{
Parse NVD Feed CVEs, descriptions and CWE ids
}
\seealso{
Other parsers: 
\code{\link{parse_bugzilla_perceval_rest_issue_comments}()},
\code{\link{parse_bugzilla_perceval_traditional_issue_comments}()},
\code{\link{parse_commit_message_id}()},
\code{\link{parse_dependencies}()},
\code{\link{parse_dv8_clusters}()},
\code{\link{parse_gitlog}()},
\code{\link{parse_jira}()},
\code{\link{parse_mbox}()}
}
\concept{parsers}
