% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{parse_r_function_dependencies}
\alias{parse_r_function_dependencies}
\title{Parse R Function Dependencies}
\usage{
parse_r_function_dependencies(parsed_r_file, function_definition)
}
\arguments{
\item{parsed_r_file}{A parsed R file (see \code{\link{parse_rfile_ast}})}

\item{function_definition}{A list of function definitions (see \code{\link{parse_r_function_definition}})}
}
\description{
Identify function calls in an R File.
}
\keyword{internal}
