% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{parse_understand_dependencies}
\alias{parse_understand_dependencies}
\title{Parse XML from Understand DB}
\usage{
parse_understand_dependencies(scitools_path, understand_dir)
}
\arguments{
\item{scitools_path}{path to the scitools binary `und`}

\item{understand_dir}{path to the built Understand project folder used in \code{\link{build_understand_project}}}
}
\description{
This function parses the data in the Understand build folder to export the parse_type dependencies into a network
}
\seealso{
Other parsers: 
\code{\link{build_understand_project}()},
\code{\link{export_understand_dependencies}()},
\code{\link{parse_bugzilla_perceval_rest_issue_comments}()},
\code{\link{parse_bugzilla_perceval_traditional_issue_comments}()},
\code{\link{parse_bugzilla_rest_comments}()},
\code{\link{parse_bugzilla_rest_issues_comments}()},
\code{\link{parse_bugzilla_rest_issues}()},
\code{\link{parse_commit_message_id}()},
\code{\link{parse_dependencies}()},
\code{\link{parse_dv8_clusters}()},
\code{\link{parse_gitlog}()},
\code{\link{parse_jira_latest_date}()},
\code{\link{parse_jira_rss_xml}()},
\code{\link{parse_jira}()},
\code{\link{parse_mbox}()},
\code{\link{parse_nvdfeed}()}
}
\concept{parsers}
