% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\name{process_gz_to_mbox_in_folder}
\alias{process_gz_to_mbox_in_folder}
\title{Process .gz files in a folder and convert them to .mbox}
\usage{
process_gz_to_mbox_in_folder(save_folder_path, verbose = TRUE)
}
\arguments{
\item{save_folder_path}{The path to the folder containing both .gz and .mbox files.}

\item{verbose}{if TRUE, prints diagnostic messages during processing.}
}
\value{
A list of the .mbox files that were created or updated.
}
\description{
This function scans a specified folder for any .gz files, unzips them,
and renames them to the .mbox format. After unzipping, the original .gz files are deleted.
If a .mbox file with the same name already exists, it will be overwritten.
This makes sure that all the files in the folder are in .mbox format, ready for parsing.
}
