% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{productivity_author_churn}
\alias{productivity_author_churn}
\title{Productivity Author Churn}
\usage{
productivity_author_churn(project_git_log, quit_lag = 90)
}
\arguments{
\item{project_git_log}{a parsed git log obtained from \code{\link{parse_gitlog}}}

\item{quit_lag}{The number of days to look back for the rolling window (90 day default)}
}
\value{
a five column data.table of the form author_name_email | author_datetimetz | lines_added | 
lines_removed | author_churn. 
At each timestamp, the lines_added, lines_removed, and author_churn are the totals across all rows 
for that author within the prior 90 days ending at that timestamp.
}
\description{
Calculates the churn per author in a rolling window based on the git log.
}
\seealso{
\code{\link{parse_gitlog}} to obtain additions and deletions from gitlog

Other metrics: 
\code{\link{commit_message_id_coverage}()},
\code{\link{metric_churn}()},
\code{\link{metric_churn_per_commit_interval}()},
\code{\link{metric_churn_per_commit_per_file}()},
\code{\link{metric_file_bug_churn}()},
\code{\link{metric_file_bug_frequency}()},
\code{\link{metric_file_churn}()},
\code{\link{metric_file_non_bug_churn}()},
\code{\link{metric_file_non_bug_frequency}()},
\code{\link{productivity_author_commits}()}
}
\concept{metrics}
