% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{productivity_author_commits}
\alias{productivity_author_commits}
\title{Productivity Author Commits Metric}
\usage{
productivity_author_commits(project_git, lag = 90)
}
\arguments{
\item{project_git}{a parsed git log obtained from \code{\link{parse_gitlog}}}

\item{lag}{the number of days to look back for the rolling window (90 day default)}
}
\value{
a three column data.table of the form author_name_email | author_datetimetz | author_total_commits.
For each row (each timestamp in author_datetimetz), author_total_commits is the number of unique 
commits by that author in the preceding 90 days ending at that timestamp.
}
\description{
Counts the unique number of commits per author in a rolling window based on
the git log.
}
\seealso{
\code{\link{parse_gitlog}}

Other metrics: 
\code{\link{commit_message_id_coverage}()},
\code{\link{jira_author_communication_count}()},
\code{\link{metric_churn}()},
\code{\link{metric_churn_per_commit_interval}()},
\code{\link{metric_churn_per_commit_per_file}()},
\code{\link{metric_file_bug_churn}()},
\code{\link{metric_file_bug_frequency}()},
\code{\link{metric_file_churn}()},
\code{\link{metric_file_non_bug_churn}()},
\code{\link{metric_file_non_bug_frequency}()},
\code{\link{productivity_author_churn}()}
}
\concept{metrics}
