% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{recolor_network_by_community}
\alias{recolor_network_by_community}
\title{Re-color OSLOM Community IDs}
\usage{
recolor_network_by_community(network, community)
}
\arguments{
\item{network}{A network returned by transform_to_network_* functions.}

\item{community}{A community detection returned by \code{community_oslom}.}
}
\description{
Re-color a graph color column using the OSLOM communities
}
\references{
Finding statistically significant communities in networks
A. Lancichinetti, F. Radicchi, J.J. Ramasco and
S. Fortunato PLoS ONE 6, e18961 (2011).
}
\seealso{
Other community: 
\code{\link{community_oslom}()}
}
\concept{community}
