% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\name{refresh_mod_mbox}
\alias{refresh_mod_mbox}
\title{Refresh mbox files}
\usage{
refresh_mod_mbox(
  archive_url,
  mailing_list,
  archive_type,
  from_year,
  save_folder_path,
  verbose = FALSE
)
}
\arguments{
\item{archive_url}{A url pointing to the mod_mbox mailing list directory (e.g. "http://mail-archives.apache.org/mod_mbox/apr-dev") without trailing slashes}

\item{mailing_list}{Name of the project mailing list (e.g. apr-dev) in the mod_mbox directory}

\item{archive_type}{Name of the archive that the project mailing list is archived in (e.g. apache)}

\item{from_year}{First year in the range to be downloaded in case there are no mod_mbox files already downloaded}

\item{save_folder_path}{the full *folder* path where the monthly downloaded mbox will be stored.}

\item{verbose}{Prints progress during execution}
}
\description{
Uses the adopted file name convention by \code{\link{download_mod_mbox_per_month}} to identify
the latest downloaded mbox year i and month j. It deletes the mbox file of year i and month j,
then redownloads it along with the remaining months past j up to 12. Then, it calls
\code{\link{download_mod_mbox_per_month}} with from_year being year i+1 and to_year being
the current real-life year so that all newer mbox files are downloaded.
}
\details{
If the directory is empty, then it downloads all mbox files starting from a definable starting year to
the current real-life year.
}
