% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smells.R
\name{smell_detect_ragequit}
\alias{smell_detect_ragequit}
\title{Detect Ragequit Social Smell.}
\usage{
smell_detect_ragequit(ragequit_data, quit_lag = 90)
}
\arguments{
\item{ragequit_data}{A data table containing four columns; timestamp of a commit, author, the
author's commit messages, and a sentiment value (integer) assigned to the commit messages}

\item{quit_lag}{The number of days since a developer's last commit message}
}
\description{
Detect potential ragequits by identifying cases where, 
after a period of inactivity, a developer's final commit message expressed negative sentiment.
}
\references{
Wouter Mulder (2025). Am I finished yet? A discovery of burnout and
ragequits within open-source projects. (Master thesis, Jheronimus Academy of Data Science).
}
