% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{transform_commit_message_id_to_network}
\alias{transform_commit_message_id_to_network}
\title{Transform parsed git repo commit messages id and files into an edgelist}
\usage{
transform_commit_message_id_to_network(
  project_git,
  commit_message_id_regex,
  weight = NULL,
  weight_agg = mean
)
}
\arguments{
\item{project_git}{A parsed git project by \code{\link{parse_gitlog}}.}

\item{commit_message_id_regex}{the regex to extract the id from the commit message}
}
\description{
Transform parsed git repo commit messages id and files into an edgelist
}
\seealso{
Other edgelists: 
\code{\link{transform_cve_cwe_file_to_network}()},
\code{\link{transform_dependencies_to_network}()},
\code{\link{transform_dependencies_to_sdsmj}()},
\code{\link{transform_gitlog_to_bipartite_network}()},
\code{\link{transform_gitlog_to_entity_bipartite_network}()},
\code{\link{transform_gitlog_to_entity_temporal_network}()},
\code{\link{transform_gitlog_to_hdsmj}()},
\code{\link{transform_gitlog_to_multimodal_network}()},
\code{\link{transform_gitlog_to_temporal_network}()},
\code{\link{transform_reply_to_bipartite_network}()},
\code{\link{transform_temporal_gitlog_to_adsmj}()},
\code{\link{transform_understand_dependencies_to_network}()}
}
\concept{edgelists}
