% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vulnerabilities.R
\name{transform_cve_cwe_file_to_network}
\alias{transform_cve_cwe_file_to_network}
\title{Transform parsed cveid and nvdfeed into a network}
\usage{
transform_cve_cwe_file_to_network(project_cve, nvd_feed)
}
\arguments{
\item{project_cve}{A parsed cve edgelist by \code{\link{transform_commit_message_id_to_network}}.}

\item{nvd_feed}{Parsed  nvdfeed by \code{\link{parse_nvdfeed}}.}
}
\description{
Transform parsed cveid and nvdfeed into a network
}
\seealso{
Other edgelists: 
\code{\link{transform_commit_message_id_to_network}()},
\code{\link{transform_dependencies_to_network}()},
\code{\link{transform_dependencies_to_sdsmj}()},
\code{\link{transform_gitlog_to_bipartite_network}()},
\code{\link{transform_gitlog_to_entity_bipartite_network}()},
\code{\link{transform_gitlog_to_entity_temporal_network}()},
\code{\link{transform_gitlog_to_hdsmj}()},
\code{\link{transform_gitlog_to_heterogeneous_network}()},
\code{\link{transform_gitlog_to_temporal_network}()},
\code{\link{transform_reply_to_bipartite_network}()},
\code{\link{transform_temporal_gitlog_to_adsmj}()},
\code{\link{transform_understand_dependencies_to_network}()}
}
\concept{edgelists}
