% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{transform_gitlog_to_bipartite_network}
\alias{transform_gitlog_to_bipartite_network}
\title{Transform parsed git log into a bipartite network}
\usage{
transform_gitlog_to_bipartite_network(
  project_git,
  mode = c("author-file", "committer-file", "commit-file", "author-committer")
)
}
\arguments{
\item{project_git}{A parsed git project by \code{\link{parse_gitlog}}.}

\item{mode}{The network of interest: author-file, committer-file, commit-file, author-committer}
}
\description{
Constructs a bipartite graph from a parsed git log where one node type
represents authors, committers, or commits and the other represents files.
The transform function is responsible for building the nodes and edgelist
tables before passing them to \code{\link{model_bipartite_graph}}.
}
\seealso{
Other edgelists: 
\code{\link{transform_commit_message_id_to_network}()},
\code{\link{transform_cve_cwe_file_to_network}()},
\code{\link{transform_dependencies_to_network}()},
\code{\link{transform_dependencies_to_sdsmj}()},
\code{\link{transform_gitlog_to_entity_bipartite_network}()},
\code{\link{transform_gitlog_to_entity_temporal_network}()},
\code{\link{transform_gitlog_to_hdsmj}()},
\code{\link{transform_gitlog_to_heterogeneous_network}()},
\code{\link{transform_gitlog_to_temporal_network}()},
\code{\link{transform_reply_to_bipartite_network}()},
\code{\link{transform_temporal_gitlog_to_adsmj}()},
\code{\link{transform_understand_dependencies_to_network}()}
}
\concept{edgelists}
