% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{transform_gitlog_to_bipartite_network}
\alias{transform_gitlog_to_bipartite_network}
\title{Transform parsed git repo into an edgelist}
\usage{
transform_gitlog_to_bipartite_network(
  project_git,
  mode = c("author-file", "committer-file", "commit-file", "author-committer")
)
}
\arguments{
\item{project_git}{A parsed git project by \code{\link{parse_gitlog}}.}

\item{mode}{The network of interest: author-entity, committer-entity, commit-entity, author-committer}
}
\description{
Transform parsed git repo into an edgelist
}
\seealso{
Other edgelists: 
\code{\link{transform_commit_message_id_to_network}()},
\code{\link{transform_cve_cwe_file_to_network}()},
\code{\link{transform_dependencies_to_network}()},
\code{\link{transform_dependencies_to_sdsmj}()},
\code{\link{transform_gitlog_to_entity_bipartite_network}()},
\code{\link{transform_gitlog_to_entity_temporal_network}()},
\code{\link{transform_gitlog_to_hdsmj}()},
\code{\link{transform_gitlog_to_temporal_network}()},
\code{\link{transform_reply_to_bipartite_network}()}
}
\concept{edgelists}
