% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{transform_r_dependencies_to_network}
\alias{transform_r_dependencies_to_network}
\title{Transform parsed R dependencies into a graph}
\usage{
transform_r_dependencies_to_network(
  r_dependencies_edgelist,
  dependency_type = c("function", "file")
)
}
\arguments{
\item{r_dependencies_edgelist}{A parsed R folder by \code{\link{parse_r_dependencies}}.}

\item{dependency_type}{The type of dependency to be parsed: Function or File}
}
\description{
Transform parsed R dependencies into a graph
}
