% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{transform_reply_to_bipartite_network}
\alias{transform_reply_to_bipartite_network}
\title{Transform parsed mbox or parsed jira replies into a network}
\usage{
transform_reply_to_bipartite_network(project_reply)
}
\arguments{
\item{project_reply}{A parsed mbox by \code{\link{parse_mbox}} or \code{\link{parse_jira_replies}}.}
}
\description{
Transform parsed mbox or parsed jira replies into a network
}
\seealso{
Other edgelists: 
\code{\link{transform_commit_message_id_to_network}()},
\code{\link{transform_cve_cwe_file_to_network}()},
\code{\link{transform_dependencies_to_network}()},
\code{\link{transform_dependencies_to_sdsmj}()},
\code{\link{transform_gitlog_to_bipartite_network}()},
\code{\link{transform_gitlog_to_entity_bipartite_network}()},
\code{\link{transform_gitlog_to_entity_temporal_network}()},
\code{\link{transform_gitlog_to_hdsmj}()},
\code{\link{transform_gitlog_to_temporal_network}()}
}
\concept{edgelists}
