% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{transform_understand_dependencies_to_network}
\alias{transform_understand_dependencies_to_network}
\title{Transform Understand Dependencies}
\usage{
transform_understand_dependencies_to_network(parsed, weight_types)
}
\arguments{
\item{parsed}{Parsed table from \code{\link{parse_understand_dependencies}}}

\item{weight_types}{The weight types as defined in Depends. Accepts single string and vector input}
}
\description{
This function subsets a parsed table from parse_understand_dependencies
}
\seealso{
Other edgelists: 
\code{\link{transform_commit_message_id_to_network}()},
\code{\link{transform_cve_cwe_file_to_network}()},
\code{\link{transform_dependencies_to_network}()},
\code{\link{transform_dependencies_to_sdsmj}()},
\code{\link{transform_gitlog_to_bipartite_network}()},
\code{\link{transform_gitlog_to_entity_bipartite_network}()},
\code{\link{transform_gitlog_to_entity_temporal_network}()},
\code{\link{transform_gitlog_to_hdsmj}()},
\code{\link{transform_gitlog_to_heterogeneous_network}()},
\code{\link{transform_gitlog_to_temporal_network}()},
\code{\link{transform_reply_to_bipartite_network}()},
\code{\link{transform_temporal_gitlog_to_adsmj}()}
}
\concept{edgelists}
